% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{adduct_has_explicit_loss}
\alias{adduct_has_explicit_loss}
\title{Return TRUE when an adduct string explicitly carries a given neutral loss term}
\usage{
adduct_has_explicit_loss(adduct, loss_formula)
}
\description{
Return TRUE when an adduct string explicitly carries a given neutral loss term
}
\keyword{internal}
