% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ADDUCT_MASSES}
\alias{ADDUCT_MASSES}
\title{Exact masses of common adduct components (Daltons)}
\format{
An object of class \code{list} of length 13.
}
\usage{
ADDUCT_MASSES
}
\description{
Exact masses of common adduct components (Daltons)
}
\keyword{datasets}
