% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ADDUCT_REGEX_PATTERN}
\alias{ADDUCT_REGEX_PATTERN}
\title{Regular expression pattern for parsing adduct notation
Matches patterns like: [M+H]+, [2M+Na]+, [M-H2O+H]+, [2M1+H]2+
Groups: (1) multimer, (2) isotope, (3) modifications, (4) charge count, (5) charge sign}
\format{
An object of class \code{character} of length 1.
}
\usage{
ADDUCT_REGEX_PATTERN
}
\description{
Regular expression pattern for parsing adduct notation
Matches patterns like: [M+H]+, [2M+Na]+, [M-H2O+H]+, [2M1+H]2+
Groups: (1) multimer, (2) isotope, (3) modifications, (4) charge count, (5) charge sign
}
\keyword{internal}
