% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates_utils.R
\name{all_single_strings}
\alias{all_single_strings}
\title{Check if all elements in a list are single character strings}
\usage{
all_single_strings(x_list)
}
\arguments{
\item{x_list}{Named list of values to check}
}
\value{
Named logical vector indicating which elements are single strings
}
\description{
Vectorized check for lists of parameters that should all be
single character strings. More efficient than repeated individual checks.
}
\examples{
\dontrun{
params <- list(
  file1 = "path.txt",
  file2 = c("a", "b"),
  file3 = "another.txt"
)
all_single_strings(params)
# file1: TRUE, file2: FALSE, file3: TRUE
}
}
\keyword{internal}
