% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms1.R
\name{annotate_ms1}
\alias{annotate_ms1}
\title{Annotate MS1}
\usage{
annotate_ms1(
  annotationTable = metadata_table_spectral_annotation,
  structureExactMassTable = structure_exact_mass_table,
  structureOrganismPairsTable = structure_organism_pairs_table,
  adducts = unlist(adducts_list[[ms_mode]]),
  neutralLosses = neutral_losses_table,
  msMode = ms_mode,
  tolerancePpm = tolerance_ppm,
  toleranceRt = tolerance_rt,
  candidatesInitial = candidates_initial
)
}
\arguments{
\item{annotationTable}{TODO}

\item{structureExactMassTable}{TODO}

\item{structureOrganismPairsTable}{TODO}

\item{adducts}{TODO}

\item{neutralLosses}{TODO}

\item{msMode}{TODO}

\item{tolerancePpm}{TODO}

\item{toleranceRt}{TODO}

\item{candidatesInitial}{TODO}
}
\value{
TODO
}
\description{
Annotate MS1
}
\examples{
TODO
}
