% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_non_ms1.R
\name{annotate_non_ms1}
\alias{annotate_non_ms1}
\title{Annotate non MS1}
\usage{
annotate_non_ms1(
  annotationTable = metadata_table_spectral_annotation,
  candidatesInitial = candidates_initial
)
}
\arguments{
\item{annotationTable}{TODO}

\item{candidatesInitial}{TODO}
}
\value{
TODO
}
\description{
Annotate non MS1
}
\examples{
TODO
}
