% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{append_component_weak_hypotheses}
\alias{append_component_weak_hypotheses}
\title{Recover weak (component-M coherent) hypotheses for unresolved nodes}
\usage{
append_component_weak_hypotheses(
  node_hypotheses,
  features_table,
  neutral_losses_list,
  clusters,
  baseline_adducts,
  tolerance_ppm,
  tolerance_dalton
)
}
\description{
Recover weak (component-M coherent) hypotheses for unresolved nodes
}
\keyword{internal}
