% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{apply_modifier_to_adducts}
\alias{apply_modifier_to_adducts}
\title{Apply a modifier formula to adduct strings and simplify notation}
\usage{
apply_modifier_to_adducts(adducts, modifier, sign)
}
\description{
Apply a modifier formula to adduct strings and simplify notation
}
\keyword{internal}
