% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_merged.R
\name{apply_taxonomic_filter}
\alias{apply_taxonomic_filter}
\title{Apply Taxonomic Filter}
\usage{
apply_taxonomic_filter(table_keys, table_org_tax_ott, level, value)
}
\description{
Internal helper to filter library by taxonomic criteria.
}
\keyword{internal}
