% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_universe.R
\name{ATOMIC_MONOISOTOPIC_MASS}
\alias{ATOMIC_MONOISOTOPIC_MASS}
\title{Monoisotopic neutral mass (Da) for the atomic / radical building blocks
used to compose charge carriers.}
\usage{
ATOMIC_MONOISOTOPIC_MASS
}
\description{
Monoisotopic neutral mass (Da) for the atomic / radical building blocks
used to compose charge carriers.
}
\keyword{internal}
