% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{biological_cleaning}
\alias{biological_cleaning}
\title{Title}
\usage{
biological_cleaning(
  annotationTableWeightedBio = annotation_table_weighted_bio,
  structureOrganismPairsTable = structure_organism_pairs_table,
  edgesTable = edges_table
)
}
\arguments{
\item{edgesTable}{}
}
\value{

}
\description{
Title
}
