% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{build_adduct_lookup_from_strings}
\alias{build_adduct_lookup_from_strings}
\title{Build an adduct lookup directly from attributed adduct strings}
\usage{
build_adduct_lookup_from_strings(adducts)
}
\description{
Build an adduct lookup directly from attributed adduct strings
}
\keyword{internal}
