% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{build_adduct_pair_differences}
\alias{build_adduct_pair_differences}
\title{Build all ordered adduct-pair differences for single-charge adducts}
\usage{
build_adduct_pair_differences(add_clu_table, tolerance_ppm, max_mz)
}
\description{
Build all ordered adduct-pair differences for single-charge adducts
}
\keyword{internal}
