% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{build_annotate_masses_annotations}
\alias{build_annotate_masses_annotations}
\title{Match, enrich, and prune annotate_masses annotations}
\usage{
build_annotate_masses_annotations(
  node_hypotheses,
  library_em,
  structure_table,
  tolerance_ppm,
  tolerance_dalton,
  adduct_edges,
  baseline_adduct,
  coverage_mode = "best_supported_conflict_free"
)
}
\description{
Match, enrich, and prune annotate_masses annotations
}
\keyword{internal}
