% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{build_annotate_masses_candidate_hypotheses}
\alias{build_annotate_masses_candidate_hypotheses}
\title{Build and expand node-level candidate hypotheses for annotate_masses}
\usage{
build_annotate_masses_candidate_hypotheses(
  adduct_edges,
  cluster_edges,
  loss_edges,
  evidence_hypotheses,
  preassigned,
  baseline_adducts,
  features_table,
  multi_adducts,
  tolerance_ppm,
  tolerance_dalton
)
}
\description{
Build and expand node-level candidate hypotheses for annotate_masses
}
\keyword{internal}
