% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_coverage.R
\name{build_annotate_masses_coverage_report}
\alias{build_annotate_masses_coverage_report}
\title{Build a feature-level coverage report for annotate_masses}
\usage{
build_annotate_masses_coverage_report(annotations, baseline_adduct)
}
\description{
Build a feature-level coverage report for annotate_masses
}
\keyword{internal}
