% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{build_annotate_masses_ion_tables}
\alias{build_annotate_masses_ion_tables}
\title{Resolve adduct tables used by annotate_masses ion discovery}
\usage{
build_annotate_masses_ion_tables(
  adducts_list,
  clusters_list,
  solvents_list,
  ms_mode
)
}
\description{
Resolve adduct tables used by annotate_masses ion discovery
}
\keyword{internal}
