% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chemo.R
\name{build_mini_taxonomy_table}
\alias{build_mini_taxonomy_table}
\title{Build mini-tier taxonomy table}
\usage{
build_mini_taxonomy_table(
  df_percentile,
  score_chemical_cla_kingdom,
  score_chemical_cla_superclass,
  score_chemical_cla_class,
  score_chemical_cla_parent,
  score_chemical_npc_pathway,
  score_chemical_npc_superclass,
  score_chemical_npc_class
)
}
\arguments{
\item{df_percentile}{Percentile-filtered candidate table}

\item{score_chemical_cla_kingdom}{ClassyFire kingdom weight}

\item{score_chemical_cla_superclass}{ClassyFire superclass weight}

\item{score_chemical_cla_class}{ClassyFire class weight}

\item{score_chemical_cla_parent}{ClassyFire direct parent weight}

\item{score_chemical_npc_pathway}{NPClassifier pathway weight}

\item{score_chemical_npc_superclass}{NPClassifier superclass weight}

\item{score_chemical_npc_class}{NPClassifier class weight}
}
\value{
Mini-tier taxonomy table keyed by feature_id
}
\description{
Build mini-tier taxonomy table
}
\keyword{internal}
