% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{build_modifier_mass_table}
\alias{build_modifier_mass_table}
\title{Compute direct mass-diff table from cluster/solvent/loss formulas}
\usage{
build_modifier_mass_table(modifiers, out_col, strip_label = FALSE)
}
\description{
Compute direct mass-diff table from cluster/solvent/loss formulas
}
\keyword{internal}
