% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate entropy}
\usage{
calculate_entropy(
  index,
  target,
  frags,
  ms2_tolerance,
  ppm_tolerance,
  threshold = 0.1
)
}
\arguments{
\item{index}{Index}

\item{target}{Target}

\item{frags}{Fragments}

\item{ms2_tolerance}{MS2 tolerance}

\item{ppm_tolerance}{ppm tolerance}

\item{threshold}{Threshold}
}
\description{
This function is part of the creation of edges
}
\examples{
NULL
}
