% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy_score.R
\name{calculate_entropy_score}
\alias{calculate_entropy_score}
\title{Calculate entropy score}
\usage{
calculate_entropy_score(
  lib_ids,
  lib_precursors,
  lib_spectra,
  query_ids,
  query_precursors,
  query_spectra,
  dalton,
  ppm,
  threshold,
  approx
)
}
\arguments{
\item{lib_ids}{Lib Ids}

\item{lib_precursors}{Lib precursors}

\item{lib_spectra}{Lib spectra}

\item{query_ids}{Query Ids}

\item{query_precursors}{Query precursors}

\item{query_spectra}{Query spectra}

\item{dalton}{Dalton}

\item{ppm}{Ppm}

\item{threshold}{Threshold}

\item{approx}{Approx}
}
\description{
This function applies similarity calculation to a list of
spectra to obtain entropy scores
}
\examples{
NULL
}
