% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mass_from_adduct.R
\name{calculate_mass_from_adduct}
\alias{calculate_mass_from_adduct}
\title{Parse adduct}
\usage{
calculate_mass_from_adduct(adduct_string, mass, electron_mass = 0.0005485799)
}
\arguments{
\item{adduct_string}{Adduct to be parsed}

\item{mass}{Mass}

\item{electron_mass}{Electron mass}
}
\value{
An exact mass
}
\description{
This function parses adducts
}
\examples{
NULL
}
