% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{calculate_modification_masses}
\alias{calculate_modification_masses}
\title{Calculate total mass change from modifications}
\usage{
calculate_modification_masses(elements, signs, multiplicities)
}
\description{
Calculate total mass change from modifications
}
\keyword{internal}
