% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_universe.R
\name{calculate_mz_from_neutral_mass}
\alias{calculate_mz_from_neutral_mass}
\title{Vectorized expected-m/z from neutral mass + universe (inverse).}
\usage{
calculate_mz_from_neutral_mass(
  neutral_mass,
  n_mer,
  z,
  adduct_mass,
  adduct_mass_per_monomer = 0
)
}
\description{
Vectorized expected-m/z from neutral mass + universe (inverse).
}
\keyword{internal}
