% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R, R/mass_consistency.R
\name{calculate_net_mod_mass_from_text}
\alias{calculate_net_mod_mass_from_text}
\title{Estimate net modification mass directly from adduct text}
\usage{
calculate_net_mod_mass_from_text(adduct)

calculate_net_mod_mass_from_text(adduct)
}
\description{
Estimate net modification mass directly from adduct text

Estimate net modification mass directly from adduct text
}
\keyword{internal}
