% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_similarity.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Calculate similarity}
\usage{
calculate_similarity(
  method,
  query_spectrum,
  target_spectrum,
  query_precursor,
  target_precursor,
  dalton,
  ppm
)
}
\arguments{
\item{method}{Method}

\item{query_spectrum}{Query spectrum}

\item{target_spectrum}{Target spectrum}

\item{query_precursor}{Query precursor}

\item{target_precursor}{Target precursor}

\item{dalton}{Dalton}

\item{ppm}{Ppm}
}
\description{
This function applies similarity calculation to a list of
spectra to obtain similarity scores
}
\examples{
NULL
}
