% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{canonicalize_adduct_notation}
\alias{canonicalize_adduct_notation}
\title{Canonicalize one adduct string}
\usage{
canonicalize_adduct_notation(adduct)
}
\description{
Canonicalize one adduct string
}
\keyword{internal}
