% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_universe_formula.R
\name{carrier_token_mass}
\alias{carrier_token_mass}
\title{Compute the neutral monoisotopic mass of a (possibly compound) carrier
token. For "H4N" (ammonium) we sum 4\emph{H + N; for "Na-2H" we add Na and
subtract 2}H, etc.}
\usage{
carrier_token_mass(parsed_carrier)
}
\description{
Compute the neutral monoisotopic mass of a (possibly compound) carrier
token. For "H4N" (ammonium) we sum 4\emph{H + N; for "Na-2H" we add Na and
subtract 2}H, etc.
}
\keyword{internal}
