% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_params_small.R
\name{change_params_small}
\alias{change_params_small}
\title{Change Params Small}
\usage{
change_params_small(
  fil_pat = NULL,
  fil_fea_raw = NULL,
  fil_met_raw = NULL,
  fil_sir_raw = NULL,
  fil_spe_raw = NULL,
  ms_pol = NULL,
  org_tax = NULL,
  hig_con = NULL,
  summarize = NULL
)
}
\arguments{
\item{fil_pat}{The pattern identifying your whole job. You can put whatever you want. STRING}

\item{fil_fea_raw}{The path to the file containing your features' intensities. Can be generated by mzmine, SLAW, or SIRIUS. STRING}

\item{fil_met_raw}{The path to the file containing your metadata. If your experiment contains a single taxon, you can provide it below instead. STRING}

\item{fil_sir_raw}{The directory containing the sirius annotations. STRING}

\item{fil_spe_raw}{The path to the file containing your features' spectra. Can contain MS1 and/or MS2 spectra. STRING}

\item{ms_pol}{The polarity used. Must be either "pos" or "neg". STRING}

\item{org_tax}{If your experiment contains a single taxon, its scientific name. "Homo sapiens". STRING}

\item{hig_con}{Filter high confidence candidates only. BOOLEAN}

\item{summarize}{Summarize all candidates per feature to a single row. BOOLEAN}
}
\value{
YAML file with changed parameters.
}
\description{
This function helps changing convenience parameters.
}
\examples{
\dontrun{
copy_backbone()
tima::change_params_small(
  fil_pat = "myExamplePattern",
  fil_fea_raw = "myExampleDir/myExampleFeatures.csv",
  fil_met_raw = "myExampleDir2SomeWhereElse/myOptionalMetadata.tsv",
  fil_sir_raw = "myExampleDir3/myAwesomeSiriusProject.zip",
  fil_spe_raw = "myBeautifulSpectra.mgf",
  ms_pol = "pos",
  org_tax = "Gentiana lutea",
  hig_con = TRUE,
  summarize = FALSE
)
}
}
