% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_collapse.R
\name{clean_collapse}
\alias{clean_collapse}
\title{Clean collapse}
\usage{
clean_collapse(grouped_df, cols = NULL)
}
\arguments{
\item{grouped_df}{Grouped dataframe to be collapsed}

\item{cols}{Character vector of column name(s) to apply collapse to.
If NULL (default), applies to all columns.}
}
\value{
A cleaned and collapsed dataframe with unique values collapsed
per group
}
\description{
This function collapses a grouped dataframe and trims it.
It removes NA values, collapses unique values with " $ " separator,
trims whitespace, and converts empty strings to NA.
}
\examples{
NULL
}
