% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_spectra.R
\name{cleanup_spectra}
\alias{cleanup_spectra}
\title{Cleanup spectra}
\usage{
cleanup_spectra(spectra)
}
\arguments{
\item{spectra}{Spectra object to clean}
}
\value{
Spectra object containing the imported spectra
}
\description{
This function cleans up spectra (keeps only merged MS2)
}
\examples{
NULL
}
