% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_metadata.R
\name{complement_metadata}
\alias{complement_metadata}
\title{Complement metadata}
\usage{
complement_metadata(df)
}
\arguments{
\item{df}{TODO}
}
\value{
TODO
}
\description{
Complement metadata
}
\examples{
TODO
}
