% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_structures_metadata.R
\name{complement_structures_metadata}
\alias{complement_structures_metadata}
\title{Complement structures metadata}
\usage{
complement_structures_metadata(
  df,
  str_2D_3D = paths$data$interim$libraries$merged$structures$dd_ddd,
  str_met = paths$data$interim$libraries$merged$structures$metadata,
  str_nam = paths$data$interim$libraries$merged$structures$names,
  str_tax_cla = paths$data$interim$libraries$merged$structures$taxonomies$classyfire,
  str_tax_npc = paths$data$interim$libraries$merged$structures$taxonomies$npc,
  quickmode = TRUE
)
}
\arguments{
\item{df}{Data frame with structural metadata to be complemented}

\item{str_2D_3D}{File containing 2D and 3D structures}

\item{str_met}{File containing structures metadata}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\value{
Data frame with complemented structural metadata
}
\description{
This function complement structural metadata
}
\examples{
NULL
}
