% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_merged.R
\name{complete_organism_taxonomy}
\alias{complete_organism_taxonomy}
\title{Complete Organism Taxonomy}
\usage{
complete_organism_taxonomy(table_keys, table_org_tax_ott)
}
\description{
Internal helper to fill missing organism taxonomy.
}
\keyword{internal}
