% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{COMPRESSED_EXTENSIONS}
\alias{COMPRESSED_EXTENSIONS}
\title{Supported compressed file extensions}
\format{
An object of class \code{character} of length 3.
}
\usage{
COMPRESSED_EXTENSIONS
}
\description{
Supported compressed file extensions
}
\keyword{datasets}
