% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{ELECTRON_MASS_DALTONS}
\title{Physical and Chemical Constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELECTRON_MASS_DALTONS
}
\description{
This module contains centralized physical and chemical constants
used throughout the TIMA package. Centralizing these values ensures
consistency, facilitates maintenance, and makes the code self-documenting.
}
\keyword{(CODATA}
\keyword{2018}
\keyword{Daltons}
\keyword{Electron}
\keyword{in}
\keyword{internal}
\keyword{mass}
\keyword{recommended}
\keyword{value)}
