% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_collapse.R
\name{convert_lists_to_characters}
\alias{convert_lists_to_characters}
\title{Convert list columns to character}
\usage{
convert_lists_to_characters(df)
}
\description{
Convert list columns to character
}
\keyword{internal}
