% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_backbone.R
\name{copy_backbone}
\alias{copy_backbone}
\title{Copy backbone}
\usage{
copy_backbone(cache_dir = fs::path_home(".tima"), package = "tima")
}
\arguments{
\item{cache_dir}{Cache directory}

\item{package}{Package}
}
\description{
This function copies backbone
}
\examples{
NULL
}
