% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adducts.R
\name{create_adducts}
\alias{create_adducts}
\title{Create adducts}
\usage{
create_adducts(
  masses_table = get("masses_table", envir = parent.frame()),
  adducts_table = get("adducts_table", envir = parent.frame()),
  ms_mode
)
}
\arguments{
\item{masses_table}{Table containing the masses of the adducts}

\item{adducts_table}{Table containing the adducts}

\item{ms_mode}{MS mode}
}
\value{
A tidytable with columns "adduct" and "adduct_mass"
}
\description{
This function creates adducts
}
\examples{
NULL
}
