% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adducts_neg.R
\name{create_adducts_neg}
\alias{create_adducts_neg}
\title{Create adducts negative}
\usage{
create_adducts_neg(massesTable, adductsTable)
}
\arguments{
\item{massesTable}{Table containing the masses of the adducts}

\item{adductsTable}{Table containing the adducts}
}
\description{
This function creates negative adducts
}
\examples{
NULL
}
