% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adducts_pos.R
\name{create_adducts_pos}
\alias{create_adducts_pos}
\title{Create adducts positive}
\usage{
create_adducts_pos(massesTable, adductsTable)
}
\arguments{
\item{massesTable}{Table containing the masses of the adducts}

\item{adductsTable}{Table containing the adducts}
}
\description{
This function creates positive adducts
}
\examples{
NULL
}
