% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_collapse.R
\name{create_collapse_function}
\alias{create_collapse_function}
\title{Create collapse function with separator}
\usage{
create_collapse_function(separator)
}
\description{
Create collapse function with separator
}
\keyword{internal}
