% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_components.R
\name{create_components}
\alias{create_components}
\title{Create components}
\usage{
create_components(
  input = params$files$networks$spectral$edges$processed,
  output = params$files$networks$spectral$components$raw,
  parameters = params
)
}
\arguments{
\item{input}{Input file(s) containing edges}

\item{output}{Output file.}

\item{parameters}{Params}
}
\description{
This function create components from edges
}
\examples{
NULL
}
