% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_parallel.R
\name{create_edges_parallel}
\alias{create_edges_parallel}
\title{Create edges parallel}
\usage{
create_edges_parallel(
  index,
  frags,
  precs,
  nspecs,
  ms2_tolerance,
  ppm_tolerance,
  p = NA,
  parallel
)
}
\arguments{
\item{index}{Indices}

\item{frags}{Fragments}

\item{precs}{Precursors}

\item{nspecs}{Number of spectra}

\item{ms2_tolerance}{MS2 tolerance}

\item{ppm_tolerance}{ppm tolerance}

\item{p}{Progressor}

\item{parallel}{Parallel}
}
\description{
This function is slow so it had to be parallelized
}
\examples{
NULL
}
