% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_progress.R
\name{create_edges_progress}
\alias{create_edges_progress}
\title{Create edges progress}
\usage{
create_edges_progress(
  y,
  id = x,
  s1 = cbind(mz = frags[[x]][, 1], intensity = frags[[x]][, 2])
)
}
\arguments{
\item{y}{Indices of target spectra}

\item{s1}{Query spectrum}

\item{x}{Index of query spectra}
}
\description{
This function is slow so it outputs the progression of the creation of edges
}
\examples{
NULL
}
