% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_smiles.R
\name{create_empty_smiles_template}
\alias{create_empty_smiles_template}
\title{Create empty SMILES result template}
\usage{
create_empty_smiles_template(smiles_colname)
}
\description{
Create empty SMILES result template
}
\keyword{internal}
