% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_spectra.R
\name{create_empty_sop_library}
\alias{create_empty_sop_library}
\title{Create Empty SOP Library Template}
\usage{
create_empty_sop_library()
}
\value{
Data frame with SOP columns, 1 row of NAs.
}
\description{
Internal helper returning an empty SOP structure template.
}
\keyword{internal}
