% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables_sop.R
\name{create_empty_sop_tables}
\alias{create_empty_sop_tables}
\title{Create empty SOP tables structure}
\usage{
create_empty_sop_tables()
}
\description{
Create empty SOP tables structure
}
\keyword{internal}
