% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_spectra.R
\name{create_empty_spectral_library}
\alias{create_empty_spectral_library}
\title{Create Empty Spectral Library Template}
\usage{
create_empty_spectral_library()
}
\value{
Spectra object with one fake spectrum.
}
\description{
Internal helper returning an empty Spectra object with expected
columns when input files are missing.
}
\keyword{internal}
