% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnps_tables.R
\name{create_fake_gnps_files}
\alias{create_fake_gnps_files}
\title{Create Fake GNPS Output Files}
\usage{
create_fake_gnps_files(
  filename,
  path_features,
  path_metadata,
  path_spectra,
  path_interim_a,
  path_interim_f
)
}
\description{
Internal helper to create empty GNPS files for testing.
}
\keyword{internal}
