% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_BATCH_SIZE}
\alias{DEFAULT_BATCH_SIZE}
\title{Default batch size for parallel processing}
\usage{
DEFAULT_BATCH_SIZE
}
\description{
Default batch size for parallel processing
}
\keyword{internal}
