% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_BEST_PERCENTILE}
\alias{DEFAULT_BEST_PERCENTILE}
\title{Default percentile for top candidate selection}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_BEST_PERCENTILE
}
\description{
Default percentile for top candidate selection
}
\keyword{datasets}
