% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_BIO_SCORES}
\alias{DEFAULT_BIO_SCORES}
\title{Default biological score weights by taxonomic rank
Lower ranks (more specific) get higher scores}
\format{
An object of class \code{list} of length 15.
}
\usage{
DEFAULT_BIO_SCORES
}
\description{
Default biological score weights by taxonomic rank
Lower ranks (more specific) get higher scores
}
\keyword{internal}
