% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CANDIDATES_FINAL}
\alias{DEFAULT_CANDIDATES_FINAL}
\title{Default number of final candidates to retain}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_CANDIDATES_FINAL
}
\description{
Default number of final candidates to retain
}
\keyword{datasets}
