% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CHEM_SCORES}
\alias{DEFAULT_CHEM_SCORES}
\title{Default chemical taxonomy score weights}
\format{
An object of class \code{list} of length 7.
}
\usage{
DEFAULT_CHEM_SCORES
}
\description{
Default chemical taxonomy score weights
}
\keyword{internal}
