% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HC_SCORE_BIO_MIN}
\alias{DEFAULT_HC_SCORE_BIO_MIN}
\title{Default minimum biological score threshold for high-confidence filter}
\usage{
DEFAULT_HC_SCORE_BIO_MIN
}
\description{
Default minimum biological score threshold for high-confidence filter
}
\keyword{internal}
