% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HC_SCORE_CHEM_MIN}
\alias{DEFAULT_HC_SCORE_CHEM_MIN}
\title{Minimum chemical score for high confidence annotations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_HC_SCORE_CHEM_MIN
}
\description{
Minimum chemical score for high confidence annotations
}
\keyword{datasets}
