% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HC_SCORE_FINAL_MIN}
\alias{DEFAULT_HC_SCORE_FINAL_MIN}
\title{Default minimum final (weighted chemical) score threshold for high-confidence
filter}
\usage{
DEFAULT_HC_SCORE_FINAL_MIN
}
\description{
Default minimum final (weighted chemical) score threshold for high-confidence
filter
}
\keyword{internal}
