% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HC_SCORE_INITIAL_MIN}
\alias{DEFAULT_HC_SCORE_INITIAL_MIN}
\title{Default minimum initial (pseudo) score threshold for high-confidence filter}
\usage{
DEFAULT_HC_SCORE_INITIAL_MIN
}
\description{
Default minimum initial (pseudo) score threshold for high-confidence filter
}
\keyword{internal}
