% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HC_SCORE_SPECTRAL_MIN}
\alias{DEFAULT_HC_SCORE_SPECTRAL_MIN}
\title{Default minimum spectral similarity score threshold for high-confidence
filter}
\usage{
DEFAULT_HC_SCORE_SPECTRAL_MIN
}
\description{
Default minimum spectral similarity score threshold for high-confidence
filter
}
\keyword{internal}
