% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HE_MAX_RT_ERROR_MIN}
\alias{DEFAULT_HE_MAX_RT_ERROR_MIN}
\title{Default maximum allowed retention time error (minutes) for high-evidence
filter}
\usage{
DEFAULT_HE_MAX_RT_ERROR_MIN
}
\description{
Default maximum allowed retention time error (minutes) for high-evidence
filter
}
\keyword{internal}
