% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HE_SCORE_BIO_MIN}
\alias{DEFAULT_HE_SCORE_BIO_MIN}
\title{Default minimum biological score threshold for high-evidence filter}
\usage{
DEFAULT_HE_SCORE_BIO_MIN
}
\description{
Default minimum biological score threshold for high-evidence filter
}
\keyword{internal}
