% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HE_SCORE_FINAL_COVERAGE_MIN}
\alias{DEFAULT_HE_SCORE_FINAL_COVERAGE_MIN}
\title{Default minimum coverage required when using weighted final score for the
high-evidence filter}
\usage{
DEFAULT_HE_SCORE_FINAL_COVERAGE_MIN
}
\description{
Default minimum coverage required when using weighted final score for the
high-evidence filter
}
\keyword{internal}
