% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DEFAULT_HE_SCORE_SPECTRAL_MIN}
\alias{DEFAULT_HE_SCORE_SPECTRAL_MIN}
\title{Default minimum spectral similarity score threshold for high-evidence
filter}
\usage{
DEFAULT_HE_SCORE_SPECTRAL_MIN
}
\description{
Default minimum spectral similarity score threshold for high-evidence
filter
}
\keyword{internal}
