% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_MINIMAL_MS1_CONDITION}
\alias{DEFAULT_MINIMAL_MS1_CONDITION}
\title{Default MS1 score combination logic for filtering
Either "OR" (at least one threshold met) or "AND" (both thresholds met)}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_MINIMAL_MS1_CONDITION
}
\description{
Default MS1 score combination logic for filtering
Either "OR" (at least one threshold met) or "AND" (both thresholds met)
}
\keyword{internal}
