% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_NUM_WORKERS}
\alias{DEFAULT_NUM_WORKERS}
\title{Default number of parallel workers (NULL = auto-detect)}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
DEFAULT_NUM_WORKERS
}
\description{
Default number of parallel workers (NULL = auto-detect)
}
\keyword{internal}
