% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_TOLERANCE_PPM_MS1}
\alias{DEFAULT_TOLERANCE_PPM_MS1}
\title{Default mass tolerance in parts per million (ppm) for MS1}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_TOLERANCE_PPM_MS1
}
\description{
Default mass tolerance in parts per million (ppm) for MS1
}
\keyword{datasets}
