% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_TOLERANCE_PPM_MS2}
\alias{DEFAULT_TOLERANCE_PPM_MS2}
\title{Default mass tolerance in ppm for MS2}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_TOLERANCE_PPM_MS2
}
\description{
Default mass tolerance in ppm for MS2
}
\keyword{internal}
