% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_TOLERANCE_RT_ADDUCTS}
\alias{DEFAULT_TOLERANCE_RT_ADDUCTS}
\title{Default retention time tolerance in minutes for adduct grouping}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_TOLERANCE_RT_ADDUCTS
}
\description{
Default retention time tolerance in minutes for adduct grouping
}
\keyword{datasets}
