% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_TOLERANCE_RT_LIBRARY}
\alias{DEFAULT_TOLERANCE_RT_LIBRARY}
\title{Default retention time tolerance in minutes for library matching}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_TOLERANCE_RT_LIBRARY
}
\description{
Default retention time tolerance in minutes for library matching
}
\keyword{internal}
