% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_coverage.R
\name{derive_annotate_masses_coverage_path}
\alias{derive_annotate_masses_coverage_path}
\title{Derive the annotate_masses coverage-report file path}
\usage{
derive_annotate_masses_coverage_path(output_annotations)
}
\description{
Derive the annotate_masses coverage-report file path
}
\keyword{internal}
