% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{derive_primary_secondary_annotations}
\alias{derive_primary_secondary_annotations}
\title{Enforce one primary ion species per feature while keeping non-conflicting
secondary alternatives.}
\usage{
derive_primary_secondary_annotations(annotations, baseline_adducts)
}
\description{
Enforce one primary ion species per feature while keeping non-conflicting
secondary alternatives.
}
\keyword{internal}
