% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_collapse.R
\name{determine_columns_to_process}
\alias{determine_columns_to_process}
\title{Determine which columns to process}
\usage{
determine_columns_to_process(df, cols)
}
\description{
Determine which columns to process
}
\keyword{internal}
