% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{discover_annotate_masses_edge_sets}
\alias{discover_annotate_masses_edge_sets}
\title{Discover all edge classes used by annotate_masses}
\usage{
discover_annotate_masses_edge_sets(
  pairs,
  features_table,
  transition_tables,
  adducts,
  evidence_seed_adducts,
  clusters,
  neutral_losses_list,
  ms_mode,
  tolerance_ppm,
  tolerance_rt,
  tolerance_dalton,
  exact_masses
)
}
\description{
Discover all edge classes used by annotate_masses
}
\keyword{internal}
