% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_groups.R
\name{dist_get}
\alias{dist_get}
\title{Get the distance between two elements in a distance matrix}
\usage{
dist_get(d, idx1, idx2)
}
\arguments{
\item{d}{distance matrix}

\item{idx1}{index of the first element}

\item{idx2}{index of the second element}
}
\value{
distance between the two elements
}
\description{
Credit goes to usedist package
}
\examples{
dist_get(d, idx1, idx2)
}
