% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_groups.R
\name{dist_groups}
\alias{dist_groups}
\title{Dist groups}
\usage{
dist_groups(d, g)
}
\arguments{
\item{d}{A distance object or matrix}

\item{g}{A grouping vector for the observations in the distance object.
Must have length equal to the number of observations.}
}
\value{
A data frame containing distance information between pairs of
observations with columns:
\item{Item1}{Index of first observation}
\item{Item2}{Index of second observation}
\item{Group1}{Group label of first observation}
\item{Group2}{Group label of second observation}
\item{Label}{Factor indicating if distance is within or between groups}
\item{Distance}{Numeric distance value rounded to 5 digits}
}
\description{
This function calculates pairwise distances between observations
and annotates them with group membership information
}
\examples{
NULL
}
