% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organism_taxonomy_ott.R
\name{.build_taxonomy_lineage}
\alias{.build_taxonomy_lineage}
\title{Build taxonomy lineage data frame}
\usage{
.build_taxonomy_lineage(taxon_info, taxon_lineage, ott_ids)
}
\arguments{
\item{taxon_info}{List of taxonomy information from rotl}

\item{taxon_lineage}{List of lineage information from rotl}

\item{ott_ids}{Integer vector of OTT IDs}
}
\value{
Data frame with complete taxonomy lineage
}
\description{
Constructs taxonomy lineage from OTT taxon info
}
\keyword{internal}
