% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy_and_similarity.R
\name{.calc_lib_spectrum_similarity}
\alias{.calc_lib_spectrum_similarity}
\title{Calculate similarity for one library spectrum against query}
\usage{
.calc_lib_spectrum_similarity(
  lib_idx,
  lib_spectra_sub,
  lib_precursors_sub,
  current_spectrum,
  current_precursor,
  method,
  dalton,
  ppm
)
}
\description{
Calculate similarity for one library spectrum against query
}
\keyword{internal}
