% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges.R
\name{.calculate_target_similarity}
\alias{.calculate_target_similarity}
\title{Calculate similarity for a target spectrum}
\usage{
.calculate_target_similarity(
  target_idx,
  frags,
  precs,
  method,
  ms2_tolerance,
  ppm_tolerance,
  query_spectrum,
  query_precursor
)
}
\description{
Calculate similarity for a target spectrum
}
\keyword{internal}
