% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy_and_similarity.R
\name{.count_matched_peaks}
\alias{.count_matched_peaks}
\title{Count matched peaks between query and sorted library spectra}
\usage{
.count_matched_peaks(query_mz, lib_mz_sorted, dalton, ppm)
}
\description{
Count matched peaks between query and sorted library spectra
}
\keyword{internal}
