% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organism_taxonomy_ott.R
\name{.create_batches}
\alias{.create_batches}
\title{Create batches from a vector}
\usage{
.create_batches(items, batch_size)
}
\arguments{
\item{items}{Character vector to batch}

\item{batch_size}{Integer size of each batch}
}
\value{
List of character vectors (batches)
}
\description{
Splits a vector into batches of specified size
}
\keyword{internal}
