% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organism_taxonomy_ott.R
\name{.extract_genus_names}
\alias{.extract_genus_names}
\title{Extract genus name from organism name}
\usage{
.extract_genus_names(organism_names)
}
\arguments{
\item{organism_names}{Character vector of organism names}
}
\value{
Character vector of genus names
}
\description{
Extracts the first word (genus) from binomial/trinomial names
}
\keyword{internal}
