% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spectra.R
\name{.extract_intensity_values}
\alias{.extract_intensity_values}
\title{Extract intensity values from peak data matrix}
\usage{
.extract_intensity_values(peakData)
}
\description{
Extract intensity values from peak data matrix
}
\keyword{internal}
