% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spectra.R
\name{.extract_mz_values}
\alias{.extract_mz_values}
\title{Extract m/z values from peak data matrix}
\usage{
.extract_mz_values(peakData)
}
\description{
Extract m/z values from peak data matrix
}
\keyword{internal}
