% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_tables.R
\name{.load_features_file}
\alias{.load_features_file}
\title{Load features file with error handling}
\usage{
.load_features_file(features)
}
\description{
Load features file with error handling
}
\keyword{internal}
