% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.merge_tima_annotations_to_mztab}
\alias{.merge_tima_annotations_to_mztab}
\title{Merge TIMA annotations into mzTab object}
\usage{
.merge_tima_annotations_to_mztab(mztab_obj, tima_annotations)
}
\arguments{
\item{mztab_obj}{mzTab object from RmzTabM or NULL to create new}

\item{tima_annotations}{Data frame with filtered TIMA annotations}
}
\value{
Updated mzTab object
}
\description{
Internal function to merge TIMA annotation results into
mzTab-M Small Molecule table structure.
}
\keyword{internal}
