% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_matching_peaks.R
\name{.ms2_matching_peaks}
\alias{.ms2_matching_peaks}
\title{Calculate the number of matching peaks in two spectra}
\usage{
.ms2_matching_peaks(x, y, ..., cutoff = 0)
}
\arguments{
\item{x}{a matrix with two columns: m/z and intensity values for the first spectrum}

\item{y}{a matrix with two columns: m/z and intensity values for the first spectrum}

\item{...}{...}

\item{cutoff}{a numeric value indicating the minimum intensity required for a peak to be considered}
}
\value{
Integer indicating the number of matching peaks between x and y
}
\description{
TODO
}
\details{
Courtesy of Michele Stravs (ORCID = "0000-0002-1426-8572")
}
\examples{
NULL
}
