% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_matching_peaks.R
\name{.ms2_matching_peaks_fraction}
\alias{.ms2_matching_peaks_fraction}
\title{Calculate the fraction of matching peaks in two spectra}
\usage{
.ms2_matching_peaks_fraction(x, ...)
}
\arguments{
\item{x}{a matrix with two columns: m/z and intensity values for the first spectrum}

\item{...}{...}
}
\value{
Numeric value indicating the fraction of matching peaks between x and y
}
\description{
TODO
}
\examples{
NULL
}
