% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mztab.R
\name{.mztab_assay_sample_map}
\alias{.mztab_assay_sample_map}
\title{Build assay-to-sample-name mapping from metadata}
\usage{
.mztab_assay_sample_map(metadata_df)
}
\description{
Build assay-to-sample-name mapping from metadata
}
\keyword{internal}
