% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_comments.R
\name{.mztab_build_ambiguity_comments}
\alias{.mztab_build_ambiguity_comments}
\title{Build COM lines summarizing ambiguity cardinalities in SML output}
\usage{
.mztab_build_ambiguity_comments(results)
}
\description{
Build COM lines summarizing ambiguity cardinalities in SML output
}
\keyword{internal}
