% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_build_edges_comments}
\alias{.mztab_build_edges_comments}
\title{Build COM lines embedding edges into mzTab text exports}
\usage{
.mztab_build_edges_comments(edges_file, chunk_size = 500L)
}
\description{
Build COM lines embedding edges into mzTab text exports
}
\keyword{internal}
