% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_build_sme}
\alias{.mztab_build_sme}
\title{Build SME (Small Molecule Evidence) table – one row per evidence}
\usage{
.mztab_build_sme(results, smf_table, xrefs_index = NULL)
}
\description{
Build SME (Small Molecule Evidence) table – one row per evidence
}
\keyword{internal}
