% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_build_smf}
\alias{.mztab_build_smf}
\title{Build SMF (Small Molecule Feature) table – one row per feature_id}
\usage{
.mztab_build_smf(results)
}
\description{
Build SMF (Small Molecule Feature) table – one row per feature_id
}
\keyword{internal}
