% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_comments.R
\name{.mztab_edges_json_rows}
\alias{.mztab_edges_json_rows}
\title{Build JSON edge rows for mzTab JSON export}
\usage{
.mztab_edges_json_rows(edges_file)
}
\description{
Build JSON edge rows for mzTab JSON export
}
\keyword{internal}
