% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_ensure_cols}
\alias{.mztab_ensure_cols}
\title{Ensure columns exist and are character}
\usage{
.mztab_ensure_cols(tbl, cols)
}
\description{
Ensure columns exist and are character
}
\keyword{internal}
