% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_merge_sme_rows}
\alias{.mztab_merge_sme_rows}
\title{Merge SME rows and return id map from new IDs to final IDs}
\usage{
.mztab_merge_sme_rows(sme_base, sme_new, smf_id_map)
}
\description{
Merge SME rows and return id map from new IDs to final IDs
}
\keyword{internal}
