% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_merge_smf_rows}
\alias{.mztab_merge_smf_rows}
\title{Merge SMF rows and return id map from new IDs to final IDs}
\usage{
.mztab_merge_smf_rows(smf_base, smf_new)
}
\description{
Merge SMF rows and return id map from new IDs to final IDs
}
\keyword{internal}
