% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_merge.R
\name{.mztab_merge_sml_rows}
\alias{.mztab_merge_sml_rows}
\title{Merge SML rows and enrich references for duplicates}
\usage{
.mztab_merge_sml_rows(sml_base, sml_new, smf_id_map, sme_id_map)
}
\description{
Merge SML rows and enrich references for duplicates
}
\keyword{internal}
