% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_merge.R
\name{.mztab_merge_tables}
\alias{.mztab_merge_tables}
\title{Merge TIMA-derived tables into a base mzTab}
\usage{
.mztab_merge_tables(base, sml_new, smf_new, sme_new)
}
\description{
Merge TIMA-derived tables into a base mzTab
}
\keyword{internal}
