% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_merge.R
\name{.mztab_next_numeric_id}
\alias{.mztab_next_numeric_id}
\title{Find next numeric ID from an ID column}
\usage{
.mztab_next_numeric_id(ids)
}
\description{
Find next numeric ID from an ID column
}
\keyword{internal}
