% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_metadata.R
\name{.mztab_opt_colname}
\alias{.mztab_opt_colname}
\title{Convert a source column name into an mzTab optional column name.}
\usage{
.mztab_opt_colname(x)
}
\description{
Convert a source column name into an mzTab optional column name.
}
\keyword{internal}
