% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_tables.R
\name{.mztab_pluck}
\alias{.mztab_pluck}
\title{Safely extract a column or return "null" vector (vectorized utility)}
\usage{
.mztab_pluck(df, col)
}
\description{
Safely extract a column or return "null" vector (vectorized utility)
}
\keyword{internal}
