% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_pluck_na}
\alias{.mztab_pluck_na}
\title{Safely extract a column, replacing NA/null with "null" (vectorized utility)}
\usage{
.mztab_pluck_na(df, col)
}
\description{
Safely extract a column, replacing NA/null with "null" (vectorized utility)
}
\keyword{internal}
