% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_merge.R
\name{.mztab_refresh_smf_sme_refs}
\alias{.mztab_refresh_smf_sme_refs}
\title{Refresh SMF SME_ID_REFS and ambiguity code from SME definitions}
\usage{
.mztab_refresh_smf_sme_refs(smf, sme)
}
\description{
Refresh SMF SME_ID_REFS and ambiguity code from SME definitions
}
\keyword{internal}
