% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_refresh_smf_sml_refs}
\alias{.mztab_refresh_smf_sml_refs}
\title{Refresh SMF SML_ID_REFS from SML definitions}
\usage{
.mztab_refresh_smf_sml_refs(smf, sml)
}
\description{
Refresh SMF SML_ID_REFS from SML definitions
}
\keyword{internal}
