% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_remap_ref_ids}
\alias{.mztab_remap_ref_ids}
\title{Remap pipe-separated references according to id map}
\usage{
.mztab_remap_ref_ids(x, id_map)
}
\description{
Remap pipe-separated references according to id map
}
\keyword{internal}
