% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_split_ref_ids}
\alias{.mztab_split_ref_ids}
\title{Split reference list encoded with pipe separator}
\usage{
.mztab_split_ref_ids(x)
}
\description{
Split reference list encoded with pipe separator
}
\keyword{internal}
