% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_union_ref_ids}
\alias{.mztab_union_ref_ids}
\title{Union two pipe-separated reference lists}
\usage{
.mztab_union_ref_ids(a, b)
}
\description{
Union two pipe-separated reference lists
}
\keyword{internal}
