% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab_metadata.R
\name{.mztab_write_section}
\alias{.mztab_write_section}
\title{Write a single mzTab-M section (header + data rows)}
\usage{
.mztab_write_section(con, header_prefix, row_prefix, tbl)
}
\description{
Write a single mzTab-M section (header + data rows)
}
\keyword{internal}
