% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_tables.R
\name{.normalize_rt_column}
\alias{.normalize_rt_column}
\title{Normalize RT column to minutes (auto-detect unit if possible)}
\usage{
.normalize_rt_column(tbl, name_rt)
}
\description{
Normalize RT column to minutes (auto-detect unit if possible)
}
\keyword{internal}
