% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy_and_similarity.R
\name{.process_query_against_library}
\alias{.process_query_against_library}
\title{Process one query spectrum against library}
\usage{
.process_query_against_library(
  spectrum_idx,
  query_spectra,
  query_precursors,
  query_ids,
  lib_spectra,
  lib_precursors,
  lib_ids,
  approx,
  dalton,
  ppm,
  method,
  threshold
)
}
\description{
Process one query spectrum against library
}
\keyword{internal}
