% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges.R
\name{.process_query_spectrum}
\alias{.process_query_spectrum}
\title{Process query spectrum against all subsequent spectra}
\usage{
.process_query_spectrum(
  index,
  nspecs,
  frags,
  precs,
  method,
  ms2_tolerance,
  ppm_tolerance,
  threshold,
  matched_peaks
)
}
\description{
Process query spectrum against all subsequent spectra
}
\keyword{internal}
