% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_tables.R
\name{.select_intensity_columns}
\alias{.select_intensity_columns}
\title{Select intensity columns and metadata (auto-detect format)}
\usage{
.select_intensity_columns(tbl, name_features, name_rt, name_mz, name_adduct)
}
\description{
Select intensity columns and metadata (auto-detect format)
}
\keyword{internal}
