% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_tables.R
\name{.standardize_column_names}
\alias{.standardize_column_names}
\title{Standardize column names by removing format-specific prefixes/suffixes}
\usage{
.standardize_column_names(tbl)
}
\description{
Standardize column names by removing format-specific prefixes/suffixes
}
\keyword{internal}
