% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.write_mztab_file}
\alias{.write_mztab_file}
\title{Write mzTab object to file}
\usage{
.write_mztab_file(mztab_obj, output_file)
}
\arguments{
\item{mztab_obj}{mzTab object}

\item{output_file}{Character string path for output}
}
\value{
Invisible NULL
}
\description{
Internal function to write mzTab object to file in
mzTab-M format.
}
\keyword{internal}
