% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_files.R
\name{download_example_files_impl}
\alias{download_example_files_impl}
\title{Download selected example files using injected dependencies}
\usage{
download_example_files_impl(
  example,
  in_cache,
  paths,
  download_file,
  download_sirius,
  enter_cache
)
}
\description{
Download selected example files using injected dependencies
}
\keyword{internal}
