% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download file}
\usage{
download_file(url, export, limit = 600)
}
\arguments{
\item{url}{URL of the file to be downloaded}

\item{export}{File path where the file should be saved}

\item{limit}{Timeout limit (in seconds)}
}
\description{
This function download files
}
\examples{
download_file(url = "https://github.com/taxonomicallyinformedannotation/tima-example-files/raw/main/metadata.tsv", export = "data/source/examples/metadata.tsv")
}
