% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnps_tables.R
\name{download_gnps_file}
\alias{download_gnps_file}
\title{Download GNPS File}
\usage{
download_gnps_file(url, file_path, file_type)
}
\description{
Internal helper to download a single GNPS file.
}
\keyword{internal}
