% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{download_with_error_handling}
\alias{download_with_error_handling}
\title{Download file with error handling}
\usage{
download_with_error_handling(url, export)
}
\description{
Download file with error handling
}
\keyword{internal}
