% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{ELECTRON_MASS_DALTONS}
\alias{ELECTRON_MASS_DALTONS}
\title{Electron mass in Daltons (CODATA 2018 recommended value)}
\usage{
ELECTRON_MASS_DALTONS
}
\description{
Electron mass in Daltons (CODATA 2018 recommended value)
}
\keyword{internal}
