% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_evidence.R
\name{empty_evidence_table}
\alias{empty_evidence_table}
\title{Empty hypothesis table with the canonical schema.}
\usage{
empty_evidence_table()
}
\description{
Empty hypothesis table with the canonical schema.
}
\keyword{internal}
