% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R, R/mass_consistency.R
\name{enforce_graph_adduct_consistency}
\alias{enforce_graph_adduct_consistency}
\title{Enforce graph-level adduct consistency on edge hypotheses}
\usage{
enforce_graph_adduct_consistency(df_add)

enforce_graph_adduct_consistency(df_add)
}
\description{
Enforce graph-level adduct consistency on edge hypotheses

Enforce graph-level adduct consistency on edge hypotheses
}
\keyword{internal}
