% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{enrich_with_structure_metadata}
\alias{enrich_with_structure_metadata}
\title{Decorate annotation rows with structural / taxonomic metadata}
\usage{
enrich_with_structure_metadata(combined, structure_table)
}
\description{
Decorate annotation rows with structural / taxonomic metadata
}
\keyword{internal}
