% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{ensure_directory_exists}
\alias{ensure_directory_exists}
\title{Ensure directory exists, creating it if necessary}
\usage{
ensure_directory_exists(dir_path)
}
\description{
Ensure directory exists, creating it if necessary
}
\keyword{internal}
