% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs_utils.R
\name{eval_glue_safely}
\alias{eval_glue_safely}
\title{Evaluate glue-style formatting safely}
\usage{
eval_glue_safely(...)
}
\arguments{
\item{...}{Message parts to combine}
}
\value{
Character string with glue-style interpolation applied
}
\description{
Evaluate glue-style formatting safely
}
\keyword{internal}
