% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{expand_with_modifier}
\alias{expand_with_modifier}
\title{Add +mod or -mod suffix to existing adduct strings on selected features}
\usage{
expand_with_modifier(
  base_hyps,
  edges,
  edge_feature_col,
  mod_col,
  mod_sign,
  strip_label = FALSE
)
}
\description{
Add +mod or -mod suffix to existing adduct strings on selected features
}
\keyword{internal}
