% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_output.R
\name{export_output}
\alias{export_output}
\title{Export output}
\usage{
export_output(x, file)
}
\arguments{
\item{x}{Data frame to be exported}

\item{file}{Character string path to the output file}
}
\value{
The path to the exported file
}
\description{
This function creates the output directory if it doesn't exist
and exports a data frame to a tab-delimited file
}
\examples{
\dontrun{
export_output(x = data.frame(), file = "output/file.tsv")
unlink("output", recursive = TRUE)
}
}
