% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_spectra.R
\name{export_spectra}
\alias{export_spectra}
\title{Export spectra}
\usage{
export_spectra(file, spectra, dir = ".", cmps = NULL, metad = NULL)
}
\arguments{
\item{file}{TODO}

\item{spectra}{TODO}

\item{dir}{TODO}

\item{cmps}{TODO}

\item{metad}{TODO}
}
\value{
TODO
}
\description{
Export spectra
}
\examples{
TODO
}
