% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_spectra.R
\name{export_spectra_2}
\alias{export_spectra_2}
\title{Export spectra 2}
\usage{
export_spectra_2(file, spectra, meta)
}
\arguments{
\item{file}{TODO}

\item{spectra}{TODO}

\item{meta}{TODO}
}
\value{
TODO
}
\description{
Export spectra 2
}
\examples{
TODO
}
