% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_spectra_rds.R
\name{export_spectra_rds}
\alias{export_spectra_rds}
\title{Export spectra RDS}
\usage{
export_spectra_rds(file, spectra)
}
\arguments{
\item{file}{File where spectra will be exported.}

\item{spectra}{The spectra object where spectra are stored}
}
\description{
This function exports spectra in RDS
}
\examples{
NULL
}
