% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{extract_charge_info}
\alias{extract_charge_info}
\title{Extract charge information from regex matches}
\usage{
extract_charge_info(charge_count_match, charge_sign_match)
}
\description{
Extract charge information from regex matches
}
\keyword{internal}
