% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{extract_directory_path}
\alias{extract_directory_path}
\title{Extract directory path from file or directory path}
\usage{
extract_directory_path(path)
}
\description{
Extract directory path from file or directory path
}
\keyword{internal}
