% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{extract_isotope_shift}
\alias{extract_isotope_shift}
\title{Extract isotope shift from regex match}
\usage{
extract_isotope_shift(match_value)
}
\description{
Extract isotope shift from regex match
}
\keyword{internal}
