% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{extract_multimer_count}
\alias{extract_multimer_count}
\title{Extract multimer count from regex match}
\usage{
extract_multimer_count(match_value)
}
\description{
Extract multimer count from regex match
}
\keyword{internal}
