% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{extract_preassigned_adducts}
\alias{extract_preassigned_adducts}
\title{Extract any non-NA pre-assigned adducts as explicit hypotheses}
\usage{
extract_preassigned_adducts(features_table)
}
\description{
Extract any non-NA pre-assigned adducts as explicit hypotheses
}
\keyword{internal}
