% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_smiles.R
\name{extract_unique_smiles}
\alias{extract_unique_smiles}
\title{Extract unique SMILES from dataframe}
\usage{
extract_unique_smiles(df, smiles_colname)
}
\description{
Extract unique SMILES from dataframe
}
\keyword{internal}
