% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_annotations_columns.R
\name{fake_annotations_columns}
\alias{fake_annotations_columns}
\title{Fake annotations columns}
\usage{
fake_annotations_columns()
}
\description{
This function fakes annotations columns
}
\examples{
NULL
}
