% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_ecmdb.R
\name{fake_ecmdb}
\alias{fake_ecmdb}
\title{Fake ECMDB}
\usage{
fake_ecmdb(export)
}
\arguments{
\item{export}{Path to save the file to}
}
\description{
This function fakes ECMDB in case the download failed
}
\examples{
NULL
}
