% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_edges.R
\name{fake_edges}
\alias{fake_edges}
\title{Title}
\usage{
fake_edges(
  input = params$input,
  output = params$output,
  name_source = params$source_name,
  name_target = params$target_name
)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}

\item{name_source}{TODO}

\item{name_target}{TODO}
}
\value{
TODO
}
\description{
Title
}
