% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_features_edges.R
\name{fake_features_edges}
\alias{fake_features_edges}
\title{Fake edges}
\usage{
fake_features_edges(
  input = params$files$features$raw,
  output = params$files$networks$spectral$edges$processed,
  name_feature = params$names$features
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{name_feature}{Name of the features column}
}
\description{
This function fakes edges in case none are given
}
\examples{
NULL
}
