% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_hmdb.R
\name{fake_hmdb}
\alias{fake_hmdb}
\title{Fake HMDB}
\usage{
fake_hmdb(export)
}
\arguments{
\item{export}{Path to save the file to}
}
\description{
This function fakes HMDB in case the download failed
}
\examples{
NULL
}
