% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_lotus.R
\name{fake_lotus}
\alias{fake_lotus}
\title{Fake LOTUS}
\usage{
fake_lotus(export)
}
\arguments{
\item{export}{Path to save the file to}
}
\description{
This function fakes LOTUS in case the download failed
}
\examples{
NULL
}
