% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_high_confidence_only.R
\name{filter_high_confidence_only}
\alias{filter_high_confidence_only}
\title{Filter high confidence only
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
filter_high_confidence_only(
  df,
  score_bio_min = 0.9,
  score_ini_min = 0.9,
  score_final_min = 0.45
)
}
\arguments{
\item{df}{Dataframe}

\item{score_bio_min}{Minimal biological score. Current default to 0.9.}

\item{score_ini_min}{Minimal initial score. Current default to 0.9.}

\item{score_final_min}{Minimal final score. Current default to 0.45.}
}
\value{
A dataframe containing only high confidence annotations
}
\description{
This function filters highly confident annotations only.
}
\examples{
NULL
}
