% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nitrogen_rule.R
\name{filter_nitrogen_rule}
\alias{filter_nitrogen_rule}
\title{Filter nitrogen rule}
\usage{
filter_nitrogen_rule(df_annotated_final, features_table, filter_nitro)
}
\arguments{
\item{df_annotated_final}{Table containing your MS1 annotations}

\item{features_table}{Feature table containing your mzs}

\item{filter_nitro}{Filter according to Nitrogen rule. Boolean}
}
\value{
A table containing filtered MS1 annotations
}
\description{
This function filters annotations according to nitrogen rule
}
\examples{
NULL
}
