% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_reals.R
\name{find_matching_columns}
\alias{find_matching_columns}
\title{Find columns matching any pattern}
\usage{
find_matching_columns(column_names, patterns)
}
\description{
Find columns matching any pattern
}
\keyword{internal}
