% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_adduct_neg.R
\name{form_adducts_neg}
\alias{form_adducts_neg}
\title{Title}
\usage{
form_adducts_neg(massesTable, adductsTable)
}
\arguments{
\item{massesTable}{TODO}

\item{adductsTable}{TODO}
}
\value{
TODO
}
\description{
Title
}
