% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_adducts_pos.R
\name{form_adducts_pos}
\alias{form_adducts_pos}
\title{Form adducts positive}
\usage{
form_adducts_pos(massesTable, adductsTable)
}
\arguments{
\item{massesTable}{TODO}

\item{adductsTable}{TODO}
}
\value{
TODO
}
\description{
Form adducts positive
}
\examples{
TODO
}
