% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_helpers.R
\name{format_percent}
\alias{format_percent}
\title{Format percentage}
\usage{
format_percent(x, digits = 1)
}
\arguments{
\item{x}{Numeric value between 0 and 1}

\item{digits}{Number of decimal places}
}
\value{
Formatted string (e.g., "75.5\%")
}
\description{
Format percentage
}
\keyword{internal}
