% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_universe_formula.R
\name{formula_mass}
\alias{formula_mass}
\title{Compute the monoisotopic mass of a parsed atomic-formula vector.}
\usage{
formula_mass(formula_vec)
}
\description{
Compute the monoisotopic mass of a parsed atomic-formula vector.
}
\keyword{internal}
