% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_enforcement.R
\name{formula_satisfies_loss_requirements}
\alias{formula_satisfies_loss_requirements}
\title{Check whether formula contains all atoms required by adduct losses}
\usage{
formula_satisfies_loss_requirements(formula, loss_requirements)
}
\description{
Check whether formula contains all atoms required by adduct losses
}
\keyword{internal}
