% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark.R
\name{get_benchmark}
\alias{get_benchmark}
\title{Get benchmark}
\usage{
get_benchmark(
  url = paths$urls$benchmarking_set,
  export = paths$data$source$benchmark$set
)
}
\arguments{
\item{url}{a character string containing the URL of the benchmarking set}

\item{export}{a character string containing the file path where the benchmarking set should be exported}
}
\description{
Get benchmark
}
\examples{
get_benchmark(url = "https://myurl.com/benchmark.csv", export = "data/source/benchmark/set.csv")
}
