% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constant}
\alias{get_constant}
\title{Get a constant value by name with validation}
\usage{
get_constant(name, default = NULL)
}
\arguments{
\item{name}{Character string name of the constant}

\item{default}{Default value to return if constant not found}
}
\value{
The constant value
}
\description{
Get a constant value by name with validation
}
\keyword{internal}
