% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_feature_table.R
\name{get_example_feature_table}
\alias{get_example_feature_table}
\title{Get example feature table}
\usage{
get_example_feature_table(
  url = paths$urls$example_feature_table,
  export = paths$data$interim$annotations$example_feature_table
)
}
\arguments{
\item{url}{the URL of the example feature table file}

\item{export}{the file path to where the example feature table should be saved}
}
\description{
Get example feature table
}
\examples{
get_example_feature_table(url = "https://myurl.com/features.csv", export = "directory/table.tsv")
}
