% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_features.R
\name{get_example_features}
\alias{get_example_features}
\title{Get example features}
\usage{
get_example_features(
  url = paths$urls$examples$features,
  export = paths$data$source$examples$features
)
}
\arguments{
\item{url}{URL of the example feature table file}

\item{export}{File path to where the example feature table should be saved}
}
\description{
This function gets an example feature table to work with
}
\examples{
NULL
}
