% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_files.R
\name{get_example_files}
\alias{get_example_files}
\title{Get example files}
\usage{
get_example_files()
}
\value{
Example files.
}
\description{
This function downloads example files
}
\examples{
NULL
}
