% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_metadata.R
\name{get_example_metadata}
\alias{get_example_metadata}
\title{Get example metadata}
\usage{
get_example_metadata(
  url = paths$urls$examples$metadata,
  export = paths$data$source$examples$metadata
)
}
\arguments{
\item{url}{URL of the example metadata table file}

\item{export}{File path to where the example metadata table should be saved}
}
\description{
This function gets an example metadata table to work with
}
\examples{
NULL
}
