% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_spectra.R
\name{get_example_spectra}
\alias{get_example_spectra}
\title{Get example spectra}
\usage{
get_example_spectra(
  url = paths$url$examples$mgf,
  export = paths$data$source$examples$spectra
)
}
\arguments{
\item{url}{URL of the MGF file to be downloaded}

\item{export}{File path where the MGF file should be saved}
}
\description{
This function gets an example mgf file to work with
}
\examples{
NULL
}
