% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnps_tables.R
\name{get_gnps_tables}
\alias{get_gnps_tables}
\title{Get GNPS Tables}
\usage{
get_gnps_tables(
  gnps_job_id,
  workflow = "fbmn",
  path_source = paths$data$source$path,
  path_interim_a = paths$data$interim$annotations$path,
  path_interim_f = paths$data$interim$features$path
)
}
\arguments{
\item{gnps_job_id}{GNPS job ID}

\item{workflow}{Character string indicating the type of workflow, either "fbmn" or "classical"}

\item{path_source}{Path to store the source files}

\item{path_interim_a}{Path to store the interim annotations file}

\item{path_interim_f}{Path to store the interim features files}
}
\value{
The downloaded GNPS tables
}
\description{
This function gets GNPS tables from corresponding job ID.
}
\examples{
NULL
}
