% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isdb.R
\name{get_isdb}
\alias{get_isdb}
\title{Title}
\usage{
get_isdb(
  url_pos = paths$urls$isdb$pos,
  url_neg = paths$urls$isdb$neg,
  export_pos = paths$data$source$libraries$isdb$pos,
  export_neg = paths$data$source$libraries$isdb$neg
)
}
\arguments{
\item{url_pos}{TODO}

\item{url_neg}{TODO}

\item{export_pos}{TODO}

\item{export_neg}{TODO}
}
\value{
TODO
}
\description{
Title
}
