% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_version_from_zenodo.R
\name{get_last_version_from_zenodo}
\alias{get_last_version_from_zenodo}
\title{Get last version from Zenodo}
\usage{
get_last_version_from_zenodo(doi, pattern, path)
}
\arguments{
\item{doi}{TODO}

\item{pattern}{TODO}

\item{path}{TODO}
}
\value{
TODO
}
\description{
Credit goes to partially to https://inbo.github.io/inborutils/
}
\examples{
TODO
}
