% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_massbank_spectra.R
\name{get_massbank_spectra}
\alias{get_massbank_spectra}
\title{Get MassBank spectra}
\usage{
get_massbank_spectra(
  output_dir = "data/source/libraries/spectra/exp",
  mb_file = parse_yaml_paths()$urls$massbank$file,
  mb_url = parse_yaml_paths()$urls$massbank$url,
  mb_version = parse_yaml_paths()$urls$massbank$version
)
}
\arguments{
\item{output_dir}{Output where to store the spectra}

\item{mb_file}{MassBank file}

\item{mb_url}{MassBank URL}

\item{mb_version}{MassBank version}
}
\description{
This function gets MassBank spectra
}
\examples{
NULL
}
