% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organism_taxonomy_ott.R
\name{get_organism_taxonomy_ott}
\alias{get_organism_taxonomy_ott}
\title{Get organism taxonomy (Open Tree of Life Taxonomy)}
\usage{
get_organism_taxonomy_ott(df)
}
\arguments{
\item{df}{Dataframe containing your organism(s) name(s)}
}
\description{
This function retrieves taxonomy from the Open Tree of Life taxonomy
}
\examples{
NULL
}
