% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chemo.R
\name{get_taxonomy_weights}
\alias{get_taxonomy_weights}
\title{Extract Weighted Taxonomy Scores}
\usage{
get_taxonomy_weights()
}
\value{
Named list of weights
}
\description{
Internal helper to extract level weights from parent environment
(matching weight_chemo parameters).
}
\keyword{internal}
