% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wrappers.R
\name{gnps_chain_dp_batch_wrapper}
\alias{gnps_chain_dp_batch_wrapper}
\title{Wrapper for the C function "gnps_chain_dp_batch"}
\usage{
gnps_chain_dp_batch_wrapper(
  x,
  xPrecursorMz,
  y_list,
  yPrecursorMz,
  tolerance,
  ppm
)
}
\description{
Wrapper for the C function "gnps_chain_dp_batch"
}
\keyword{internal}
