% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_id.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator (like \verb{??} in other languages)}
\usage{
a \%||\% b
}
\description{
Null-coalescing operator (like \verb{??} in other languages)
}
\keyword{internal}
