% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_names_sirius.R
\name{harmonize_names_sirius}
\alias{harmonize_names_sirius}
\title{Harmonize names sirius}
\usage{
harmonize_names_sirius(x)
}
\arguments{
\item{x}{Character string containing a name}
}
\value{
Character string with the name modified according
to the rules specified in the function
}
\description{
This function harmonizes the names of Sirius
outputs to make them compatible
}
\examples{
\dontrun{
harmonized_name <- harmonize_names_sirius("My_name")
}
}
