% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_spectra.R
\name{harmonize_spectra_polarity}
\alias{harmonize_spectra_polarity}
\title{Harmonize Spectra for a Given Polarity}
\usage{
harmonize_spectra_polarity(
  spectra_extracted,
  mode,
  metad,
  col_ad,
  col_ce,
  col_ci,
  col_em,
  col_in,
  col_io,
  col_ik,
  col_il,
  col_mf,
  col_na,
  col_po,
  col_sm,
  col_sn,
  col_si,
  col_sp,
  col_sy,
  col_xl
)
}
\arguments{
\item{spectra_extracted}{\link{list} List of extracted spectra data frames.}

\item{mode}{\link{character} Character "pos" or "neg".}

\item{metad}{\link{character} Character library metadata name.}

\item{col_ad}{\link{character} Adduct field name}

\item{col_ce}{\link{character} Collision energy field name}

\item{col_ci}{\link{character} Compound ID field name}

\item{col_em}{\link{character} Exact mass field name}

\item{col_in}{\link{character} InChI field name}

\item{col_io}{\link{character} InChI without stereochemistry field name}

\item{col_ik}{\link{character} InChIKey field name}

\item{col_il}{\link{character} InChIKey connectivity layer field name}

\item{col_mf}{\link{character} Molecular formula field name}

\item{col_na}{\link{character} Compound name field name}

\item{col_po}{\link{character} Polarity field name}

\item{col_sm}{\link{character} SMILES field name}

\item{col_sn}{\link{character} SMILES without stereochemistry field name}

\item{col_si}{\link{character} Spectrum ID field name}

\item{col_sp}{\link{character} SPLASH field name}

\item{col_sy}{\link{character} Synonyms field name}

\item{col_xl}{\link{character} xLogP field name}
}
\value{
Data frame with harmonized spectra.
}
\description{
Internal helper to harmonize extracted spectra for a specific
polarity mode, applying column mappings and fixing precursor/inchikey issues.
}
\keyword{internal}
