% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_latest_version.R
\name{install_latest_version}
\alias{install_latest_version}
\title{Install latest version}
\usage{
install_latest_version(test = FALSE)
}
\arguments{
\item{test}{Flag for tests}
}
\description{
This function installs the latest version
}
\examples{
NULL
}
