% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_version_from_zenodo.R
\name{is_download_needed}
\alias{is_download_needed}
\title{Check if File Download is Needed}
\usage{
is_download_needed(path, zenodo_size)
}
\arguments{
\item{path}{Character local file path}

\item{zenodo_size}{Numeric size in bytes from Zenodo}
}
\value{
Logical TRUE if download needed, FALSE otherwise
}
\description{
Internal helper to determine if download is required by comparing sizes.
}
\keyword{internal}
