% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{is_valid_binary_format}
\alias{is_valid_binary_format}
\title{Check if header indicates valid binary format}
\usage{
is_valid_binary_format(header_raw)
}
\description{
Check if header indicates valid binary format
}
\keyword{internal}
