% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ISOTOPE_MASS_SHIFT_DALTONS}
\alias{ISOTOPE_MASS_SHIFT_DALTONS}
\title{Isotope mass shift in Daltons}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ISOTOPE_MASS_SHIFT_DALTONS
}
\description{
Represents the average mass difference between isotopes,
primarily the 13C - 12C difference (1.0033548 Da). This is used for
calculating masses of isotopically labelled compounds (M+1, M+2, etc.).
The value is very close to the neutron-proton mass difference.
}
\keyword{internal}
