% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wrappers.R
\name{join_gnps_wrapper}
\alias{join_gnps_wrapper}
\title{Wrapper for the C function "join_gnps"}
\usage{
join_gnps_wrapper(x, y, xPrecursorMz, yPrecursorMz, tolerance, ppm)
}
\arguments{
\item{x}{Numeric vector or matrix for query masses.}

\item{y}{Numeric vector or matrix for target masses.}

\item{xPrecursorMz}{Numeric vector of precursor values for queries.}

\item{yPrecursorMz}{Numeric vector of precursor values for targets.}

\item{tolerance}{Numeric value specifying the tolerance in daltons.}

\item{ppm}{Numeric value specifying the tolerance in ppm.}
}
\value{
The result returned from the C function "join_gnps".
}
\description{
Wrapper for the C function "join_gnps"
}
