% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{join_multi_with_addlossed}
\alias{join_multi_with_addlossed}
\title{Join multi-adduct candidates with add/loss inferred masses}
\usage{
join_multi_with_addlossed(df_multi, df_addlossed_rdy)
}
\description{
Join multi-adduct candidates with add/loss inferred masses
}
\keyword{internal}
