% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_peaks.R
\name{keep_peaks}
\alias{keep_peaks}
\title{Keep peaks}
\usage{
keep_peaks(x, prop)
}
\arguments{
\item{x}{PeakData}

\item{prop}{Minimal ratio to the max peak}
}
\description{
This function keeps a proportion of peaks
}
\examples{
NULL
}
