% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_mandatory.R
\name{label_mandatory}
\alias{label_mandatory}
\title{Label mandatory}
\usage{
label_mandatory(label)
}
\arguments{
\item{label}{Label}
}
\description{
This function adds an asterisk to mandatory inputs
}
\examples{
NULL
}
