% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_merged.R
\name{load_and_merge_libraries}
\alias{load_and_merge_libraries}
\title{Load and Merge Library Files}
\usage{
load_and_merge_libraries(files, cache)
}
\description{
Internal helper to load and combine library files.
}
\keyword{internal}
