% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{load_annotate_masses_runtime_inputs}
\alias{load_annotate_masses_runtime_inputs}
\title{Load prepared features and structural library for annotate_masses}
\usage{
load_annotate_masses_runtime_inputs(
  features_table,
  tolerance_rt,
  baseline_adduct,
  library,
  str_stereo,
  str_met,
  str_tax_cla,
  str_tax_npc,
  tolerance_ppm
)
}
\description{
Load prepared features and structural library for annotate_masses
}
\keyword{internal}
