% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_smiles.R
\name{load_smiles_cache}
\alias{load_smiles_cache}
\title{Load SMILES cache or create empty template}
\usage{
load_smiles_cache(cache, smiles_colname)
}
\description{
Load SMILES cache or create empty template
}
\keyword{internal}
