% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs_utils.R
\name{log_complete}
\alias{log_complete}
\title{Log operation completion}
\usage{
log_complete(ctx, ...)
}
\arguments{
\item{ctx}{tima_log_context object from log_operation()}

\item{...}{Named results to log}
}
\value{
The input context (invisibly)
}
\description{
Logs the successful completion of an operation with results and timing.
}
