% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{log_consistency_audit}
\alias{log_consistency_audit}
\title{Log the consistency-audit attribute attached by
\code{enforce_graph_adduct_consistency()}.}
\usage{
log_consistency_audit(audit)
}
\description{
Log the consistency-audit attribute attached by
\code{enforce_graph_adduct_consistency()}.
}
\keyword{internal}
