% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_debug.R
\name{log_debug}
\alias{log_debug}
\title{Log debug}
\usage{
log_debug(...)
}
\arguments{
\item{...}{one or more values to be logged}
}
\description{
Log debug
}
\examples{
log_debug("This is a debug message")
}
