% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LOG_LEVELS}
\alias{LOG_LEVELS}
\title{Log levels available (in order of severity)}
\format{
An object of class \code{character} of length 6.
}
\usage{
LOG_LEVELS
}
\description{
Log levels available (in order of severity)
}
\keyword{internal}
