% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs_utils.R
\name{log_metadata}
\alias{log_metadata}
\title{Log intermediate operation metadata}
\usage{
log_metadata(ctx, ...)
}
\arguments{
\item{ctx}{tima_log_context object from log_operation()}

\item{...}{Named metadata to add and log}
}
\value{
The updated context (invisibly)
}
\description{
Adds metadata to an operation context and logs it.
}
