% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_pipe.R
\name{log_pipe}
\alias{log_pipe}
\title{Log pipe}
\usage{
log_pipe(x, ...)
}
\arguments{
\item{x}{value for the pipe}

\item{...}{one or more values to be logged}
}
\value{
Message for debugging
}
\description{
Simple helper for debugging between pipes
}
\examples{
NULL
}
