% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs_utils.R
\name{logging_wrappers}
\alias{logging_wrappers}
\alias{log_trace}
\alias{log_debug}
\alias{log_info}
\alias{log_warn}
\alias{log_error}
\alias{log_fatal}
\alias{log_success}
\title{Logging wrapper functions for lgr compatibility}
\usage{
log_trace(msg, ...)

log_debug(msg, ...)

log_info(msg, ...)

log_warn(msg, ...)

log_error(msg, ...)

log_fatal(msg, ...)

log_success(msg, ...)
}
\arguments{
\item{msg}{Message template (use sprintf format: "Value: \%s")}

\item{...}{Values to substitute into msg via sprintf}
}
\value{
NULL (invisibly)
}
\description{
Simple logging wrappers that use sprintf-style formatting
}
\keyword{internal}
