% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{match_adduct_regex}
\alias{match_adduct_regex}
\title{Match adduct string against regex pattern}
\usage{
match_adduct_regex(adduct_string, regex)
}
\description{
Match adduct string against regex pattern
}
\keyword{internal}
