% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_hypothesis.R
\name{match_pairs_to_adduct_diffs}
\alias{match_pairs_to_adduct_diffs}
\title{Match feature pairs against pairwise adduct-mass differences}
\usage{
match_pairs_to_adduct_diffs(pairs, adduct_diffs)
}
\description{
Match feature pairs against pairwise adduct-mass differences
}
\keyword{internal}
