% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{MAX_PATH_LENGTH}
\alias{MAX_PATH_LENGTH}
\title{Maximum path length (OS-dependent, conservative estimate)}
\usage{
MAX_PATH_LENGTH
}
\description{
Maximum path length (OS-dependent, conservative estimate)
}
\keyword{internal}
