% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_TOLERANCE_PPM}
\alias{MAX_TOLERANCE_PPM}
\title{Maximum allowed mass tolerance in ppm
Tolerances above this value may produce excessive false positives}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_TOLERANCE_PPM
}
\description{
Maximum allowed mass tolerance in ppm
Tolerances above this value may produce excessive false positives
}
\keyword{internal}
