% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{MAX_TOLERANCE_RT_ADDUCTS}
\alias{MAX_TOLERANCE_RT_ADDUCTS}
\title{Maximum allowed retention time tolerance for adduct grouping (minutes)}
\usage{
MAX_TOLERANCE_RT_ADDUCTS
}
\description{
Maximum allowed retention time tolerance for adduct grouping (minutes)
}
\keyword{internal}
