% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_params.R
\name{merge_lists_recursive}
\alias{merge_lists_recursive}
\title{Recursively merge two lists}
\usage{
merge_lists_recursive(default, user)
}
\arguments{
\item{default}{Default list}

\item{user}{User list to override defaults}
}
\value{
Merged list
}
\description{
Deep merge where user values override defaults
}
\keyword{internal}
