% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{MIN_MASS_DALTONS}
\alias{MIN_MASS_DALTONS}
\title{Minimum mass value (Da) - used for validation}
\usage{
MIN_MASS_DALTONS
}
\description{
Minimum mass value (Da) - used for validation
}
\keyword{internal}
