% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_annotation.R
\name{ms1_annotation}
\alias{ms1_annotation}
\title{Title}
\usage{
ms1_annotation(
  annotationTable = metadata_table_spectral_annotation,
  structureExactMassTable = structure_exact_mass_table,
  structureOrganismPairsTable = structure_organism_pairs_table,
  adducts = unlist(adducts_list[[ms_mode]]),
  neutralLosses = neutral_losses_table,
  msMode = ms_mode,
  tolerancePpm = tolerance_ppm,
  toleranceRt = tolerance_rt,
  candidatesInitial = candidates_initial
)
}
\arguments{
\item{annotationTable}{TODO}

\item{structureExactMassTable}{TODO}

\item{structureOrganismPairsTable}{TODO}

\item{adducts}{TODO}

\item{neutralLosses}{TODO}

\item{msMode}{TODO}

\item{tolerancePpm}{TODO}

\item{toleranceRt}{TODO}

\item{candidatesInitial}{TODO}
}
\value{
TODO
}
\description{
Title
}
