% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_ms1_annotation.R
\name{non_ms1_annotation}
\alias{non_ms1_annotation}
\title{Non MS1 annotation}
\usage{
non_ms1_annotation(
  annotationTable = metadata_table_spectral_annotation,
  candidatesInitial = candidates_initial
)
}
\arguments{
\item{annotationTable}{TODO}

\item{candidatesInitial}{TODO}
}
\value{
TODO
}
\description{
Non MS1 annotation
}
\examples{
TODO
}
