% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{normalize_modifier_terms}
\alias{normalize_modifier_terms}
\title{Normalize modifier terms to a flat character vector}
\usage{
normalize_modifier_terms(x)
}
\description{
Normalize modifier terms to a flat character vector
}
\keyword{internal}
