% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{parse_modification_components}
\alias{parse_modification_components}
\title{Parse modification components (elements, signs, multiplicities)}
\usage{
parse_modification_components(modifications_clean)
}
\description{
Parse modification components (elements, signs, multiplicities)
}
\keyword{internal}
