% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{parse_single_adduct}
\alias{parse_single_adduct}
\title{Parse a single adduct (internal helper)}
\usage{
parse_single_adduct(adduct_string, regex, failed_parse)
}
\description{
Parse a single adduct (internal helper)
}
\keyword{internal}
