% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_params.R
\name{parse_yaml_params}
\alias{parse_yaml_params}
\title{Parse YAML parameters}
\usage{
parse_yaml_params()
}
\value{
A list containing the parameters specified in the YAML files
}
\description{
Parse YAML parameters
}
\examples{
params <- parse_yaml_params()
}
