% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_version_from_zenodo.R
\name{parse_zenodo_content}
\alias{parse_zenodo_content}
\title{Parse Zenodo API Response}
\usage{
parse_zenodo_content(api_url)
}
\arguments{
\item{api_url}{Character URL to Zenodo API endpoint}
}
\value{
List with parsed JSON content
}
\description{
Internal helper to parse JSON content from Zenodo API.
}
\keyword{internal}
