% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates_utils.R
\name{predicates_utils}
\alias{predicates_utils}
\title{Utility predicates for validation}
\description{
Collection of reusable predicate functions for common
validation patterns. These predicates follow SOLID principles and
can be easily composed and tested.
}
\keyword{internal}
