% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_adducts.R
\name{prepare_adducts}
\alias{prepare_adducts}
\title{Title}
\usage{
prepare_adducts(
  adducts_input = params$input,
  adducts_table_input = paths$data$source$adducts,
  config_output_path = paths$data$interim$config$path,
  adducts_output_path = paths$data$interim$adducts$path,
  output_name = params$output,
  masses_pos_output_path = paths$data$interim$adducts$pos,
  masses_neg_output_path = paths$data$interim$adducts$neg
)
}
\arguments{
\item{adducts_input}{TODO}

\item{adducts_table_input}{TODO}

\item{config_output_path}{TODO}

\item{adducts_output_path}{TODO}

\item{output_name}{TODO}

\item{masses_pos_output_path}{TODO}

\item{masses_neg_output_path}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
