% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations_gnps.R
\name{prepare_annotations_gnps}
\alias{prepare_annotations_gnps}
\title{Prepare annotations GNPS}
\usage{
prepare_annotations_gnps(
  input = get_params(step =
    "prepare_annotations_gnps")$files$annotations$raw$spectral$gnps,
  output = get_params(step =
    "prepare_annotations_gnps")$files$annotations$prepared$structural$gnps,
  str_stereo = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$metadata,
  str_tax_cla = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$taxonomies$npc
)
}
\arguments{
\item{input}{\link{character} Character string or vector of paths to GNPS annotation files}

\item{output}{\link{character} Character string path for prepared GNPS annotations output}

\item{str_stereo}{\link{character} Character string path to structures stereochemistry file}

\item{str_met}{\link{character} Character string path to structures metadata file}

\item{str_tax_cla}{\link{character} Character string path to ClassyFire taxonomy file}

\item{str_tax_npc}{\link{character} Character string path to NPClassifier taxonomy file}
}
\value{
Character string path to prepared GNPS annotations
}
\description{
This function prepares GNPS spectral library matching results
by standardizing column names, integrating structure metadata, and
formatting for downstream TIMA annotation workflows.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_annotations_gnps()
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation: 
\code{\link{prepare_annotations_mzmine}()},
\code{\link{prepare_annotations_sirius}()},
\code{\link{prepare_annotations_spectra}()},
\code{\link{prepare_features_components}()},
\code{\link{prepare_features_edges}()},
\code{\link{prepare_features_tables}()},
\code{\link{prepare_libraries_rt}()},
\code{\link{prepare_libraries_sop_bigg}()},
\code{\link{prepare_libraries_sop_closed}()},
\code{\link{prepare_libraries_sop_ecmdb}()},
\code{\link{prepare_libraries_sop_hmdb}()},
\code{\link{prepare_libraries_sop_lotus}()},
\code{\link{prepare_libraries_sop_merged}()},
\code{\link{prepare_libraries_spectra}()},
\code{\link{prepare_params}()},
\code{\link{prepare_taxa}()}
}
\concept{preparation}
