% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations_mzmine.R
\name{prepare_annotations_mzmine}
\alias{prepare_annotations_mzmine}
\title{Prepare annotations mzmine}
\usage{
prepare_annotations_mzmine(
  input = get_params(step =
    "prepare_annotations_mzmine")$files$annotations$raw$spectral$mzmine,
  output = get_params(step =
    "prepare_annotations_mzmine")$files$annotations$prepared$structural$mzmine,
  str_stereo = get_params(step =
    "prepare_annotations_mzmine")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "prepare_annotations_mzmine")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "prepare_annotations_mzmine")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "prepare_annotations_mzmine")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "prepare_annotations_mzmine")$files$libraries$sop$merged$structures$taxonomies$npc
)
}
\arguments{
\item{input}{Character string or vector of paths to mzmine annotation files}

\item{output}{Character string path for prepared mzmine annotations output}

\item{str_stereo}{Character string path to structures stereochemistry file}

\item{str_met}{Character string path to structures metadata file}

\item{str_nam}{Character string path to structures names file}

\item{str_tax_cla}{Character string path to ClassyFire taxonomy file}

\item{str_tax_npc}{Character string path to NPClassifier taxonomy file}
}
\value{
Character string path to prepared mzmine annotations
}
\description{
This function prepares mzmine spectral library matching results
by standardizing column names, integrating structure metadata, and
formatting for downstream TIMA annotation workflows.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_annotations_mzmine()
unlink("data", recursive = TRUE)
}
}
