% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_closed.R
\name{prepare_closed}
\alias{prepare_closed}
\title{Prepare closed}
\usage{
prepare_closed(
  input = params$files$libraries$sop$raw$closed,
  output = params$files$libraries$sop$processed,
  parameters = params
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{parameters}{params}
}
\description{
Prepare closed
}
\examples{
NULL
}
