% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_config.R
\name{prepare_config}
\alias{prepare_config}
\title{Prepare config}
\usage{
prepare_config(
  filename = params$files$pattern,
  gnps_job_id = params$gnps$id,
  ms_mode = params$ms$polarity,
  taxon = params$organisms$taxon,
  parameters = params
)
}
\arguments{
\item{filename}{Name of the file}

\item{gnps_job_id}{GNPS job ID}

\item{ms_mode}{MS ionization mode. 'pos' or 'neg'}

\item{taxon}{Name of a taxon you want to enforce}

\item{parameters}{Params}
}
\description{
This function prepares main config
}
\examples{
NULL
}
