% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_edges.R
\name{prepare_edges}
\alias{prepare_edges}
\title{Title}
\usage{
prepare_edges(
  tool = tool,
  gnps_job = params$gnps,
  input = params$input,
  output = params$output,
  name_source = params$source_name,
  name_target = params$target_name
)
}
\arguments{
\item{tool}{TODO}

\item{gnps_job}{TODO}

\item{input}{TODO}

\item{output}{TODO}

\item{name_source}{TODO}

\item{name_target}{TODO}
}
\value{
TODO
}
\description{
Title
}
