% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_classification.R
\name{prepare_features_classification}
\alias{prepare_features_classification}
\title{Prepare features classification}
\usage{
prepare_features_classification(
  library = params$library,
  input = params$input,
  output = params$output,
  quickmode = params$quickmode
)
}
\arguments{
\item{library}{Library to be used to classify features}

\item{input}{Input file}

\item{output}{Output file}

\item{quickmode}{Boolean. If TRUE, runs fast but does not complete missing data}
}
\description{
This function prepares the features chemical classes for further use
}
\examples{
NULL
}
