% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_components.R
\name{prepare_features_components}
\alias{prepare_features_components}
\title{Prepare features components}
\usage{
prepare_features_components(
  input = get_params(step =
    "prepare_features_components")$files$networks$spectral$components$raw,
  output = get_params(step =
    "prepare_features_components")$files$networks$spectral$components$prepared
)
}
\arguments{
\item{input}{\link{character} Character vector of paths to input component files.
Can be
a single file or multiple files that will be combined.}

\item{output}{\link{character} Character string path where prepared components
should be saved}
}
\value{
Character string path to the prepared features' components file
}
\description{
This function prepares molecular network component (cluster)
assignments by loading, standardizing, and formatting component IDs
for each feature. Components represent groups of related features in
the molecular network.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
github <- "https://raw.githubusercontent.com/"
repo <- "taxonomicallyinformedannotation/tima-example-files/main/"
dir <- paste0(github, repo)
input <- get_params(step =
    "prepare_features_components")$files$networks$spectral$components$raw
get_file(url = paste0(dir, input), export = input)
prepare_features_components(input = input)
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation: 
\code{\link{prepare_annotations_gnps}()},
\code{\link{prepare_annotations_mzmine}()},
\code{\link{prepare_annotations_sirius}()},
\code{\link{prepare_annotations_spectra}()},
\code{\link{prepare_features_edges}()},
\code{\link{prepare_features_tables}()},
\code{\link{prepare_libraries_rt}()},
\code{\link{prepare_libraries_sop_bigg}()},
\code{\link{prepare_libraries_sop_closed}()},
\code{\link{prepare_libraries_sop_ecmdb}()},
\code{\link{prepare_libraries_sop_hmdb}()},
\code{\link{prepare_libraries_sop_lotus}()},
\code{\link{prepare_libraries_sop_merged}()},
\code{\link{prepare_libraries_spectra}()},
\code{\link{prepare_params}()},
\code{\link{prepare_taxa}()}
}
\concept{preparation}
