% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_features.R
\name{prepare_features_table}
\alias{prepare_features_table}
\title{Coerce, harmonize, and prepare the features table for downstream steps}
\usage{
prepare_features_table(features_table, tolerance_rt)
}
\description{
Coerce, harmonize, and prepare the features table for downstream steps
}
\keyword{internal}
