% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gnps.R
\name{prepare_gnps}
\alias{prepare_gnps}
\title{Prepare GNPS}
\usage{
prepare_gnps(
  gnps_job_id = params$gnps$id,
  output = params$files$annotations$pretreated,
  parameters = params
)
}
\arguments{
\item{gnps_job_id}{GNPS job ID}

\item{output}{Output file}

\item{parameters}{Parameters}
}
\description{
This function prepares GNPS obtained annotations for further use
}
\examples{
NULL
}
