% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isdb.R
\name{prepare_isdb}
\alias{prepare_isdb}
\title{Prepare ISDB}
\usage{
prepare_isdb(input = params$input, output = params$output)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}
}
\value{
TODO
}
\description{
Prepare ISDB
}
\examples{
TODO
}
