% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isdb_lotus.R
\name{prepare_isdb_lotus}
\alias{prepare_isdb_lotus}
\title{Prepare LOTUS In Silico DataBase}
\usage{
prepare_isdb_lotus(
  input_pos = paths$data$source$spectra$lotus$pos,
  input_neg = paths$data$source$spectra$lotus$neg,
  output_pos = paths$data$interim$spectra$lotus$pos,
  output_neg = paths$data$interim$spectra$lotus$neg,
  export_sqlite = TRUE
)
}
\arguments{
\item{input_pos}{File containing positive ionization LOTUS ISDB}

\item{input_neg}{File containing negative ionization LOTUS ISDB}

\item{output_pos}{Output for positive mode}

\item{output_neg}{Output for negative mode}

\item{export_sqlite}{Boolean. Export to SQLite? TRUE or FALSE}
}
\description{
This function prepares the LOTUS in silico generated spectra to be used for spectral matching
}
\examples{
NULL
}
