% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_bigg.R
\name{prepare_libraries_sop_bigg}
\alias{prepare_libraries_sop_bigg}
\title{Prepare libraries of structure organism pairs BiGG}
\usage{
prepare_libraries_sop_bigg(
  output = get_params(step =
    "prepare_libraries_sop_bigg")$files$libraries$sop$prepared$bigg
)
}
\arguments{
\item{output}{Character string path for prepared BiGG library output}
}
\value{
Character string path to prepared BiGG structure-organism pairs
}
\description{
This function prepares BiGG (Biochemical, Genetic and
Genomic) structure-organism pairs by querying BiGG models and PubChem
for metabolite information, extracting chemical structures, and
formatting for TIMA annotation workflows.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_bigg()
unlink("data", recursive = TRUE)
}
}
