% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_closed.R
\name{prepare_libraries_sop_closed}
\alias{prepare_libraries_sop_closed}
\title{Prepare libraries of structure organism pairs CLOSED}
\usage{
prepare_libraries_sop_closed(
  input = get_params(step =
    "prepare_libraries_sop_closed")$files$libraries$sop$raw$closed,
  output = get_params(step =
    "prepare_libraries_sop_closed")$files$libraries$sop$prepared$closed
)
}
\arguments{
\item{input}{\link{character} Character string path to input closed library file}

\item{output}{\link{character} Character string path where prepared library should
be saved}
}
\value{
Character string path to the prepared structure-organism pairs
library
}
\description{
This function prepares closed (private/restricted) structure-
organism pair libraries by formatting columns, rounding values, and
standardizing structure. Falls back to an empty template if the closed
resource is not accessible.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_closed()
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation:
\code{\link[=prepare_annotations_gnps]{prepare_annotations_gnps()}},
\code{\link[=prepare_annotations_mzmine]{prepare_annotations_mzmine()}},
\code{\link[=prepare_annotations_sirius]{prepare_annotations_sirius()}},
\code{\link[=prepare_annotations_spectra]{prepare_annotations_spectra()}},
\code{\link[=prepare_features_components]{prepare_features_components()}},
\code{\link[=prepare_features_edges]{prepare_features_edges()}},
\code{\link[=prepare_features_tables]{prepare_features_tables()}},
\code{\link[=prepare_libraries_rt]{prepare_libraries_rt()}},
\code{\link[=prepare_libraries_sop_bigg]{prepare_libraries_sop_bigg()}},
\code{\link[=prepare_libraries_sop_ecmdb]{prepare_libraries_sop_ecmdb()}},
\code{\link[=prepare_libraries_sop_hmdb]{prepare_libraries_sop_hmdb()}},
\code{\link[=prepare_libraries_sop_lotus]{prepare_libraries_sop_lotus()}},
\code{\link[=prepare_libraries_sop_merged]{prepare_libraries_sop_merged()}},
\code{\link[=prepare_libraries_spectra]{prepare_libraries_spectra()}},
\code{\link[=prepare_params]{prepare_params()}},
\code{\link[=prepare_taxa]{prepare_taxa()}}
}
\concept{preparation}
