% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_hmdb.R
\name{prepare_libraries_sop_hmdb}
\alias{prepare_libraries_sop_hmdb}
\title{Prepare libraries of structure organism pairs HMDB}
\usage{
prepare_libraries_sop_hmdb(
  input = get_params(step = "prepare_libraries_sop_hmdb")$files$libraries$sop$raw$hmdb,
  output = get_params(step =
    "prepare_libraries_sop_hmdb")$files$libraries$sop$prepared$hmdb
)
}
\arguments{
\item{input}{\link{character} Character string path to HMDB SDF zip file}

\item{output}{\link{character} Character string path for prepared HMDB library output}
}
\value{
Character string path to prepared HMDB structure-organism pairs
}
\description{
This function prepares HMDB (Human Metabolome Database)
structure-organism pairs by parsing SDF files, extracting metadata,
and formatting for TIMA annotation workflows.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_hmdb()
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation: 
\code{\link{prepare_annotations_gnps}()},
\code{\link{prepare_annotations_mzmine}()},
\code{\link{prepare_annotations_sirius}()},
\code{\link{prepare_annotations_spectra}()},
\code{\link{prepare_features_components}()},
\code{\link{prepare_features_edges}()},
\code{\link{prepare_features_tables}()},
\code{\link{prepare_libraries_rt}()},
\code{\link{prepare_libraries_sop_bigg}()},
\code{\link{prepare_libraries_sop_closed}()},
\code{\link{prepare_libraries_sop_ecmdb}()},
\code{\link{prepare_libraries_sop_lotus}()},
\code{\link{prepare_libraries_sop_merged}()},
\code{\link{prepare_libraries_spectra}()},
\code{\link{prepare_params}()},
\code{\link{prepare_taxa}()}
}
\concept{preparation}
