% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_hmdb_like.R
\name{prepare_libraries_sop_hmdb_like}
\alias{prepare_libraries_sop_hmdb_like}
\title{Prepare HMDB-like structure organism pairs library}
\usage{
prepare_libraries_sop_hmdb_like(
  input,
  output,
  source_name,
  organism_name = "Homo sapiens",
  organism_taxonomy = list(organism_taxonomy_ottid = "770315", organism_taxonomy_01domain
    = "Eukaryota", organism_taxonomy_02kingdom = "Metazoa", organism_taxonomy_03phylum =
    "Chordata", organism_taxonomy_04class = "Mammalia", organism_taxonomy_05order =
    "Primates", organism_taxonomy_06family = "Hominidae", organism_taxonomy_07tribe =
    NA_character_, organism_taxonomy_08genus = "Homo", organism_taxonomy_09species =
    "Homo sapiens", organism_taxonomy_10varietas = NA_character_),
  tag = NA_character_
)
}
\arguments{
\item{input}{\link{character} Character string path to SDF zip file}

\item{output}{\link{character} Character string path for prepared SOP output}

\item{source_name}{\link{character} Character label used in logs}

\item{organism_name}{\link{character} Character organism name attached to all rows}

\item{organism_taxonomy}{\link{list} Named list containing organism taxonomy
fields}

\item{tag}{\link{character} Character scalar tag propagated through SOP pipeline}
}
\value{
Character string path to prepared SOP file
}
\description{
Internal helper for HMDB ecosystem databases distributed as
SDF archives with HMDB-style fields.
}
\keyword{internal}
