% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_lotus.R
\name{prepare_libraries_sop_lotus}
\alias{prepare_libraries_sop_lotus}
\title{Prepare libraries of structure organism pairs LOTUS}
\usage{
prepare_libraries_sop_lotus(
  input = get_params(step = "prepare_libraries_sop_lotus")$files$libraries$sop$raw$lotus,
  output = get_params(step =
    "prepare_libraries_sop_lotus")$files$libraries$sop$prepared$lotus
)
}
\arguments{
\item{input}{\link{character} Character string path to the raw LOTUS data file}

\item{output}{\link{character} Character string path for the prepared output file}
}
\value{
Character string path to the prepared structure-organism pairs
library file
}
\description{
This function prepares the LOTUS (LOng-lasting, cUraTed
collection
of cOnnectivity daTa for natural products) structure-organism pairs database.
It standardizes columns, extracts 2D InChIKeys, rounds numeric values,
and removes duplicates.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_lotus()
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation:
\code{\link[=prepare_annotations_gnps]{prepare_annotations_gnps()}},
\code{\link[=prepare_annotations_mzmine]{prepare_annotations_mzmine()}},
\code{\link[=prepare_annotations_sirius]{prepare_annotations_sirius()}},
\code{\link[=prepare_annotations_spectra]{prepare_annotations_spectra()}},
\code{\link[=prepare_features_components]{prepare_features_components()}},
\code{\link[=prepare_features_edges]{prepare_features_edges()}},
\code{\link[=prepare_features_tables]{prepare_features_tables()}},
\code{\link[=prepare_libraries_rt]{prepare_libraries_rt()}},
\code{\link[=prepare_libraries_sop_bigg]{prepare_libraries_sop_bigg()}},
\code{\link[=prepare_libraries_sop_closed]{prepare_libraries_sop_closed()}},
\code{\link[=prepare_libraries_sop_ecmdb]{prepare_libraries_sop_ecmdb()}},
\code{\link[=prepare_libraries_sop_hmdb]{prepare_libraries_sop_hmdb()}},
\code{\link[=prepare_libraries_sop_merged]{prepare_libraries_sop_merged()}},
\code{\link[=prepare_libraries_spectra]{prepare_libraries_spectra()}},
\code{\link[=prepare_params]{prepare_params()}},
\code{\link[=prepare_taxa]{prepare_taxa()}}
}
\concept{preparation}
