% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_pubchemlite.R
\name{prepare_libraries_sop_pubchemlite}
\alias{prepare_libraries_sop_pubchemlite}
\title{Prepare libraries of structure organism pairs PubChem Lite}
\usage{
prepare_libraries_sop_pubchemlite(
  input = get_params(step =
    "prepare_libraries_sop_pubchemlite")$files$libraries$sop$raw$pubchemlite,
  output = get_params(step =
    "prepare_libraries_sop_pubchemlite")$files$libraries$sop$prepared$pubchemlite
)
}
\arguments{
\item{input}{\link{character} Character string path to PubChem Lite CSV file}

\item{output}{\link{character} Character string path for prepared SOP output}
}
\value{
Character string path to prepared SOP file
}
\description{
This function prepares the PubChem Lite CCSbase export for
exposomics as a xenobiotic structure-organism pairs library.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_pubchemlite()
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation:
\code{\link[=prepare_annotations_gnps]{prepare_annotations_gnps()}},
\code{\link[=prepare_annotations_mzmine]{prepare_annotations_mzmine()}},
\code{\link[=prepare_annotations_mztab]{prepare_annotations_mztab()}},
\code{\link[=prepare_annotations_sirius]{prepare_annotations_sirius()}},
\code{\link[=prepare_annotations_spectra]{prepare_annotations_spectra()}},
\code{\link[=prepare_features_components]{prepare_features_components()}},
\code{\link[=prepare_features_edges]{prepare_features_edges()}},
\code{\link[=prepare_features_tables]{prepare_features_tables()}},
\code{\link[=prepare_libraries_rt]{prepare_libraries_rt()}},
\code{\link[=prepare_libraries_sop_bigg]{prepare_libraries_sop_bigg()}},
\code{\link[=prepare_libraries_sop_closed]{prepare_libraries_sop_closed()}},
\code{\link[=prepare_libraries_sop_ecmdb]{prepare_libraries_sop_ecmdb()}},
\code{\link[=prepare_libraries_sop_hmdb]{prepare_libraries_sop_hmdb()}},
\code{\link[=prepare_libraries_sop_lotus]{prepare_libraries_sop_lotus()}},
\code{\link[=prepare_libraries_sop_merged]{prepare_libraries_sop_merged()}},
\code{\link[=prepare_libraries_spectra]{prepare_libraries_spectra()}},
\code{\link[=prepare_params]{prepare_params()}},
\code{\link[=prepare_taxa]{prepare_taxa()}},
\code{\link[=read_mztab]{read_mztab()}}
}
\concept{preparation}
