% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lotus.R
\name{prepare_lotus}
\alias{prepare_lotus}
\title{Prepare LOTUS}
\usage{
prepare_lotus(
  input = paths$data$source$libraries$sop$lotus,
  output = paths$data$interim$libraries$lotus
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\description{
This function prepares the LOTUS structure-organism pairs
}
\examples{
NULL
}
