% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sirius.R
\name{prepare_sirius}
\alias{prepare_sirius}
\title{Prepare sirius}
\usage{
prepare_sirius(
  input_directory = params$files$annotations$raw$sirius,
  npc = params$tools$taxonomies$chemical,
  output = params$files$annotations$pretreated,
  parameters = params
)
}
\arguments{
\item{input_directory}{Directory containing the Sirius results}

\item{npc}{Boolean. NPClassifier classes computed. TRUE or FALSE}

\item{output}{Output where to save prepared results}

\item{parameters}{Params}
}
\description{
This function prepares Sirius results to make them compatible
}
\examples{
NULL
}
