% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_spectral_matches.R
\name{prepare_spectral_matches}
\alias{prepare_spectral_matches}
\title{Prepare spectral matches}
\usage{
prepare_spectral_matches(input = params$input, output = params$output)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\description{
This function prepares the spectral matches obtained previously to make them compatible
}
\examples{
NULL
}
