% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{process_annotations}
\alias{process_annotations}
\title{Process annotations}
\usage{
process_annotations(
  library = params$library,
  name = params$name,
  gnps = params$annotation$gnps,
  isdb = params$annotation$isdb,
  sirius = params$annotation$sirius,
  taxa = params$taxa,
  edges = params$edges,
  output = params$output,
  candidates_initial = params$top_k$initial,
  candidates_final = params$top_k$final,
  weight_spectral = params$weights$spectral,
  weight_chemical = params$weights$chemical,
  weight_biological = params$weights$biological,
  score_chemical_pathway = params$scores$chemical$pathway,
  score_chemical_superclass = params$scores$chemical$superclass,
  score_chemical_class = params$scores$chemical$class,
  score_biological_domain = params$scores$biological$domain,
  score_biological_kingdom = params$scores$biological$kingdom,
  score_biological_phylum = params$scores$biological$phylum,
  score_biological_class = params$scores$biological$class,
  score_biological_order = params$scores$biological$order,
  score_biological_infraorder = params$scores$biological$infraorder,
  score_biological_family = params$scores$biological$family,
  score_biological_subfamily = params$scores$biological$subfamily,
  score_biological_tribe = params$scores$biological$tribe,
  score_biological_subtribe = params$scores$biological$subtribe,
  score_biological_genus = params$scores$biological$genus,
  score_biological_subgenus = params$scores$biological$subgenus,
  score_biological_species = params$scores$biological$species,
  score_biological_subspecies = params$scores$biological$subspecies,
  score_biological_variety = params$scores$biological$variety,
  ms_mode = params$ms$mode,
  ms_level = params$ms$level,
  annotate = params$ms$annotate,
  tolerance_ppm = params$ms$tolerance$ppm,
  tolerance_rt = params$ms$tolerance$rt,
  adducts_list = params$ms$adducts,
  minimal_ms1_bio = params$scores$biological$minimal,
  minimal_ms1_chemo = params$scores$chemical$minimal,
  ms1_only = params$ms$ms1only,
  force = params$force
)
}
\arguments{
\item{library}{Library to be used to perform MS1 annotation}

\item{name}{Name of the adducts}

\item{gnps}{GNPS annotation results}

\item{isdb}{ISDB annotation results}

\item{sirius}{SIRIUS annotation results}

\item{taxa}{Prepared taxed features file}

\item{edges}{Prepared edges file}

\item{output}{Output file}

\item{candidates_initial}{Number of initial candidates to keep}

\item{candidates_final}{Number of final candidates to keep}

\item{weight_spectral}{Weight for the spectral score}

\item{weight_chemical}{Weight for the biological score}

\item{weight_biological}{Weight for the chemical consistency score}

\item{score_chemical_pathway}{Score for a \code{pathway} match (should be lower than \code{superclass})}

\item{score_chemical_superclass}{Score for a \code{superclass} match (should be lower than \code{class})}

\item{score_chemical_class}{Score for a \code{class} match (should be the highest)}

\item{score_biological_domain}{Score for a \code{domain} match (should be lower than \code{kingdom})}

\item{score_biological_kingdom}{Score for a \code{kingdom} match (should be lower than \code{phylum})}

\item{score_biological_phylum}{Score for a \code{phylum} match (should be lower than \code{class})}

\item{score_biological_class}{Score for a \code{class} match (should be lower than \code{order})}

\item{score_biological_order}{Score for a \code{order} match (should be lower than \code{infraorder})}

\item{score_biological_infraorder}{Score for a \code{infraorder} match (should be lower than \code{order})}

\item{score_biological_family}{Score for a \code{family} match (should be lower than \code{subfamily})}

\item{score_biological_subfamily}{Score for a \code{subfamily} match (should be lower than \code{family})}

\item{score_biological_tribe}{Score for a \code{tribe} match (should be lower than \code{subtribe})}

\item{score_biological_subtribe}{Score for a \code{subtribe} match (should be lower than \code{genus})}

\item{score_biological_genus}{Score for a \code{genus} match (should be lower than \code{subgenus})}

\item{score_biological_subgenus}{Score for a \code{subgenus} match (should be lower than \code{species})}

\item{score_biological_species}{Score for a \code{species} match (should be lower than \code{subspecies})}

\item{score_biological_subspecies}{Score for a \code{subspecies} match (should be lower than \code{variety})}

\item{score_biological_variety}{Score for a \code{variety} match (should be the highest)}

\item{ms_mode}{MS ionization mode. Should be 'pos' or 'neg'}

\item{ms_level}{1}

\item{annotate}{Boolean. Perform MS1 annotation completion or not}

\item{tolerance_ppm}{Tolerance in ppm for MS1 annotation}

\item{tolerance_rt}{Tolerance in retention time (minute) for adducts attribution}

\item{adducts_list}{Adducts list}

\item{minimal_ms1_bio}{Minimal biological score to keep MS1 based annotation}

\item{minimal_ms1_chemo}{Minimal chemical score to keep MS1 based annotation}

\item{ms1_only}{Boolean. Keep only MS1 annotations}

\item{force}{Force parameters. Use it at your own risk}
}
\description{
This function weights and eventually complements initial annotations.
}
\examples{
NULL
}
\seealso{
annotate_ms1 weight_bio weight_chemo
}
