% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{process_modifications}
\alias{process_modifications}
\title{Process adduct modifications and calculate total mass change}
\usage{
process_modifications(modifications_raw)
}
\description{
Process adduct modifications and calculate total mass change
}
\keyword{internal}
