% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectra.R
\name{process_spectra}
\alias{process_spectra}
\title{Process spectra}
\usage{
process_spectra(
  input = params$input,
  library = params$library,
  output = params$output,
  method = params$ms$similarity$method,
  threshold = params$ms$similarity$threshold,
  ppm = params$ms$tolerance$ppm,
  dalton = params$ms$tolerance$dalton,
  npeaks = params$ms$peaks$absolute,
  rpeaks = params$ms$peaks$ratio,
  condition = params$ms$condition,
  quickmode = params$quickmode,
  parallel = params$parallel
)
}
\arguments{
\item{input}{Query file containing spectra. Currently an '.mgf' file}

\item{library}{Library containing spectra to match against. Can be '.mgf' or '.sqlite' (Spectra formatted)}

\item{output}{Output file.}

\item{method}{Method to be used to perform spectral comparison}

\item{threshold}{Minimal similarity to report}

\item{ppm}{Relative ppm tolerance to be used}

\item{dalton}{Absolute Dalton tolerance to be used}

\item{npeaks}{Absolute minimum number of peaks to be matched}

\item{rpeaks}{Relative minimum number of peaks to be matched}

\item{condition}{Condition to be fulfilled. Either 'OR' or 'AND' (mass and peaks minima).}

\item{quickmode}{Boolean. Quick mode goes faster but does not output matched peaks and ratios.}

\item{parallel}{Boolean. Process in parallel}
}
\description{
This function performs spectral matching.
}
\details{
It takes two files as input. A query file that will be matched against a library file.
Multiple comparison distances are available
('gnps', 'navdist','ndotproduct','neuclidean', 'nspectraangle' (See MsCoreUtils for details)).
Number of matched peaks and their ratio are also available if using 'quickmode=FALSE'.
Parallel processing is also made available.
}
\examples{
NULL
}
