% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PROTON_MASS_DALTONS}
\alias{PROTON_MASS_DALTONS}
\title{Proton mass in Daltons}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PROTON_MASS_DALTONS
}
\description{
Proton mass in Daltons
}
\keyword{internal}
