% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edges.R
\name{read_edges}
\alias{read_edges}
\title{Read edges}
\usage{
read_edges(id)
}
\arguments{
\item{id}{Character string containing a GNPS job ID}
}
\value{
Data frame containing the network edges for the specified GNPS job
}
\description{
This function reads the edges table from GNPS
}
\examples{
NULL
}
