% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnps.R
\name{read_features}
\alias{read_features}
\title{#' Title
#'
#' @param file
#' @param path
#'
#' @return
#' @export
#'
#' @examples
download_file <- function(file, path = "NULL") {
if (is.null(path)) {
error("Must enter a path")
}
f <- RCurl::CFILE(path, mode = "wb")
RCurl::curlPerform(url = file, writedata = f@ref)
close(f)
}
Title}
\usage{
read_features(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
#' Title
#'
#' @param file
#' @param path
#'
#' @return
#' @export
#'
#' @examples
download_file <- function(file, path = "NULL") {
if (is.null(path)) {
error("Must enter a path")
}
f <- RCurl::CFILE(path, mode = "wb")
RCurl::curlPerform(url = file, writedata = f@ref)
close(f)
}
Title
}
