% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(id)
}
\arguments{
\item{id}{Character string containing a GNPS job ID}
}
\value{
Data frame containing the metadata for the specified GNPS job
}
\description{
This function reads the metadata table from GNPS
}
\examples{
NULL
}
