% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgf_opti.R
\name{read_mgf_opti}
\alias{read_mgf_opti}
\title{Read MGF opti}
\usage{
read_mgf_opti(
  f,
  msLevel = 2L,
  mapping = Spectra::spectraVariableMapping(MsBackendMgf::MsBackendMgf())
)
}
\arguments{
\item{f}{A character string specifying the path to a single MGF file.}

\item{msLevel}{An integer specifying the MS level. Default is 2L.}

\item{mapping}{A named character vector mapping MGF fields to standard spectra variables.
Defaults to \code{Spectra::spectraVariableMapping(MsBackendMgf::MsBackendMgf())}.}
}
\value{
A \code{DataFrame} containing the parsed spectra data.
}
\description{
This function reads a Mascot Generic Format (MGF) file while mimicking the
\code{MsBackendMgf} implementation but using significantly lower memory,
making it more suitable for large MGF files.
}
\examples{
NULL
}
