% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nap.R
\name{read_nap}
\alias{read_nap}
\title{Read NAP}
\usage{
read_nap(id)
}
\arguments{
\item{id}{a character string containing a GNPS job ID}
}
\value{
a data frame containing the node attributes table for the specified GNPS job
}
\description{
TODO
}
\examples{
NULL
}
