% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{Read results}
\usage{
read_results(id, workflow = params$gnps$workflow)
}
\arguments{
\item{id}{Character string containing the job ID of the GNPS task}

\item{workflow}{Character string indicating the type of workflow used in the GNPS task. Supported values are 'fbmn' and 'classical'}
}
\value{
Data frame containing the results of the GNPS task
}
\description{
This function reads GNPS results table
}
\examples{
NULL
}
