% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_annotations.R
\name{rearrange_annotations}
\alias{rearrange_annotations}
\title{Rearrange Annotation Tables}
\usage{
rearrange_annotations(
  annotation_table,
  formula_table,
  canopus_table,
  edges_table
)
}
\arguments{
\item{annotation_table}{Data frame with annotations}

\item{formula_table}{Data frame with formula data}

\item{canopus_table}{Data frame with canopus data}

\item{edges_table}{Data frame with edges}
}
\value{
Data frame with rearranged annotations
}
\description{
Internal helper to reorganize and merge annotation tables.
}
\keyword{internal}
