% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{rebuild_hill_formula}
\alias{rebuild_hill_formula}
\title{Rebuild a Hill-order atomic formula string from named counts}
\usage{
rebuild_hill_formula(atom_counts)
}
\description{
Rebuild a Hill-order atomic formula string from named counts
}
\keyword{internal}
