% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{reduce_ammonium_loss_pairs}
\alias{reduce_ammonium_loss_pairs}
\title{Reduce (+H4N, -loss-containing-H3N) pairs to proton-plus-residual-loss form}
\usage{
reduce_ammonium_loss_pairs(net)
}
\description{
Reduce (+H4N, -loss-containing-H3N) pairs to proton-plus-residual-loss form
}
\keyword{internal}
