% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{remove_comments}
\alias{remove_comments}
\title{Remove comments from modification string}
\usage{
remove_comments(modification_string)
}
\description{
Remove comments from modification string
}
\keyword{internal}
