% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chemo.R
\name{remove_compound_names}
\alias{remove_compound_names}
\title{Remove Compound Names from Results}
\usage{
remove_compound_names(results_list, compounds_names)
}
\arguments{
\item{results_list}{Named list with full, filtered, mini data frames}

\item{compounds_names}{Logical, if FALSE remove names}
}
\value{
Modified results list
}
\description{
Internal helper to optionally remove compound names from
all result tiers.
}
\keyword{internal}
