% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_gnps_job_id.R
\name{replace_gnps_job_id}
\alias{replace_gnps_job_id}
\title{Replace GNPS job ID}
\usage{
replace_gnps_job_id(
  job,
  default = yamls_default$`config/prepare_config`$gnps$id,
  repl = params$gnps$id
)
}
\arguments{
\item{job}{a character string containing a GNPS job ID}

\item{default}{a default values for a GNPS job ID}

\item{repl}{a user-specified values for a GNPS job ID}
}
\value{
Character string with the GNPS job ID modified according to the rules specified in the function
}
\description{
This function replaces the default GNPS job ID given in the example by a user-specified one
}
\examples{
replaced_job_id <- replace_gnps_job_id(job = "MyGNPSJobIDXXX", default = "XXX", repl = "Foo")
}
