% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_id.R
\name{replace_id_in_filename}
\alias{replace_id_in_filename}
\title{Replace ID in filename}
\usage{
replace_id_in_filename(filename, old_id, new_id)
}
\description{
Replace ID in filename
}
\keyword{internal}
