% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{resolve_adduct_consistency_config}
\alias{resolve_adduct_consistency_config}
\title{Resolve \verb{adduct_consistency*} config defaults}
\usage{
resolve_adduct_consistency_config(
  adduct_consistency,
  adduct_min_support,
  adduct_consistency_min_degree
)
}
\description{
Resolve \verb{adduct_consistency*} config defaults
}
\keyword{internal}
