% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_coverage.R
\name{resolve_annotate_masses_coverage_mode}
\alias{resolve_annotate_masses_coverage_mode}
\title{Resolve annotate_masses coverage mode}
\usage{
resolve_annotate_masses_coverage_mode(coverage_mode)
}
\description{
Resolve annotate_masses coverage mode
}
\keyword{internal}
