% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{retain_supported_single_m_edges}
\alias{retain_supported_single_m_edges}
\title{Keep only supported edges that are coherent with a single neutral M}
\usage{
retain_supported_single_m_edges(
  adduct_edges,
  cluster_edges,
  loss_edges,
  annotations,
  tolerance_ppm,
  tolerance_dalton
)
}
\description{
Keep only supported edges that are coherent with a single neutral M
}
\keyword{internal}
