% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_reals.R
\name{round_reals}
\alias{round_reals}
\title{Round reals}
\usage{
round_reals(df, dig = 5L, cols = c("structure_exact_mass", "structure_xlogp"))
}
\arguments{
\item{df}{Dataframe to use}

\item{dig}{Number of digits}

\item{cols}{Columns}
}
\description{
This function rounds reals in columns of a dataframe
}
\examples{
NULL
}
