% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_spectra_benchmark.R
\name{sanitize_spectra_benchmark}
\alias{sanitize_spectra_benchmark}
\title{Sanitize spectra benchmark}
\usage{
sanitize_spectra_benchmark(
  sp,
  mgf_pos_path,
  mgf_neg_path,
  meta_pos_path = "data/interim/benchmark/benchmark_meta_pos.tsv",
  meta_neg_path = "data/interim/benchmark/benchmark_meta_neg.tsv"
)
}
\arguments{
\item{sp}{Spectra}

\item{mgf_pos_path}{Path to store the positive spectra}

\item{mgf_neg_path}{Path to store the negative spectra}

\item{meta_pos_path}{Path to store the positive metadata}

\item{meta_neg_path}{Path to store the negative metadata}
}
\description{
This function sanitizes the benchmark spectra
}
\details{
Because they are still quite dirty
}
\examples{
NULL
}
