% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{score_modifier_relation_from_hypotheses}
\alias{score_modifier_relation_from_hypotheses}
\title{Score how well node hypotheses support one modifier interpretation}
\usage{
score_modifier_relation_from_hypotheses(
  src_hypotheses,
  dest_hypotheses,
  modifier,
  relation
)
}
\description{
Score how well node hypotheses support one modifier interpretation
}
\keyword{internal}
