% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_utils.R
\name{select_annotations_complement_working_columns}
\alias{select_annotations_complement_working_columns}
\title{Select working columns for annotation complementation}
\usage{
select_annotations_complement_working_columns(df)
}
\description{
Select working columns for annotation complementation
}
\keyword{internal}
