% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sirius_columns.R
\name{select_sirius_columns}
\alias{select_sirius_columns}
\title{Select sirius columns}
\usage{
select_sirius_columns(df)
}
\arguments{
\item{df}{Dataframe}
}
\description{
This function selects sirius columns
}
\examples{
NULL
}
