% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sirius_columns.R
\name{select_sirius_columns_2}
\alias{select_sirius_columns_2}
\title{Select sirius columns 2}
\usage{
select_sirius_columns_2(df)
}
\arguments{
\item{df}{Dataframe}
}
\description{
This function selects sirius columns (2)
}
\examples{
NULL
}
