% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sop_columns.R
\name{select_sop_columns}
\alias{select_sop_columns}
\title{Select SOP columns}
\usage{
select_sop_columns(df)
}
\arguments{
\item{df}{Dataframe containing structure-organism pair data}
}
\value{
The dataframe with selected and renamed SOP columns in standardized format
}
\description{
This function selects and renames structure-organism pair (SOP)
columns to a standardized format, including chemical structures,
taxonomic classifications, and organism taxonomy
}
\examples{
NULL
}
