% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_utils.R
\name{select_structure_organism_pair_columns}
\alias{select_structure_organism_pair_columns}
\title{Select structure-organism pair columns}
\usage{
select_structure_organism_pair_columns(df)
}
\description{
Select structure-organism pair columns
}
\keyword{internal}
