% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_utils.R
\name{select_structure_organism_pair_loading_columns}
\alias{select_structure_organism_pair_loading_columns}
\title{Select loading columns for structure-organism pairs}
\usage{
select_structure_organism_pair_loading_columns(df)
}
\description{
Select loading columns for structure-organism pairs
}
\keyword{internal}
