% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_utils.R
\name{select_weight_annotations_working_columns}
\alias{select_weight_annotations_working_columns}
\title{Select working columns for annotation weighting}
\usage{
select_weight_annotations_working_columns(df)
}
\description{
Select working columns for annotation weighting
}
\keyword{internal}
