% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{setup_virtualenv}
\alias{setup_virtualenv}
\title{Setup Python Virtual Environment}
\usage{
setup_virtualenv(envname = "tima-env", python = NULL)
}
\arguments{
\item{envname}{Character name of virtual environment}

\item{python}{Character path to Python executable}
}
\value{
NULL (side effect: creates virtualenv and installs packages)
}
\description{
Internal helper to create and configure Python virtualenv with dependencies.
}
\keyword{internal}
