% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{show_system_messages}
\alias{show_system_messages}
\title{Display System-Specific Installation Messages}
\usage{
show_system_messages(system)
}
\arguments{
\item{system}{\link{character} OS name from Sys.info()}
}
\value{
NULL (side effect: logging)
}
\description{
Internal helper to show OS-specific installation instructions.
}
\keyword{internal}
