% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_universe_matching.R
\name{single_component_increment}
\alias{single_component_increment}
\title{Return the single component added when \code{to} differs from \code{from} by +1 term}
\usage{
single_component_increment(from, to)
}
\description{
Return the single component added when \code{to} differs from \code{from} by +1 term
}
\keyword{internal}
