% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPECTRAL_FORMATS}
\alias{SPECTRAL_FORMATS}
\title{Supported spectral data formats}
\format{
An object of class \code{character} of length 4.
}
\usage{
SPECTRAL_FORMATS
}
\description{
Supported spectral data formats
}
\keyword{datasets}
