% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{subtract_atomic_formula}
\alias{subtract_atomic_formula}
\title{Subtract one formula from another, returning the residual formula string}
\usage{
subtract_atomic_formula(formula, sub_formula = "H3N")
}
\description{
Subtract one formula from another, returning the residual formula string
}
\keyword{internal}
