% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TABLE_EXTENSIONS}
\alias{TABLE_EXTENSIONS}
\title{Supported table file extensions}
\format{
An object of class \code{character} of length 3.
}
\usage{
TABLE_EXTENSIONS
}
\description{
Supported table file extensions
}
\keyword{internal}
