% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize_spectra_benchmark.R
\name{taxize_spectra_benchmark}
\alias{taxize_spectra_benchmark}
\title{Taxize spectra benchmark}
\usage{
taxize_spectra_benchmark(input, keys, org_tax_ott, output)
}
\arguments{
\item{input}{Initial features}

\item{keys}{SOP keys}

\item{org_tax_ott}{Taxonomy}

\item{output}{Prepared features}
}
\description{
This function adds taxa to the benchmark
}
\details{
Because they are still quite dirty
}
\examples{
NULL
}
