% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TEMP_FILE_PREFIX}
\alias{TEMP_FILE_PREFIX}
\title{Temporary file prefix for write tests}
\format{
An object of class \code{character} of length 1.
}
\usage{
TEMP_FILE_PREFIX
}
\description{
Temporary file prefix for write tests
}
\keyword{internal}
