% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tima-package.R
\docType{package}
\name{tima-package}
\alias{tima}
\alias{tima-package}
\title{tima: Taxonomically Informed Metabolite Annotation}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

TIMA provides a reproducible workflow for taxonomically informed metabolite annotation from feature tables, MS/MS spectra, and optional external resources such as SIRIUS and GNPS outputs. It combines mass, spectral, taxonomic, and structural evidence into a transparent scoring framework that can be inspected step-by-step. The package targets metabolomics practitioners who need configurable, scriptable, and documented annotation pipelines for research and production settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/taxonomicallyinformedannotation/tima}
  \item \url{https://taxonomicallyinformedannotation.github.io/tima-shinylive}
  \item \url{https://taxonomicallyinformedannotation.github.io/tima}
  \item Report bugs at \url{https://github.com/taxonomicallyinformedannotation/tima/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{adafede@gmail.com} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

Authors:
\itemize{
  \item Adriano Rutz \email{adafede@gmail.com} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})
}

Other contributors:
\itemize{
  \item Pierre-Marie Allard \email{pierre-marie.allard@unifr.ch} (\href{https://orcid.org/0000-0003-3389-2191}{ORCID}) [contributor]
}

}
\keyword{internal}
