% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timaR-package.R
\docType{package}
\name{timaR-package}
\alias{timaR}
\alias{timaR-package}
\title{timaR: Taxonomically Informed Metabolite Annotation}
\description{
TIMA provides the infrastructure to perform Taxonomically Informed Metabolite Annotation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/taxonomicallyinformedannotation/tima-r}
  \item \url{https://taxonomicallyinformedannotation.github.io/tima-r/}
  \item Report bugs at \url{https://github.com/taxonomicallyinformedannotation/tima-r/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{adriano.rutz@unige.ch} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

Other contributors:
\itemize{
  \item Pierre-Marie Allard \email{pierre-marie.allard@unifr.ch} (\href{https://orcid.org/0000-0003-3389-2191}{ORCID}) [contributor]
}

}
\keyword{internal}
