% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_names_sirius.R
\name{treat_names_sirius}
\alias{treat_names_sirius}
\title{Treat names sirius}
\usage{
treat_names_sirius(x)
}
\arguments{
\item{x}{a character string containing a name}
}
\value{
a character string with the name modified according to the rules specified in the function
}
\description{
Treat names sirius
}
\examples{
treated_name <- treat_names_sirius("My_name")
}
