% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_MS_MODES}
\alias{VALID_MS_MODES}
\title{Valid ionization mode values}
\format{
An object of class \code{character} of length 2.
}
\usage{
VALID_MS_MODES
}
\description{
Valid ionization mode values
}
\keyword{datasets}
