% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_SIMILARITY_METHODS}
\alias{VALID_SIMILARITY_METHODS}
\title{Valid similarity method values}
\format{
An object of class \code{character} of length 3.
}
\usage{
VALID_SIMILARITY_METHODS
}
\description{
Valid similarity method values
}
\keyword{internal}
