% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{validate_adduct_string}
\alias{validate_adduct_string}
\title{Validate adduct string format}
\usage{
validate_adduct_string(adduct_string)
}
\description{
Validate adduct string format
}
\keyword{internal}
