% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{validate_adduct_translations}
\alias{validate_adduct_translations}
\title{Validate adduct translations structure}
\usage{
validate_adduct_translations(translations)
}
\description{
Validate adduct translations structure
}
\keyword{internal}
