% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_spectra.R
\name{validate_column_mappings}
\alias{validate_column_mappings}
\title{Validate column mapping parameters}
\usage{
validate_column_mappings(col_params)
}
\description{
Validate column mapping parameters
}
\keyword{internal}
