% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{validate_downloaded_file}
\alias{validate_downloaded_file}
\title{Validate downloaded file exists and has content}
\usage{
validate_downloaded_file(export, url)
}
\description{
Validate downloaded file exists and has content
}
\keyword{internal}
