% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{validate_file_content}
\alias{validate_file_content}
\title{Validate file content is not an HTML error page}
\usage{
validate_file_content(export, url)
}
\description{
Validate file content is not an HTML error page
}
\keyword{internal}
