% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{validate_http_response}
\alias{validate_http_response}
\title{Validate HTTP response status}
\usage{
validate_http_response(resp, url, export)
}
\description{
Validate HTTP response status
}
\keyword{internal}
