% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{verify_directory_writable}
\alias{verify_directory_writable}
\title{Verify that a directory is writable}
\usage{
verify_directory_writable(dir_path)
}
\description{
Verify that a directory is writable
}
\keyword{internal}
