% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_utils.R
\name{with_progress2}
\alias{with_progress2}
\title{Progress Bar for map2 Operations}
\usage{
with_progress2(.x, .y, .f, desc = "Processing", show_progress = TRUE, ...)
}
\arguments{
\item{.x}{First vector to iterate over}

\item{.y}{Second vector to iterate over}

\item{.f}{Function taking two arguments}

\item{desc}{Character progress description}

\item{show_progress}{Logical whether to show progress}

\item{...}{Additional arguments to .f}
}
\value{
List of results
}
\description{
Wraps purrr::map2 with progress bar
}
\keyword{internal}
