% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{write_empty_annotate_masses_outputs}
\alias{write_empty_annotate_masses_outputs}
\title{Write the empty-output sentinel files}
\usage{
write_empty_annotate_masses_outputs(output_annotations, output_edges)
}
\description{
Write the empty-output sentinel files
}
\keyword{internal}
