% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{annotate_masses}
\alias{annotate_masses}
\title{Annotate masses}
\usage{
annotate_masses(
  features = get_params(step = "annotate_masses")$files$features$prepared,
  output_annotations = get_params(step =
    "annotate_masses")$files$annotations$prepared$structural$ms1,
  output_edges = get_params(step = "annotate_masses")$files$networks$spectral$edges$raw,
  name_source = get_params(step = "annotate_masses")$names$source,
  name_target = get_params(step = "annotate_masses")$names$target,
  library = get_params(step = "annotate_masses")$files$libraries$sop$merged$keys,
  str_stereo = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$npc,
  adducts_list = get_params(step = "annotate_masses")$ms$adducts,
  clusters_list = get_params(step = "annotate_masses")$ms$clusters,
  neutral_losses_list = get_params(step = "annotate_masses")$ms$neutral_losses,
  ms_mode = get_params(step = "annotate_masses")$ms$polarity,
  tolerance_ppm = get_params(step = "annotate_masses")$ms$tolerances$mass$ppm$ms1,
  tolerance_rt = get_params(step = "annotate_masses")$ms$tolerances$rt$adducts
)
}
\arguments{
\item{features}{Table containing your previous annotation to complement}

\item{output_annotations}{Output for mass based structural annotations}

\item{output_edges}{Output for mass based edges}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}

\item{library}{Library containing the keys}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}

\item{adducts_list}{List of adducts to be used}

\item{clusters_list}{List of clusters to be used}

\item{neutral_losses_list}{List of neutral losses to be used}

\item{ms_mode}{Ionization mode. Must be 'pos' or 'neg'}

\item{tolerance_ppm}{Tolerance to perform annotation. Should be <= 10 ppm}

\item{tolerance_rt}{Tolerance to group adducts. Should be <= 0.05 minutes}
}
\value{
The path to the files containing MS1 annotations and edges
}
\description{
This function annotates a feature table based on exact mass
match. It requires a structural library, its metadata, and lists of adducts,
clusters, and neutral losses to be considered. The polarity has to be \code{pos}
or \code{neg} and retention time and mass tolerances should be given. The feature
table is expected to be pre-formatted.
}
\examples{
annotate_masses(
  features = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/features/example_features.tsv",
  library = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/libraries/sop/merged/keys.tsv",
  str_stereo = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/libraries/sop/merged/structure/stereo.tsv",
  str_met = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/libraries/sop/merged/structure/metadata.tsv",
  str_nam = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/libraries/sop/merged/structure/names.tsv",
  str_tax_cla = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/libraries/sop/merged/structure/taxonomies/classyfire.tsv",
  str_tax_npc = "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima-example-files/main/data/interim/libraries/sop/merged/structure/taxonomies/npc.tsv",
  adducts_list = list(pos = c("[M+H]+"), neg = c("[M-H]-")),
  clusters_list = list(pos = c("H2O"), neg = c("H2O")),
  neutral_losses_list = c("C6H10O4 (methylpentose/desoxyhexose-H2O)"),
  ms_mode = "pos",
  tolerance_ppm = 10,
  tolerance_rt = 0.01
)
}
