% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{annotate_masses}
\alias{annotate_masses}
\title{Annotate masses}
\usage{
annotate_masses(
  features = params$files$features$prepared,
  output_annotations = params$files$annotations$prepared$structural,
  output_edges = params$files$networks$spectral$edges$raw,
  name_source = params$names$source,
  name_target = params$names$target,
  library = params$files$libraries$sop$merged$keys,
  str_stereo = params$files$libraries$sop$merged$structures$stereo,
  str_met = params$files$libraries$sop$merged$structures$metadata,
  str_nam = params$files$libraries$sop$merged$structures$names,
  str_tax_cla = params$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = params$files$libraries$sop$merged$structures$taxonomies$npc,
  name = params$files$libraries$adducts$prepared,
  adducts_list = params$ms$adducts,
  adducts_neg = params$files$libraries$adducts$neg,
  adducts_pos = params$files$libraries$adducts$pos,
  adducts_masses_list = system.file("extdata", "adducts.tsv", package = "timaR"),
  clusters_neg = params$ms$clusters$neg,
  clusters_pos = params$ms$clusters$pos,
  clusters_list = system.file("extdata", "clusters.tsv", package = "timaR"),
  neutral_losses_list = system.file("extdata", "neutral_losses.tsv", package = "timaR"),
  ms_mode = params$ms$polarity,
  tolerance_ppm = params$ms$tolerances$mass$ppm$ms1,
  tolerance_rt = params$ms$tolerances$rt$minutes,
  parameters = params
)
}
\arguments{
\item{features}{Table containing your previous annotation to complement}

\item{output_annotations}{Output for mass based structural annotations}

\item{output_edges}{Output for mass based edges}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}

\item{library}{Library containing the keys}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}

\item{name}{Name of the adducts library}

\item{adducts_list}{List of adducts to be used}

\item{adducts_neg}{Negative adducts to be used}

\item{adducts_pos}{Positive adducts to be used}

\item{adducts_masses_list}{Adducts masses}

\item{clusters_neg}{Negative clusters to be used}

\item{clusters_pos}{Positive clusters to be used}

\item{clusters_list}{List of clusters to be used}

\item{neutral_losses_list}{List of neutral losses to be used}

\item{ms_mode}{Ionization mode. Must be 'pos' or 'neg'}

\item{tolerance_ppm}{Tolerance to perform annotation. Should be ≤ 10 ppm}

\item{tolerance_rt}{Tolerance to group adducts. Should be ≤ 0.1min}

\item{parameters}{params}
}
\value{
A table containing MS1 annotations based on exact mass
}
\description{
This function annotates masses
}
\examples{
NULL
}
