% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses.R
\name{annotate_masses}
\alias{annotate_masses}
\title{Annotate masses}
\usage{
annotate_masses(
  features = params$files$features$prepared,
  output_annotations = params$files$annotations$prepared,
  output_edges = params$files$networks$spectral$edges$raw,
  name_source = params$names$source,
  name_target = params$names$target,
  library = paths$data$interim$libraries$sop$merged$keys,
  str_2D_3D = params$files$libraries$sop$merged$structures$dd_ddd,
  str_met = params$files$libraries$sop$merged$structures$metadata,
  str_nam = params$files$libraries$sop$merged$structures$names,
  str_tax_cla = params$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = params$files$libraries$sop$merged$structures$taxonomies$npc,
  name = params$files$libraries$adducts$prepared,
  adducts_list = params$ms$adducts,
  adducts_masses_list = paths$inst$extdata$adducts,
  neutral_losses_list = paths$inst$extdata$neutral_losses,
  msMode = params$ms$polarity,
  tolerancePpm = params$ms$tolerances$mass$ppm$ms1,
  toleranceRt = params$ms$tolerances$rt$minutes,
  parameters = params
)
}
\arguments{
\item{features}{Table containing your previous annotation to complement}

\item{output_annotations}{Output for mass based structural annotations}

\item{output_edges}{Output for mass based edges}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}

\item{library}{Library containing the keys}

\item{str_2D_3D}{File containing 2D and 3D structures}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}

\item{name}{Name of the adducts library}

\item{adducts_list}{List of adducts to be used}

\item{adducts_masses_list}{Adducts masses}

\item{neutral_losses_list}{List of neutral losses to be used}

\item{msMode}{Ionization mode. Must be 'pos' or 'neg'}

\item{tolerancePpm}{Tolerance to perform annotation. Should be lower than 10 ppm}

\item{toleranceRt}{Tolerance to perform adduct attribution. Should be lower than 0.1min}

\item{parameters}{params}
}
\value{
A table containing MS1 annotations based on exact mass
}
\description{
This function annotates masses
}
\examples{
NULL
}
