% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms1.R
\name{annotate_ms1}
\alias{annotate_ms1}
\title{Annotate MS1}
\usage{
annotate_ms1(
  features = params$files$features$prepared,
  output = params$files$annotations$pretreated,
  library = paths$data$interim$libraries$merged$keys,
  str_2D_3D = paths$data$interim$libraries$merged$structures$dd_ddd,
  str_met = paths$data$interim$libraries$merged$structures$metadata,
  str_nam = paths$data$interim$libraries$merged$structures$names,
  str_tax_cla = paths$data$interim$libraries$merged$structures$taxonomies$classyfire,
  str_tax_npc = paths$data$interim$libraries$merged$structures$taxonomies$npc,
  name = params$files$libraries$adducts$processed,
  adducts_list = params$ms$adducts,
  adducts_masses_list = paths$inst$extdata$adducts,
  neutral_losses_list = paths$inst$extdata$neutral_losses,
  msMode = params$ms$polarity,
  tolerancePpm = params$ms$tolerances$mass$ppm$ms1,
  toleranceRt = params$ms$tolerances$rt$minutes,
  parameters = params
)
}
\arguments{
\item{msMode}{Ionization mode. Must be 'pos' or 'neg'}

\item{tolerancePpm}{Tolerance to perform annotation. Should be lower than 10 ppm}

\item{toleranceRt}{Tolerance to perform adduct attribution. Should be lower than 0.1min}

\item{featuresTable}{Table containing your previous annotation to complement}

\item{structureExactMassTable}{Table containing the structure - exact mass pairs to perform annotation}

\item{adducts}{List of adducts to be used}

\item{adductsM}{Adducts masses}

\item{adductsTable}{Table containing the adducted library masses}

\item{neutralLosses}{List of neutral losses to be used}
}
\value{
A table containing MS1 annotations based on exact mass
}
\description{
This function performs MS1 annotation.
}
\examples{
NULL
}
\seealso{
annotate_non_ms1
}
