% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms1.R
\name{annotate_ms1}
\alias{annotate_ms1}
\title{Annotate MS1}
\usage{
annotate_ms1(
  annotationTable = annotation_table_ms2,
  structureExactMassTable = structure_exact_mass_table,
  structureOrganismPairsTable = structure_organism_pairs_table,
  adducts = unlist(adducts_list[[ms_mode]]),
  adductsM = adducts_m,
  adductsTable = adducts_table,
  neutralLosses = neutral_losses_table,
  msMode = ms_mode,
  tolerancePpm = tolerance_ppm,
  toleranceRt = tolerance_rt,
  candidatesInitial = candidates_initial
)
}
\arguments{
\item{annotationTable}{Table containing your previous annotation to complement}

\item{structureExactMassTable}{Table containing the structure - exact mass pairs to perform annotation}

\item{structureOrganismPairsTable}{Table containing the structure - organism pairs}

\item{adducts}{List of adducts to be used}

\item{adductsM}{Adducts masses}

\item{adductsTable}{Table containing the adducted library masses}

\item{neutralLosses}{List of neutral losses to be used}

\item{msMode}{Ionization mode. Must be 'pos' or 'neg'}

\item{tolerancePpm}{Tolerance to perform annotation. Should be lower than 10 ppm}

\item{toleranceRt}{Tolerance to perform adduct attribution. Should be lower than 0.1min}

\item{candidatesInitial}{Number of initial candidates to keep}
}
\value{
A table containing the initial annotation complemented by additional MS1 annotations based on exact mass
}
\description{
This function complements MS2 results with MS1 annotation and filters the final results.
}
\examples{
NULL
}
\seealso{
annotate_non_ms1
}
