% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_non_ms1.R
\name{annotate_non_ms1}
\alias{annotate_non_ms1}
\title{Annotate non MS1}
\usage{
annotate_non_ms1(
  annotationTable = metadata_table_spectral_annotation,
  candidatesInitial = candidates_initial
)
}
\arguments{
\item{annotationTable}{Table containing your previous annotation}

\item{candidatesInitial}{Number of initial candidates to keep}
}
\value{
A table containing the initial annotation where only part of initial candidates are kept
}
\description{
This function filters MS2 results in case no MS1 annotation was performed
}
\examples{
NULL
}
\seealso{
annotate_ms1
}
