% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_spectra.R
\name{annotate_spectra}
\alias{annotate_spectra}
\title{Annotate spectra}
\usage{
annotate_spectra(
  input = params$files$spectral$raw,
  library = params$files$libraries$spectral$exp,
  polarity = params$ms$polarity,
  output = params$files$annotations$raw$spectral,
  threshold = params$annotations$ms2$thresholds$similarity,
  ppm = params$ms$tolerances$mass$ppm$ms2,
  dalton = params$ms$tolerances$mass$dalton$ms2,
  qutoff = params$ms$intensity$thresholds$ms2,
  parallel = params$options$parallel,
  approx = params$annotations$ms2$approx,
  parameters = params
)
}
\arguments{
\item{input}{Query file containing spectra. Currently an '.mgf' file}

\item{library}{Library containing spectra to match against.
Can be '.mgf' or '.sqlite' (Spectra formatted)}

\item{polarity}{MS polarity. Must be 'pos' or 'neg'.}

\item{output}{Output file.}

\item{threshold}{Minimal similarity to report}

\item{ppm}{Relative ppm tolerance to be used}

\item{dalton}{Absolute Dalton tolerance to be used}

\item{qutoff}{Intensity under which ms2 fragments will be removed.}

\item{parallel}{Boolean. Process in parallel}

\item{approx}{Perform matching without precursor match}

\item{parameters}{Params}

\item{condition}{Condition to be fulfilled.
Either 'OR' or 'AND' (mass and peaks minima).}
}
\description{
This function annotates spectra
}
\details{
It takes two files as input.
A query file that will be matched against a library file.
Parallel processing is also made available.
}
\examples{
NULL
}
