% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_consistency.R
\name{apply_adduct_consistency_filter}
\alias{apply_adduct_consistency_filter}
\title{Apply optional support-based adduct consistency filtering.}
\usage{
apply_adduct_consistency_filter(
  df_add,
  adduct_consistency = "conditional",
  adduct_min_support = 2L,
  adduct_consistency_min_degree = 3L
)
}
\description{
Apply optional support-based adduct consistency filtering.
}
\keyword{internal}
