% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_taxize_spectra.R
\name{benchmark_taxize_spectra}
\alias{benchmark_taxize_spectra}
\title{Benchmark taxize spectra}
\usage{
benchmark_taxize_spectra(input, keys, org_tax_ott, output)
}
\arguments{
\item{input}{Initial features}

\item{keys}{SOP keys}

\item{org_tax_ott}{Taxonomy}

\item{output}{Prepared features}
}
\value{
The path to the taxed benchmark
}
\description{
This function adds taxa to the benchmark
}
\details{
Because they are still quite dirty
}
\examples{
NULL
}
