% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate entropy}
\usage{
calculate_entropy(
  index,
  target,
  frags,
  ms2_tolerance,
  ppm_tolerance,
  threshold = 0.1
)
}
\arguments{
\item{index}{Index of the first spectrum}

\item{target}{Index of the second spectrum}

\item{frags}{List of fragments}

\item{ms2_tolerance}{MS2 tolerance}

\item{ppm_tolerance}{ppm tolerance}

\item{threshold}{Threshold value for the score}
}
\value{
A list containing the calculated entropy similarity values
or NULL if the score is below the threshold
}
\description{
This function calculates entropy similarity between two spectra
}
\examples{
NULL
}
