% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mass_of_m.R
\name{calculate_mass_of_m}
\alias{calculate_mass_of_m}
\title{Calculate mass of M}
\usage{
calculate_mass_of_m(adduct_string, mz, electron_mass = 0.0005485799)
}
\arguments{
\item{adduct_string}{Adduct to be parsed}

\item{mz}{mz}

\item{electron_mass}{Electron mass}
}
\value{
A mass
}
\description{
This function calculates the mass of M
}
\examples{
calculate_mass_of_m(mz = 123.4567, adduct_string = "[M+H]+")
calculate_mass_of_m(mz = 123.4567, adduct_string = "[M+Na]+")
calculate_mass_of_m(mz = 123.4567, adduct_string = "[2M1-C6H12O6 (hexose)+NaCl+H]2+")
}
