% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mass_of_m.R
\name{calculate_neutral_mass_formula}
\alias{calculate_neutral_mass_formula}
\title{Calculate neutral mass using the adduct formula}
\usage{
calculate_neutral_mass_formula(
  mz,
  n_charges,
  charge_sign,
  n_iso,
  mass_modifications,
  electron_mass,
  n_mer
)
}
\description{
Calculate neutral mass using the adduct formula
}
\keyword{internal}
