% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_similarity.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Calculate similarity between spectra}
\usage{
calculate_similarity(
  method,
  query_spectrum,
  target_spectrum,
  query_precursor,
  target_precursor,
  dalton,
  ppm,
  return_matched_peaks = FALSE,
  ...
)
}
\arguments{
\item{method}{Character string specifying method: "entropy", "gnps", or "cosine"}

\item{query_spectrum}{Numeric matrix with columns for mz and intensity}

\item{target_spectrum}{Numeric matrix with columns for mz and intensity}

\item{query_precursor}{Numeric precursor m/z value for query}

\item{target_precursor}{Numeric precursor m/z value for target}

\item{dalton}{Numeric Dalton tolerance for peak matching}

\item{ppm}{Numeric PPM tolerance for peak matching}

\item{return_matched_peaks}{Logical; return matched peaks count?
Not compatible with 'entropy' method. Default: FALSE}

\item{...}{Additional arguments passed to MsCoreUtils::join}
}
\value{
Numeric similarity score (0-1), or list with score and matches
if return_matched_peaks = TRUE. Returns 0.0 if calculation fails.
}
\description{
Efficiently calculates similarity scores between query and
target spectra using either entropy, cosine, or GNPS methods
}
\examples{
sp_1 <- cbind(
  mz = c(10, 36, 63, 91, 93),
  intensity = c(14, 15, 999, 650, 1)
)
precursor_1 <- 123.4567
precursor_2 <- precursor_1 + 14
sp_2 <- cbind(
  mz = c(10, 12, 50, 63, 105),
  intensity = c(35, 5, 16, 999, 450)
)
calculate_similarity(
  method = "entropy",
  query_spectrum = sp_1,
  target_spectrum = sp_2,
  query_precursor = precursor_1,
  target_precursor = precursor_2,
  dalton = 0.005,
  ppm = 10.0
)
calculate_similarity(
  method = "gnps",
  query_spectrum = sp_1,
  target_spectrum = sp_2,
  query_precursor = precursor_1,
  target_precursor = precursor_2,
  dalton = 0.005,
  ppm = 10.0,
  return_matched_peaks = TRUE
)
}
