% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_weighting.R
\name{chemical_weighting}
\alias{chemical_weighting}
\title{Title}
\usage{
chemical_weighting(
  annotationTableWeightedBioCleaned = annotation_table_weighted_bio_cleaned,
  WeightChemical = weight_chemical,
  scoreChemicalPathway = score_chemical_pathway,
  scoreChemicalSuperclass = score_chemical_superclass,
  scoreChemicalClass = score_chemical_class
)
}
\arguments{
\item{annotationTableWeightedBioCleaned}{TODO}

\item{WeightChemical}{TODO}

\item{scoreChemicalPathway}{TODO}

\item{scoreChemicalSuperclass}{TODO}

\item{scoreChemicalClass}{TODO}
}
\value{
TODO
}
\description{
Title
}
