% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bio.R
\name{clean_bio}
\alias{clean_bio}
\title{Clean bio}
\usage{
clean_bio(
  annotationTableWeightedBio = annotation_table_weighted_bio,
  structureOrganismPairsTable = structure_organism_pairs_table,
  edgesTable = edges_table,
  aNnOtAtE = annotate,
  candidatesInitial = candidates_initial,
  minimalMs1Bio = minimal_ms1_bio
)
}
\arguments{
\item{annotationTableWeightedBio}{TODO}

\item{structureOrganismPairsTable}{TODO}

\item{edgesTable}{TODO}

\item{aNnOtAtE}{TODO}

\item{candidatesInitial}{TODO}

\item{minimalMs1Bio}{TODO}
}
\value{
TODO
}
\description{
Clean bio
}
\examples{
TODO
}
