% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bio.R
\name{clean_bio}
\alias{clean_bio}
\title{Clean bio}
\usage{
clean_bio(
  annot_table_wei_bio = get("annot_table_wei_bio", envir = parent.frame()),
  edges_table = get("edges_table", envir = parent.frame()),
  candidates_initial = get("candidates_initial", envir = parent.frame())
)
}
\arguments{
\item{annot_table_wei_bio}{Table containing your
biologically weighted annotation}

\item{edges_table}{Table containing the edges between features}

\item{candidates_initial}{Number of initial candidates to keep}
}
\value{
A table containing the biologically weighted annotation
where only a given number of initial candidates are kept
}
\description{
This function cleans the results
obtained after biological weighting
}
\examples{
NULL
}
\seealso{
weight_bio
}
