% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_collapse.R
\name{clean_character_columns}
\alias{clean_character_columns}
\title{Clean character columns (trim and convert empty to NA)}
\usage{
clean_character_columns(df)
}
\description{
Clean character columns (trim and convert empty to NA)
}
\keyword{internal}
