% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_collapse.R
\name{clean_collapse}
\alias{clean_collapse}
\title{Clean collapse}
\usage{
clean_collapse(grouped_df, cols = NA)
}
\arguments{
\item{grouped_df}{Grouped dataframe}

\item{cols}{Column(s) to apply collapse to}
}
\value{
Cleaned and collapsed dataframe
}
\description{
This function collapses a grouped dataframe and trims it
}
\examples{
NULL
}
