% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_metadata.R
\name{complement_metadata}
\alias{complement_metadata}
\title{Complement metadata}
\usage{
complement_metadata(df)
}
\arguments{
\item{df}{data frame with metadata to be complemented}
}
\value{
data frame with complemented metadata
}
\description{
This function fetches the metadata from Zenodo and joins it with the input data frame.

It also renames and removes unnecessary columns.
}
\examples{
complement_metadata(df)
}
