% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_metadata_structures.R
\name{complement_metadata_structures}
\alias{complement_metadata_structures}
\title{Complement metadata of structures}
\usage{
complement_metadata_structures(
  df,
  str_stereo = params$files$libraries$merged$structures$stereo,
  str_met = params$files$libraries$merged$structures$metadata,
  str_nam = params$files$libraries$merged$structures$names,
  str_tax_cla = params$files$libraries$merged$structures$taxonomies$cla,
  str_tax_npc = params$files$libraries$merged$structures$taxonomies$npc
)
}
\arguments{
\item{df}{Data frame with structural metadata to be complemented}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\value{
Data frame with complemented structural metadata
}
\description{
This function complement structural metadata
}
\examples{
NULL
}
