% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_metadata_structures.R
\name{complement_metadata_structures}
\alias{complement_metadata_structures}
\title{Complement metadata of structures}
\usage{
complement_metadata_structures(
  df,
  str_stereo = get("str_stereo", envir = parent.frame()),
  str_met = get("str_met", envir = parent.frame()),
  str_nam = get("str_nam", envir = parent.frame()),
  str_tax_cla = get("str_tax_cla", envir = parent.frame()),
  str_tax_npc = get("str_tax_npc", envir = parent.frame())
)
}
\arguments{
\item{df}{Data frame with structural metadata to be complemented}

\item{str_stereo}{Character string path to file with structure stereochemistry}

\item{str_met}{Character string path to file with structure metadata}

\item{str_nam}{Character string path to file with structure names}

\item{str_tax_cla}{Character string path to file with ClassyFire taxonomy}

\item{str_tax_npc}{Character string path to file with NPClassifier taxonomy}
}
\value{
Data frame with enriched structural metadata
}
\description{
This function complements structural metadata by joining
stereochemistry, metadata, names, and chemical taxonomy information
from reference libraries. Enriches annotation results with comprehensive
structure information.
}
\examples{
NULL
}
