% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_metadata_structures.R
\name{complement_metadata_structures}
\alias{complement_metadata_structures}
\title{Complement metadata of structures}
\usage{
complement_metadata_structures(
  df,
  str_stereo = get("str_stereo", envir = parent.frame()),
  str_met = get("str_met", envir = parent.frame()),
  str_nam = get("str_nam", envir = parent.frame()),
  str_tax_cla = get("str_tax_cla", envir = parent.frame()),
  str_tax_npc = get("str_tax_npc", envir = parent.frame())
)
}
\arguments{
\item{df}{Data frame with structural metadata to be complemented}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\value{
Data frame with complemented structural metadata
}
\description{
This function complement structural metadata
}
\examples{
NULL
}
