% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_metadata_structures.R
\name{complement_metadata_structures}
\alias{complement_metadata_structures}
\title{Complement metadata of structures}
\usage{
complement_metadata_structures(
  df,
  str_2d_3d = paths$data$interim$libraries$merged$structures$dd_ddd,
  str_met = paths$data$interim$libraries$merged$structures$metadata,
  str_nam = paths$data$interim$libraries$merged$structures$names,
  str_tax_cla = paths$data$interim$libraries$merged$structures$taxonomies$classyfire,
  str_tax_npc = paths$data$interim$libraries$merged$structures$taxonomies$npc
)
}
\arguments{
\item{df}{Data frame with structural metadata to be complemented}

\item{str_2d_3d}{File containing 2D and 3D structures}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\value{
Data frame with complemented structural metadata
}
\description{
This function complement structural metadata
}
\examples{
NULL
}
