% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_backbone.R
\name{copy_backbone}
\alias{copy_backbone}
\title{Copy backbone}
\usage{
copy_backbone(cache_dir = ".tima", package = "tima", copy_dir = "inst")
}
\arguments{
\item{cache_dir}{Cache directory}

\item{package}{Package}

\item{copy_dir}{Copy directory}
}
\description{
This function copies backbone
}
\examples{
NULL
}
