% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adducts_pos.R
\name{create_adducts_pos}
\alias{create_adducts_pos}
\title{Create adducts positive}
\usage{
create_adducts_pos(
  masses_table = get("masses_table", envir = parent.frame()),
  adducts_table = get("adducts_table", envir = parent.frame())
)
}
\arguments{
\item{masses_table}{Table containing the masses of the adducts}

\item{adducts_table}{Table containing the adducts}
}
\description{
This function creates positive adducts
}
\examples{
NULL
}
