% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create directory}
\usage{
create_dir(export)
}
\arguments{
\item{export}{Path to the directory to be created}
}
\value{
Message indicating the status of directory creation
}
\description{
This function creates a directory at the specified path
if it does not already exist.
}
\examples{
create_dir(export = "path/to/directory_of_file")
unlink("path", recursive = TRUE)
}
