% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges.R
\name{create_edges}
\alias{create_edges}
\title{Create edges}
\usage{
create_edges(
  index,
  frags,
  precs,
  nspecs,
  ms2_tolerance,
  ppm_tolerance,
  threshold
)
}
\arguments{
\item{index}{Indices}

\item{frags}{Fragments}

\item{precs}{Precursors}

\item{nspecs}{Number of spectra}

\item{ms2_tolerance}{MS2 tolerance}

\item{ppm_tolerance}{ppm tolerance}

\item{threshold}{Threshold}
}
\description{
This function applies similarity calculation to a list of
spectra to create edges
}
\examples{
NULL
}
