% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges.R
\name{create_edges}
\alias{create_edges}
\title{Create edges}
\usage{
create_edges(
  frags,
  nspecs,
  precs,
  method,
  ms2_tolerance,
  ppm_tolerance,
  threshold,
  matched_peaks
)
}
\arguments{
\item{frags}{Fragments}

\item{nspecs}{Number of spectra}

\item{precs}{Precursors}

\item{method}{Method}

\item{ms2_tolerance}{MS2 tolerance}

\item{ppm_tolerance}{ppm tolerance}

\item{threshold}{Threshold}

\item{matched_peaks}{Matched peaks}
}
\description{
This function applies similarity calculation to a list of
spectra to create edges
}
\examples{
NULL
}
