% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_parallel.R
\name{create_edges_parallel}
\alias{create_edges_parallel}
\title{Create edges parallel}
\usage{
create_edges_parallel(
  spectra,
  ms2_tolerance = params$ms$tolerances$mass$dalton$ms2,
  ppm_tolerance = params$ms$tolerances$mass$ppm$ms2,
  parallel = params$options$parallel
)
}
\description{
This function is part of the creation of edges
}
\examples{
NULL
}
