% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_parallel.R
\name{create_edges_parallel}
\alias{create_edges_parallel}
\title{Create edges parallel}
\usage{
create_edges_parallel(index, frags, precs, nspecs)
}
\arguments{
\item{frags}{Fragments}

\item{precs}{Precursors}

\item{nspecs}{Number of spectra}

\item{x}{Indices}
}
\description{
This function part of the creation of edges
}
\examples{
NULL
}
