% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_progress.R
\name{create_edges_progress}
\alias{create_edges_progress}
\title{Create edges progress}
\usage{
create_edges_progress(xs)
}
\arguments{
\item{xs}{Indices of spectra}
}
\description{
This function is slow so it outputs the progression of the creation of edges
}
\examples{
NULL
}
