% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_progress.R
\name{create_edges_progress}
\alias{create_edges_progress}
\title{Create edges progress}
\usage{
create_edges_progress(
  index,
  frags,
  precs,
  nspecs,
  ms2_tolerance,
  ppm_tolerance,
  p
)
}
\arguments{
\item{index}{Indices}

\item{frags}{Fragments}

\item{precs}{Precursors}

\item{nspecs}{Number of spectra}

\item{p}{Progressor}
}
\description{
This function is slow so it had to be parallelized
}
\examples{
NULL
}
