% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_progress.R
\name{create_edges_progress}
\alias{create_edges_progress}
\title{Create edges progress}
\usage{
create_edges_progress(target, query, s1, fragments, precursors)
}
\arguments{
\item{target}{Indices of target spectra}

\item{query}{Index of query spectra}

\item{s1}{Query spectrum}

\item{fragments}{Fragments}

\item{precursors}{Precursors}
}
\description{
This function is slow so it outputs the progression of the creation of edges
}
\examples{
NULL
}
