% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_spectra.R
\name{create_edges_spectra}
\alias{create_edges_spectra}
\title{Create edges spectra}
\usage{
create_edges_spectra(
  input = params$files$spectral$raw,
  output = params$files$networks$spectral$edges$raw,
  name_source = params$names$source,
  name_target = params$names$target,
  threshold = params$annotations$ms2$thresholds$similarity,
  ppm = params$ms$tolerances$mass$ppm$ms2,
  dalton = params$ms$tolerances$mass$dalton$ms2,
  qutoff = params$ms$intensity$thresholds$ms2,
  parallel = params$options$parallel,
  parameters = params
)
}
\arguments{
\item{input}{Query file containing spectra. Currently an '.mgf' file}

\item{output}{Output file.}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}

\item{threshold}{Minimal similarity to report}

\item{ppm}{Relative ppm tolerance to be used}

\item{dalton}{Absolute Dalton tolerance to be used}

\item{qutoff}{Intensity under which ms2 fragments will be removed.}

\item{parallel}{Boolean. Process in parallel}

\item{parameters}{Params}

\item{method}{Method to be used to perform spectral comparison}
}
\description{
This function create edges
based on fragmentation spectra similarity
}
\examples{
NULL
}
