% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edges_spectra.R
\name{create_edges_spectra}
\alias{create_edges_spectra}
\title{Create edges spectra}
\usage{
create_edges_spectra(
  input = get_params(step = "create_edges_spectra")$files$spectral$raw,
  output = get_params(step = "create_edges_spectra")$files$networks$spectral$edges$raw,
  name_source = get_params(step = "create_edges_spectra")$names$source,
  name_target = get_params(step = "create_edges_spectra")$names$target,
  threshold = get_params(step =
    "create_edges_spectra")$annotations$thresholds$ms2$similarity$edges,
  ppm = get_params(step = "create_edges_spectra")$ms$tolerances$mass$ppm$ms2,
  dalton = get_params(step = "create_edges_spectra")$ms$tolerances$mass$dalton$ms2,
  qutoff = get_params(step = "create_edges_spectra")$ms$thresholds$ms2$intensity
)
}
\arguments{
\item{input}{Query file containing spectra. Currently an '.mgf' file}

\item{output}{Output file.}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}

\item{threshold}{Minimal similarity to report}

\item{ppm}{Relative ppm tolerance to be used}

\item{dalton}{Absolute Dalton tolerance to be used}

\item{qutoff}{Intensity under which ms2 fragments will be removed.}
}
\value{
The path to the created spectral edges
}
\description{
This function create edges
based on fragmentation spectra similarity
}
\examples{
\donttest{
tima:::copy_backbone()
go_to_cache()
get_file(
  url = get_default_paths()$urls$examples$spectra_mini,
  export = get_params(step = "create_edges_spectra")$files$spectral$raw
)
create_edges_spectra()
unlink("data", recursive = TRUE)
}
}
