% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_bio}
\alias{decorate_bio}
\title{Decorate bio}
\usage{
decorate_bio()
}
\value{
Message indicating the number of annotations weighted at each biological level
}
\description{
TODO
}
\examples{
NULL
}
