% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_bio.R
\name{decorate_bio}
\alias{decorate_bio}
\title{Decorate bio}
\usage{
decorate_bio(
  annot_table_wei_bio = get("annot_table_wei_chemo", envir = parent.frame()),
  score_biological_kingdom = get("score_biological_kingdom", envir = parent.frame()),
  score_biological_phylum = get("score_biological_phylum", envir = parent.frame()),
  score_biological_class = get("score_biological_class", envir = parent.frame()),
  score_biological_order = get("score_biological_order", envir = parent.frame()),
  score_biological_family = get("score_biological_family", envir = parent.frame()),
  score_biological_genus = get("score_biological_genus", envir = parent.frame()),
  score_biological_species = get("score_biological_species", envir = parent.frame()),
  score_biological_variety = get("score_biological_variety", envir = parent.frame())
)
}
\arguments{
\item{annot_table_wei_bio}{Table to decorate}

\item{score_biological_kingdom}{Kingdom score}

\item{score_biological_phylum}{Phylum score}

\item{score_biological_class}{Class score}

\item{score_biological_order}{Order score}

\item{score_biological_family}{Family score}

\item{score_biological_genus}{Genus score}

\item{score_biological_species}{Species score}

\item{score_biological_variety}{Variety score}
}
\value{
Message indicating the number of annotations
weighted at each biological level
}
\description{
This function outputs information about biological weighting
}
\examples{
NULL
}
