% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_chemo.R
\name{decorate_chemo}
\alias{decorate_chemo}
\title{Decorate chemo}
\usage{
decorate_chemo(
  annot_table_weighted_chemo = get("annot_table_weighted_chemo", envir = parent.frame()),
  score_chemical_cla_kingdom = get("score_chemical_cla_kingdom", envir = parent.frame()),
  score_chemical_cla_superclass = get("score_chemical_cla_superclass", envir =
    parent.frame()),
  score_chemical_cla_class = get("score_chemical_cla_class", envir = parent.frame()),
  score_chemical_cla_parent = get("score_chemical_cla_parent", envir = parent.frame()),
  score_chemical_npc_pathway = get("score_chemical_npc_pathway", envir = parent.frame()),
  score_chemical_npc_superclass = get("score_chemical_npc_superclass", envir =
    parent.frame()),
  score_chemical_npc_class = get("score_chemical_npc_class", envir = parent.frame())
)
}
\arguments{
\item{annot_table_weighted_chemo}{Table to decorate}

\item{score_chemical_cla_kingdom}{Classyfire kingdom score}

\item{score_chemical_cla_superclass}{Classyfire superclass score}

\item{score_chemical_cla_class}{Classyfire class score}

\item{score_chemical_cla_parent}{Classyfire parent score}

\item{score_chemical_npc_pathway}{NPC pathway score}

\item{score_chemical_npc_superclass}{NPC superclass score}

\item{score_chemical_npc_class}{NPC class score}
}
\value{
Message indicating the number of annotations weighted at each chemical level
}
\description{
This function outputs information about chemical weighting
}
\examples{
NULL
}
