% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_chemo.R
\name{decorate_chemo}
\alias{decorate_chemo}
\title{Decorate chemo}
\usage{
decorate_chemo(
  df = annotation_table_weighted_chemo,
  sc_cla_kin = score_chemical_cla_kingdom,
  sc_cla_sup = score_chemical_cla_superclass,
  sc_cla_cla = score_chemical_cla_class,
  sc_cla_par = score_chemical_cla_parent,
  sc_npc_pat = score_chemical_npc_pathway,
  sc_npc_sup = score_chemical_npc_superclass,
  sc_npc_cla = score_chemical_npc_class
)
}
\arguments{
\item{df}{Table to decorate}

\item{sc_cla_kin}{Classyfire kingdom score}

\item{sc_cla_sup}{Classyfire superclass score}

\item{sc_cla_cla}{Classyfire class score}

\item{sc_cla_par}{Classyfire parent score}

\item{sc_npc_pat}{NPC pathway score}

\item{sc_npc_sup}{NPC superclass score}

\item{sc_npc_cla}{NPC class score}
}
\value{
Message indicating the number of annotations weighted at each chemical level
}
\description{
This function outputs information about chemical weighting
}
\examples{
NULL
}
