% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_chemo.R
\name{decorate_chemo}
\alias{decorate_chemo}
\title{Decorate chemo}
\usage{
decorate_chemo(
  df = annotation_table_weighted_chemo,
  sc_pat = score_chemical_pathway,
  sc_sup = score_chemical_superclass,
  sc_cla = score_chemical_class
)
}
\arguments{
\item{df}{Table to decorate}

\item{sc_pat}{Pathway score}

\item{sc_sup}{Superclass score}

\item{sc_cla}{Class score}
}
\value{
Message indicating the number of annotations weighted at each chemical level
}
\description{
This function outputs information about chemical weighting
}
\examples{
NULL
}
