% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_masses.R
\name{decorate_masses}
\alias{decorate_masses}
\title{Decorate masses}
\usage{
decorate_masses(
  annotation_table_ms1 = get("annotation_table_ms1", envir = parent.frame())
)
}
\arguments{
\item{annotation_table_ms1}{Data frame containing MS1 annotation results
with columns for feature_id and candidate_structure_inchikey_connectivity_layer}
}
\value{
The input annotation table (unchanged), for use in pipelines
}
\description{
This function logs summary statistics about MS1-based
annotations, reporting the number of unique structures and features
annotated
}
\examples{
NULL
}
