% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_masses.R
\name{decorate_masses}
\alias{decorate_masses}
\title{Decorate masses}
\usage{
decorate_masses(df = annotation_table_ms1)
}
\arguments{
\item{df}{Table to decorate}
}
\value{
Message indicating the number of annotations obtained by MS1
}
\description{
This function outputs information about MS1 annotation
}
\examples{
NULL
}
