% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_ms1}
\alias{decorate_ms1}
\title{Decorate MS1}
\usage{
decorate_ms1()
}
\value{
Message indicating the number of annotations obtained by MS1
}
\description{
This function outputs informations about MS1 annotation
}
\examples{
NULL
}
