% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_BATCH_SIZE}
\alias{DEFAULT_BATCH_SIZE}
\title{Default batch size for parallel processing}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_BATCH_SIZE
}
\description{
Default batch size for parallel processing
}
\keyword{datasets}
