% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_BEST_PERCENTILE}
\alias{DEFAULT_BEST_PERCENTILE}
\title{Default percentile threshold for top candidate selection
Only candidates with scores above this percentile are retained}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_BEST_PERCENTILE
}
\description{
Default percentile threshold for top candidate selection
Only candidates with scores above this percentile are retained
}
\keyword{internal}
