% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CANDIDATES_NEIGHBORS}
\alias{DEFAULT_CANDIDATES_NEIGHBORS}
\title{Default number of neighboring candidates to retain during network annotation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_CANDIDATES_NEIGHBORS
}
\description{
Default number of neighboring candidates to retain during network annotation
}
\keyword{datasets}
