% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HC_SCORE_BIO_MIN}
\alias{DEFAULT_HC_SCORE_BIO_MIN}
\title{Minimum biological score for high confidence annotations}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_HC_SCORE_BIO_MIN

DEFAULT_HC_SCORE_BIO_MIN
}
\description{
Minimum biological score for high confidence annotations

Default minimum biological score threshold for high-confidence filter
}
\keyword{datasets}
