% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HC_SCORE_FINAL_MIN}
\alias{DEFAULT_HC_SCORE_FINAL_MIN}
\title{Default minimum final (weighted chemical) score threshold for high-confidence filter}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_HC_SCORE_FINAL_MIN
}
\description{
Default minimum final (weighted chemical) score threshold for high-confidence filter
}
\keyword{datasets}
