% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HC_SCORE_MIN_PEAKS}
\alias{DEFAULT_HC_SCORE_MIN_PEAKS}
\title{Default minimum number of matched peaks for high-confidence filter}
\format{
An object of class \code{integer} of length 1.
}
\usage{
DEFAULT_HC_SCORE_MIN_PEAKS
}
\description{
Default minimum number of matched peaks for high-confidence filter
}
\keyword{internal}
