% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HC_SCORE_SIRIUS_MIN}
\alias{DEFAULT_HC_SCORE_SIRIUS_MIN}
\title{Default minimum SIRIUS confidence score threshold for high-confidence filter}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_HC_SCORE_SIRIUS_MIN
}
\description{
Default minimum SIRIUS confidence score threshold for high-confidence filter
}
\keyword{internal}
