% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HC_SCORE_SPECTRAL_MIN}
\alias{DEFAULT_HC_SCORE_SPECTRAL_MIN}
\title{Default minimum SIRIUS confidence score threshold for high-confidence filter}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
DEFAULT_HC_SCORE_SPECTRAL_MIN
}
\description{
Default minimum SIRIUS confidence score threshold for high-confidence filter
}
\keyword{datasets}
