% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_LOG_LEVEL}
\alias{DEFAULT_LOG_LEVEL}
\title{Default log level for package operations}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_LOG_LEVEL
}
\description{
Default log level for package operations
}
\keyword{datasets}
