% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_MINIMAL_CONSISTENCY}
\alias{DEFAULT_MINIMAL_CONSISTENCY}
\title{Minimum consistency score threshold for taxonomic predictions}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_MINIMAL_CONSISTENCY
}
\description{
Minimum consistency score threshold for taxonomic predictions
}
\keyword{internal}
