% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_MINIMAL_MS1_CHEMO}
\alias{DEFAULT_MINIMAL_MS1_CHEMO}
\title{Minimum MS1 chemical score threshold}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_MINIMAL_MS1_CHEMO
}
\description{
Minimum MS1 chemical score threshold
}
\keyword{internal}
