% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_PROGRESS_INTERVAL}
\alias{DEFAULT_PROGRESS_INTERVAL}
\title{Default progress reporting interval (number of items)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_PROGRESS_INTERVAL
}
\description{
Default progress reporting interval (number of items)
}
\keyword{internal}
