% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_groups.R
\name{dist_get}
\alias{dist_get}
\title{Distance between two elements in a distance matrix}
\usage{
dist_get(d, idx1, idx2)
}
\arguments{
\item{d}{Distance matrix}

\item{idx1}{Index of the first element}

\item{idx2}{Index of the second element}
}
\value{
Distance between the two elements
}
\description{
This function calculates the distance between
two elements in a distance matrix
}
\details{
Credit goes to usedist package
}
\examples{
NULL
}
