% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_groups.R
\name{dist_groups}
\alias{dist_groups}
\title{Dist groups}
\usage{
dist_groups(d, g)
}
\arguments{
\item{d}{A distance object}

\item{g}{A grouping vector for the distance object}
}
\value{
A data frame containing distance information between pairs of observations
in the distance object, with columns for the names or indices of the observations,
the group labels for each observation, and the distance between the observations.
The label column indicates whether the distance is within a group or between groups.
}
\description{
Dist groups
}
\examples{
TODO
}
