% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mztab.R
\name{.mztab_build_sml}
\alias{.mztab_build_sml}
\title{Build SML (Small Molecule Summary) table – one row per unique compound}
\usage{
.mztab_build_sml(results, smf_table, sme_table, xrefs_index = NULL)
}
\description{
Build SML (Small Molecule Summary) table – one row per unique compound
}
\keyword{internal}
