% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_output.R
\name{export_output}
\alias{export_output}
\title{Export output}
\usage{
export_output(x, file = output, ...)
}
\arguments{
\item{x}{data frame to be exported}

\item{file}{path to the output file}

\item{...}{...}
}
\description{
This function creates the output directory if it doesn't exist and exports the data frame to a tab-delimited file.
}
\examples{
export_output(x, file = "output/file.tsv")
}
