% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_params.R
\name{export_params}
\alias{export_params}
\title{Export parameters}
\usage{
export_params(
  parameters = get("parameters", envir = parent.frame()),
  directory = parse_yaml_paths()$data$interim$params$path,
  step
)
}
\arguments{
\item{parameters}{list of parameters to be exported}

\item{directory}{directory where the YAML file will be saved}

\item{step}{step identifier to be included in the YAML file name}
}
\description{
This function writes the parameters
to a YAML file in the specified directory.
}
\examples{
NULL
}
