% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_spectra.R
\name{export_spectra}
\alias{export_spectra}
\title{Export spectra}
\usage{
export_spectra(file, spectra, dir = ".", cmps = NULL, metad = NULL)
}
\arguments{
\item{file}{File where spectra will be exported. Can be '.mgf' or '.sqlite'}

\item{spectra}{The spectra object where spectra are stored}

\item{dir}{Directory for export}

\item{cmps}{Optional. Compounds data frame as in 'CompoundDb' package}

\item{metad}{Optional. Metadata as in 'CompoundDb' package}
}
\description{
TODO
}
\examples{
NULL
}
