% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_spectra.R
\name{export_spectra_2}
\alias{export_spectra_2}
\title{Export spectra 2}
\usage{
export_spectra_2(file, spectra, meta)
}
\arguments{
\item{file}{File where spectra will be exported. Can be '.mgf' or '.sqlite'}

\item{spectra}{The spectra object where spectra are stored}

\item{meta}{Optional. Metadata as in 'CompoundDb' package}
}
\description{
This function export spectra.
}
\examples{
NULL
}
