% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spectra.R
\name{extract_spectra}
\alias{extract_spectra}
\title{Extract spectra}
\usage{
extract_spectra(object)
}
\arguments{
\item{object}{TODO}
}
\value{
TODO
}
\description{
Extract spectra
}
\examples{
TODO
}
