% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spectra.R
\name{extract_spectra}
\alias{extract_spectra}
\title{Extract spectra from a Spectra object}
\usage{
extract_spectra(object)
}
\arguments{
\item{object}{Object of class Spectra}
}
\value{
Data frame containing spectra data
}
\description{
Extract spectra from a Spectra object
}
\examples{
extract_spectra(object)
}
