% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_edges.R
\name{fake_edges}
\alias{fake_edges}
\title{Fake edges}
\usage{
fake_edges(
  input = params$input,
  output = params$output,
  name_feature = params$feature
)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}

\item{name_feature}{TODO}
}
\description{
Fake edges
}
\examples{
TODO
}
