% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_features_components.R
\name{fake_features_components}
\alias{fake_features_components}
\title{Fake features components}
\usage{
fake_features_components(
  input = params$input,
  features = params$features,
  output = params$output,
  ms_mode = params$mode,
  name_rt = params$rt_name,
  name_mz = params$precursor_name
)
}
\arguments{
\item{input}{TODO}

\item{features}{TODO}

\item{output}{TODO}

\item{ms_mode}{TODO}

\item{name_rt}{TODO}

\item{name_mz}{TODO}
}
\value{
TODO
}
\description{
Fake features components
}
\examples{
TODO
}
