% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_annotations.R
\name{filter_annotations}
\alias{filter_annotations}
\title{Filter annotations}
\usage{
filter_annotations(
  annotations = params$files$annotations$prepared,
  features = params$files$features$prepared,
  rts = params$files$libraries$temporal$prepared,
  output = params$files$annotations$filtered,
  tolerance_rt = params$ms$tolerances$rt$minutes,
  parameters = params
)
}
\arguments{
\item{annotations}{Prepared annotations file}

\item{features}{Prepared features file}

\item{rts}{Prepared retention time library}

\item{output}{Output file}

\item{tolerance_rt}{Tolerance to filter retention time}

\item{parameters}{Params}
}
\description{
This function filters initial annotations.
}
\examples{
NULL
}
