% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_annotations.R
\name{filter_annotations}
\alias{filter_annotations}
\title{Filter annotations}
\usage{
filter_annotations(
  annotations = get_params(step =
    "filter_annotations")$files$annotations$prepared$structural,
  features = get_params(step = "filter_annotations")$files$features$prepared,
  rts = get_params(step = "filter_annotations")$files$libraries$temporal$prepared,
  output = get_params(step = "filter_annotations")$files$annotations$filtered,
  tolerance_rt = get_params(step = "filter_annotations")$ms$tolerances$rt$minutes
)
}
\arguments{
\item{annotations}{Prepared annotations file}

\item{features}{Prepared features file}

\item{rts}{Prepared retention time library}

\item{output}{Output file}

\item{tolerance_rt}{Tolerance to filter retention time}
}
\description{
This function filters initial annotations.
}
\examples{
NULL
}
