% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_annotations.R
\name{filter_annotations}
\alias{filter_annotations}
\title{Filter annotations}
\usage{
filter_annotations(
  annotations = get_params(step =
    "filter_annotations")$files$annotations$prepared$structural,
  features = get_params(step = "filter_annotations")$files$features$prepared,
  rts = get_params(step = "filter_annotations")$files$libraries$temporal$prepared,
  output = get_params(step = "filter_annotations")$files$annotations$filtered,
  tolerance_rt = get_params(step = "filter_annotations")$ms$tolerances$rt$library
)
}
\arguments{
\item{annotations}{Prepared annotations file}

\item{features}{Prepared features file}

\item{rts}{Prepared retention time library}

\item{output}{Output file}

\item{tolerance_rt}{Tolerance to filter retention time}
}
\value{
The path to the filtered annotations
}
\description{
This function filters initial annotations.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
github <- "https://raw.githubusercontent.com/"
repo <- "taxonomicallyinformedannotation/tima-example-files/main/"
dir <- paste0(github, repo)
annotations <- get_params(step = "filter_annotations")$files$annotations$prepared$structural[[2]] |>
  gsub(
    pattern = ".gz",
    replacement = "",
    fixed = TRUE
  )
features <- get_params(step = "filter_annotations")$files$features$prepared |>
  gsub(
    pattern = ".gz",
    replacement = "",
    fixed = TRUE
  )
rts <- get_params(step = "filter_annotations")$files$libraries$temporal$prepared |>
  gsub(
    pattern = ".gz",
    replacement = "",
    fixed = TRUE
  )
get_file(url = paste0(dir, annotations), export = annotations)
get_file(url = paste0(dir, features), export = features)
get_file(url = paste0(dir, rts), export = rts)
filter_annotations(
  annotations = annotations,
  features = features,
  rts = rts
)
unlink("data", recursive = TRUE)
}
}
