% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_annotations.R
\name{filter_ms1_redundancy}
\alias{filter_ms1_redundancy}
\title{Filter MS1 Annotations}
\usage{
filter_ms1_redundancy(annotation_tables_list)
}
\arguments{
\item{annotation_tables_list}{Named list of annotation tables}
}
\value{
Data frame with filtered annotations
}
\description{
Internal helper to remove MS1 annotations that have spectral matches.
}
\keyword{internal}
