% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_adducts_neg.R
\name{form_adducts_neg}
\alias{form_adducts_neg}
\title{Form adducts negative}
\usage{
form_adducts_neg(massesTable, adductsTable)
}
\arguments{
\item{massesTable}{TODO}

\item{adductsTable}{TODO}
}
\value{
TODO
}
\description{
Form adducts negative
}
\examples{
TODO
}
