% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark.R
\name{get_benchmark}
\alias{get_benchmark}
\title{Title}
\usage{
get_benchmark(
  link = paths$links$benchmarking_set,
  export = paths$data$source$benchmark$set
)
}
\arguments{
\item{link}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Title
}
