% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark.R
\name{get_benchmark}
\alias{get_benchmark}
\title{Get benchmark}
\usage{
get_benchmark(
  url = paths$urls$benchmarking_set,
  export = paths$data$source$benchmark$set
)
}
\arguments{
\item{url}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Get benchmark
}
\examples{
TODO
}
