% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_paths.R
\name{get_default_paths}
\alias{get_default_paths}
\title{Get default paths}
\usage{
get_default_paths(yaml = system.file("paths.yaml", package = "tima"))
}
\arguments{
\item{yaml}{The YAML file containing the paths
(default is "paths.yaml")}
}
\value{
A list containing the paths specified in the YAML file
}
\description{
This function gets default paths
}
\examples{
get_default_paths()
}
