% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_feature_table.R
\name{get_example_feature_table}
\alias{get_example_feature_table}
\title{Title}
\usage{
get_example_feature_table(
  url = paths$urls$example_feature_table,
  export = paths$data$interim$annotations$example_feature_table
)
}
\arguments{
\item{url}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Title
}
