% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_files.R
\name{get_example_files}
\alias{get_example_files}
\title{Get example files}
\usage{
get_example_files(
  example = c("features", "metadata", "sirius", "spectra"),
  in_cache = TRUE
)
}
\arguments{
\item{example}{The example(s) you want to download}

\item{in_cache}{Flag to indicate if storing the files in cache}
}
\value{
Example files.
}
\description{
This function downloads example files
}
\examples{
get_example_files(example = c("features"), in_cache = FALSE)
unlink("data", recursive = TRUE)
}
