% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_isdb.R
\name{get_example_isdb}
\alias{get_example_isdb}
\title{Title}
\usage{
get_example_isdb(
  link = paths$links$example_isdb,
  export = paths$data$interim$annotations$example_isdb
)
}
\arguments{
\item{link}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
