% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_isdb.R
\name{get_example_isdb}
\alias{get_example_isdb}
\title{Get example ISDB}
\usage{
get_example_isdb(
  url = paths$urls$example_isdb,
  export = paths$data$interim$annotations$example_isdb
)
}
\arguments{
\item{url}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Get example ISDB
}
\examples{
TODO
}
