% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_mgf.R
\name{get_example_mgf}
\alias{get_example_mgf}
\title{Get example MGF}
\usage{
get_example_mgf(
  url = paths$url$examples$mgf,
  export = paths$data$source$examples$spectra
)
}
\arguments{
\item{url}{URL of the MGF file to be downloaded}

\item{export}{File path where the MGF file should be saved}
}
\description{
This function gets an example mgf file to work with
}
\examples{
NULL
}
