% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_mgf.R
\name{get_example_mgf}
\alias{get_example_mgf}
\title{Get example MGF}
\usage{
get_example_mgf(
  url = paths$url$example_mgf,
  export = paths$data$source$example$mgf
)
}
\arguments{
\item{url}{URL of the MGF file to be downloaded}

\item{export}{File path where the MGF file should be saved}
}
\description{
TODO
}
\examples{
NULL
}
