% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_sirius.R
\name{get_example_sirius}
\alias{get_example_sirius}
\title{Get example sirius}
\usage{
get_example_sirius(
  url = paths$url$example_sirius,
  export = paths$data$interim$annotations$example_sirius
)
}
\arguments{
\item{url}{the URL of the example sirius file to be downloaded}

\item{export}{the file path where the downloaded sirius file should be saved}
}
\description{
Get example sirius
}
\examples{
TODO
}
