% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_sirius.R
\name{get_example_sirius}
\alias{get_example_sirius}
\title{Title}
\usage{
get_example_sirius(
  link = paths$links$example_sirius,
  export = paths$data$interim$annotations$example_sirius
)
}
\arguments{
\item{link}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Title
}
