% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_sirius.R
\name{get_example_sirius}
\alias{get_example_sirius}
\title{Get example sirius}
\usage{
get_example_sirius(
  url = get_default_paths()$urls$examples$sirius,
  export = get_default_paths()$data$interim$annotations$example_sirius
)
}
\arguments{
\item{url}{URL where the example is accessible}

\item{export}{Path where to save the example}
}
\description{
This function gets example SIRIUS annotations
}
\examples{
NULL
}
