% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Get file}
\usage{
get_file(url, export, limit = 3600L)
}
\arguments{
\item{url}{Character string URL of the file to be downloaded}

\item{export}{Character string file path where the file should be saved}

\item{limit}{Integer timeout limit in seconds (default: 3600)}
}
\value{
The path to the downloaded file
}
\description{
This function downloads files from a URL with retry logic
and exponential backoff
}
\examples{
\dontrun{
git <- "https://github.com/"
org <- "taxonomicallyinformedannotation"
repo <- "tima-example-files"
branch <- "main"
file <- "example_metadata.tsv"
get_file(
  url = paste(git, org, repo, "raw", branch, file, sep = "/"),
  export = "data/source/example_metadata.tsv"
)
}
}
