% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Get file}
\usage{
get_file(url, export, limit = 1000)
}
\arguments{
\item{url}{URL of the file to be downloaded}

\item{export}{File path where the file should be saved}

\item{limit}{Timeout limit (in seconds)}
}
\description{
This function get files
}
\examples{
get_file(url = "https://github.com/taxonomicallyinformedannotation/tima-example-files/raw/main/example_metadata.tsv", export = "data/source/example_metadata.tsv")
}
