% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnps_tables.R
\name{get_gnps_tables}
\alias{get_gnps_tables}
\title{Get GNPS Tables}
\usage{
get_gnps_tables(
  gnps_job_id,
  gnps_job_example = get_default_paths()$gnps$example,
  filename,
  workflow = "fbmn",
  path_features,
  path_metadata,
  path_spectra,
  path_source = get_default_paths()$data$source$path,
  path_interim_a = get_default_paths()$data$interim$annotations$path,
  path_interim_f = get_default_paths()$data$interim$features$path
)
}
\arguments{
\item{gnps_job_id}{GNPS job ID}

\item{gnps_job_example}{GNPS job example}

\item{filename}{Name of the file}

\item{workflow}{Character string indicating the type of workflow,
either "fbmn" or "classical"}

\item{path_features}{Path to features}

\item{path_metadata}{Path to metadata}

\item{path_spectra}{Path to spectra}

\item{path_source}{Path to store the source files}

\item{path_interim_a}{Path to store the interim annotations file}

\item{path_interim_f}{Path to store the interim features files}
}
\value{
The downloaded GNPS tables
}
\description{
This function gets GNPS tables from corresponding job ID.
}
\examples{
NULL
}
