% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_version_from_zenodo.R
\name{get_last_version_from_zenodo}
\alias{get_last_version_from_zenodo}
\title{Get last version from Zenodo}
\usage{
get_last_version_from_zenodo(doi, pattern, path)
}
\arguments{
\item{doi}{DOI of the Zenodo record}

\item{pattern}{Pattern to identify the file to download}

\item{path}{Path to save the file to}
}
\value{
The path to the file
}
\description{
This function gets the last version
of a file from a Zenodo record
}
\details{
Credit goes to partially to
https://inbo.github.io/inborutils/
}
\examples{
NULL
}
