% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_version_from_zenodo.R
\name{get_last_version_from_zenodo}
\alias{get_last_version_from_zenodo}
\title{Get last version from Zenodo}
\usage{
get_last_version_from_zenodo(url, pattern, delim = ",")
}
\arguments{
\item{url}{TODO}

\item{pattern}{TODO}

\item{delim}{TODO}
}
\value{
TODO
}
\description{
Get last version from Zenodo
}
\examples{
TODO
}
