% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lotus.R
\name{get_lotus}
\alias{get_lotus}
\title{Get LOTUS}
\usage{
get_lotus(url = paths$urls$lotus, export = paths$data$source$libraries$lotus)
}
\arguments{
\item{url}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Get LOTUS
}
\examples{
TODO
}
