% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lotus.R
\name{get_lotus}
\alias{get_lotus}
\title{Title}
\usage{
get_lotus(
  input = "https://osf.io/rheq5/download",
  output = "data/source/libraries/lotus.csv.gz"
)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}
}
\value{
TODO
}
\description{
Title
}
