% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lotus.R
\name{get_lotus}
\alias{get_lotus}
\title{Title}
\usage{
get_lotus(link = paths$links$lotus, export = paths$data$source$libraries$lotus)
}
\arguments{
\item{link}{TODO}

\item{export}{TODO}
}
\value{
TODO
}
\description{
Title
}
