% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_massbank_spectra.R
\name{get_massbank_spectra}
\alias{get_massbank_spectra}
\title{Get MassBank spectra}
\usage{
get_massbank_spectra(output_dir = "data/source/libraries/spectra/exp/")
}
\arguments{
\item{output_dir}{Output where to store the spectra}
}
\description{
This function gets MassBank spectra
}
\examples{
NULL
}
