% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get parameters}
\usage{
get_params(step)
}
\arguments{
\item{step}{Name of the step being performed}
}
\value{
The parameters
}
\description{
This function gets the parameters for the job.
Combination of cli and yaml parameters
}
\examples{
NULL
}
