% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra_ids.R
\name{get_spectra_ids}
\alias{get_spectra_ids}
\title{Get Spectra ids}
\usage{
get_spectra_ids(spectra)
}
\arguments{
\item{spectra}{Spectra.}
}
\description{
This function extracts spectra IDs as they are yet not consistently named
}
\examples{
NULL
}
