% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wrappers.R
\name{gnps_wrapper}
\alias{gnps_wrapper}
\title{Wrapper for the C function "gnps"}
\usage{
gnps_wrapper(x, y)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric matrix.}
}
\value{
The result from the C function.
}
\description{
Wrapper for the C function "gnps"
}
