% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_to_cache.R
\name{go_to_cache}
\alias{go_to_cache}
\title{Go to cache}
\usage{
go_to_cache(dir = ".tima")
}
\arguments{
\item{dir}{Directory}
}
\value{
Goes to cache
}
\description{
This function goes to cache
}
\examples{
go_to_cache()
}
