% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{harmonize_adducts}
\alias{harmonize_adducts}
\title{Harmonize adducts}
\usage{
harmonize_adducts(df, adducts_colname = "adduct", adducts_translations)
}
\arguments{
\item{df}{Dataframe}

\item{adducts_colname}{Adducts colname}

\item{adducts_translations}{Adducts translations}
}
\value{
A table with harmonized adducts
}
\description{
This function annotates masses
}
\examples{
NULL
}
