% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_adducts.R
\name{harmonize_adducts}
\alias{harmonize_adducts}
\title{Harmonize adducts}
\usage{
harmonize_adducts(df, adducts_colname = "adduct", adducts_translations)
}
\arguments{
\item{df}{Dataframe containing adduct column to harmonize}

\item{adducts_colname}{Character string name of the adduct column
(default: "adduct")}

\item{adducts_translations}{Named character vector mapping original
adduct notations (names) to standardized forms (values)}
}
\value{
The dataframe with harmonized adduct column
}
\description{
This function harmonizes adduct definitions by replacing
various adduct notations with standardized forms according to a
translation table
}
\examples{
NULL
}
