% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_spectra.R
\name{harmonize_spectra}
\alias{harmonize_spectra}
\title{Harmonize spectra}
\usage{
harmonize_spectra(
  spectra,
  metad,
  mode,
  col_ad,
  col_ce,
  col_ci,
  col_em,
  col_in,
  col_io,
  col_ik,
  col_il,
  col_mf,
  col_na,
  col_po,
  col_sm,
  col_sn,
  col_si,
  col_sp,
  col_sy,
  col_xl
)
}
\arguments{
\item{spectra}{Spectra object to be harmonized}

\item{metad}{Metadata to identify the library}

\item{mode}{MS ionization mode. Must contain 'pos' or 'neg'}

\item{col_ad}{Name of the adduct in mgf}

\item{col_ce}{Name of the collision energy in mgf}

\item{col_ci}{Name of the compound id in mgf}

\item{col_em}{Name of the exact mass in mgf}

\item{col_in}{Name of the InChI in mgf}

\item{col_io}{Name of the InChI without stereo in mgf}

\item{col_ik}{Name of the InChIKey in mgf}

\item{col_il}{Name of the InChIKey without stereo in mgf}

\item{col_mf}{Name of the molecular formula in mgf}

\item{col_na}{Name of the name in mgf}

\item{col_po}{Name of the polarity in mgf}

\item{col_sm}{Name of the SMILES in mgf}

\item{col_sn}{Name of the SMILES without stereo in mgf}

\item{col_si}{Name of the spectrum id in mgf}

\item{col_sp}{Name of the SPLASH in mgf}

\item{col_sy}{Name of the synonyms in mgf}

\item{col_xl}{Name of the xlogp in mgf}
}
\description{
This function harmonizes spectra headers
}
\examples{
NULL
}
