% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_spectra.R
\name{harmonize_spectra}
\alias{harmonize_spectra}
\title{Harmonize spectra}
\usage{
harmonize_spectra(
  spectra,
  colnames = c(colname_collision_energy = NA, colname_compound_id = NA,
    colname_exact_mass = "EXACTMASS", colname_formula = "MOLECULAR_FORMULA",
    colname_inchi = "INCHI", colname_inchikey = "NAME", colname_mode = "IONMODE",
    colname_name = "NAME", colname_precursorMz = "precursorMz", colname_precursorCharge =
    "precursorCharge", colname_smiles = "SMILES", colname_spectrum_id = NA,
    colname_splash = NA, colname_synonyms = NA),
  mode = "pos"
)
}
\arguments{
\item{spectra}{Spectra object to be harmonized}

\item{colnames}{Names of the different columns corresponding to standardized headers}

\item{mode}{MS ionization mode. Must contain 'pos' or 'neg'}
}
\description{
This function harmonizes spectra headers
}
\examples{
NULL
}
