% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_spectra.R
\name{harmonize_spectra}
\alias{harmonize_spectra}
\title{Harmonize spectra}
\usage{
harmonize_spectra(
  spectra,
  metad = get("metad", envir = parent.frame()),
  mode,
  col_ad = get("col_ad", envir = parent.frame()),
  col_ce = get("col_ce", envir = parent.frame()),
  col_ci = get("col_ci", envir = parent.frame()),
  col_em = get("col_em", envir = parent.frame()),
  col_in = get("col_in", envir = parent.frame()),
  col_io = get("col_io", envir = parent.frame()),
  col_ik = get("col_ik", envir = parent.frame()),
  col_il = get("col_il", envir = parent.frame()),
  col_mf = get("col_mf", envir = parent.frame()),
  col_na = get("col_na", envir = parent.frame()),
  col_po = get("col_po", envir = parent.frame()),
  col_sm = get("col_sm", envir = parent.frame()),
  col_sn = get("col_sn", envir = parent.frame()),
  col_si = get("col_si", envir = parent.frame()),
  col_sp = get("col_sp", envir = parent.frame()),
  col_sy = get("col_sy", envir = parent.frame()),
  col_xl = get("col_xl", envir = parent.frame())
)
}
\arguments{
\item{spectra}{Data frame containing spectra data to be harmonized}

\item{metad}{Metadata to identify the library source}

\item{mode}{MS ionization mode. Must contain 'pos' or 'neg'}

\item{col_ad}{Name of the adduct field in the input spectra}

\item{col_ce}{Name of the collision energy field}

\item{col_ci}{Name of the compound ID field}

\item{col_em}{Name of the exact mass field}

\item{col_in}{Name of the InChI field}

\item{col_io}{Name of the InChI without stereochemistry field}

\item{col_ik}{Name of the InChIKey field}

\item{col_il}{Name of the InChIKey connectivity layer field}

\item{col_mf}{Name of the molecular formula field}

\item{col_na}{Name of the compound name field}

\item{col_po}{Name of the polarity field}

\item{col_sm}{Name of the SMILES field}

\item{col_sn}{Name of the SMILES without stereochemistry field}

\item{col_si}{Name of the spectrum ID field}

\item{col_sp}{Name of the SPLASH field}

\item{col_sy}{Name of the synonyms field}

\item{col_xl}{Name of the xLogP field}
}
\value{
A data frame with harmonized column names and standardized field names
}
\description{
This function harmonizes spectra header fields by mapping
library-specific column names to standardized field names. It handles
different naming conventions across spectral libraries and ensures
consistent metadata structure.
}
\examples{
NULL
}
