% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_spectra.R
\name{harmonize_spectra}
\alias{harmonize_spectra}
\title{Harmonize spectra}
\usage{
harmonize_spectra(
  spectra,
  mode,
  co_ce,
  co_ci,
  co_em,
  co_in,
  co_io,
  co_ik,
  co_il,
  co_mf,
  co_na,
  co_po,
  co_sm,
  co_sn,
  co_si,
  co_sp,
  co_sy,
  co_xl
)
}
\arguments{
\item{spectra}{Spectra object to be harmonized}

\item{mode}{MS ionization mode. Must contain 'pos' or 'neg'}

\item{co_ce}{Name of the collision energy in mgf}

\item{co_ci}{Name of the compound id in mgf}

\item{co_em}{Name of the exact mass in mgf}

\item{co_in}{Name of the InChI in mgf}

\item{co_io}{Name of the InChI 2D in mgf}

\item{co_ik}{Name of the InChIKey in mgf}

\item{co_il}{Name of the InChIKey 2D in mgf}

\item{co_mf}{Name of the molecular formula in mgf}

\item{co_na}{Name of the name in mgf}

\item{co_po}{Name of the polarity in mgf}

\item{co_sm}{Name of the SMILES in mgf}

\item{co_sn}{Name of the SMILES 2D in mgf}

\item{co_si}{Name of the spectrum id in mgf}

\item{co_sp}{Name of the SPLASH in mgf}

\item{co_sy}{Name of the synonyms in mgf}

\item{co_xl}{Name of the xlogp in mgf}
}
\description{
This function harmonizes spectra headers
}
\examples{
NULL
}
