% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_spectra.R
\name{harmonize_spectra}
\alias{harmonize_spectra}
\title{Harmonize spectra}
\usage{
harmonize_spectra(
  spectra,
  mode,
  col_ce = get("col_ce", envir = parent.frame()),
  col_ci = get("col_ci", envir = parent.frame()),
  col_em = get("col_em", envir = parent.frame()),
  col_in = get("col_in", envir = parent.frame()),
  col_io = get("col_io", envir = parent.frame()),
  col_ik = get("col_ik", envir = parent.frame()),
  col_il = get("col_il", envir = parent.frame()),
  col_mf = get("col_mf", envir = parent.frame()),
  col_na = get("col_na", envir = parent.frame()),
  col_po = get("col_po", envir = parent.frame()),
  col_sm = get("col_sm", envir = parent.frame()),
  col_sn = get("col_sn", envir = parent.frame()),
  col_si = get("col_si", envir = parent.frame()),
  col_sp = get("col_sp", envir = parent.frame()),
  col_sy = get("col_sy", envir = parent.frame()),
  col_xl = get("col_xl", envir = parent.frame())
)
}
\arguments{
\item{spectra}{Spectra object to be harmonized}

\item{mode}{MS ionization mode. Must contain 'pos' or 'neg'}

\item{col_ce}{Name of the collision energy in mgf}

\item{col_ci}{Name of the compound id in mgf}

\item{col_em}{Name of the exact mass in mgf}

\item{col_in}{Name of the InChI in mgf}

\item{col_io}{Name of the InChI 2D in mgf}

\item{col_ik}{Name of the InChIKey in mgf}

\item{col_il}{Name of the InChIKey 2D in mgf}

\item{col_mf}{Name of the molecular formula in mgf}

\item{col_na}{Name of the name in mgf}

\item{col_po}{Name of the polarity in mgf}

\item{col_sm}{Name of the SMILES in mgf}

\item{col_sn}{Name of the SMILES 2D in mgf}

\item{col_si}{Name of the spectrum id in mgf}

\item{col_sp}{Name of the SPLASH in mgf}

\item{col_sy}{Name of the synonyms in mgf}

\item{col_xl}{Name of the xlogp in mgf}
}
\description{
This function harmonizes spectra headers
}
\examples{
NULL
}
