% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_spectra.R
\name{harmonize_spectra_polarity}
\alias{harmonize_spectra_polarity}
\title{Harmonize Spectra for a Given Polarity}
\usage{
harmonize_spectra_polarity(
  spectra_extracted,
  mode,
  metad,
  col_ad,
  col_ce,
  col_ci,
  col_em,
  col_in,
  col_io,
  col_ik,
  col_il,
  col_mf,
  col_na,
  col_po,
  col_sm,
  col_sn,
  col_si,
  col_sp,
  col_sy,
  col_xl
)
}
\arguments{
\item{spectra_extracted}{List of extracted spectra data frames.}

\item{mode}{Character "pos" or "neg".}

\item{metad}{Character library metadata name.}

\item{col_ad}{Adduct field name}

\item{col_ce}{Collision energy field name}

\item{col_ci}{Compound ID field name}

\item{col_em}{Exact mass field name}

\item{col_in}{InChI field name}

\item{col_io}{InChI without stereochemistry field name}

\item{col_ik}{InChIKey field name}

\item{col_il}{InChIKey connectivity layer field name}

\item{col_mf}{Molecular formula field name}

\item{col_na}{Compound name field name}

\item{col_po}{Polarity field name}

\item{col_sm}{SMILES field name}

\item{col_sn}{SMILES without stereochemistry field name}

\item{col_si}{Spectrum ID field name}

\item{col_sp}{SPLASH field name}

\item{col_sy}{Synonyms field name}

\item{col_xl}{xLogP field name}
}
\value{
Data frame with harmonized spectra.
}
\description{
Internal helper to harmonize extracted spectra for a specific
polarity mode, applying column mappings and fixing precursor/inchikey issues.
}
\keyword{internal}
