% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_spectra.R
\name{harmonize_spectra_polarity}
\alias{harmonize_spectra_polarity}
\title{Harmonize Spectra for a Given Polarity}
\usage{
harmonize_spectra_polarity(
  spectra_extracted,
  mode,
  metad,
  col_ad,
  col_ce,
  col_ci,
  col_em,
  col_in,
  col_io,
  col_ik,
  col_il,
  col_mf,
  col_na,
  col_po,
  col_sm,
  col_sn,
  col_si,
  col_sp,
  col_sy,
  col_xl
)
}
\arguments{
\item{spectra_extracted}{List of extracted spectra data frames.}

\item{mode}{Character "pos" or "neg".}

\item{metad}{Character library metadata name.}

\item{col_*}{Character column name mappings (see parent function params).}
}
\value{
Data frame with harmonized spectra.
}
\description{
Internal helper to harmonize extracted spectra for a specific
polarity mode, applying column mappings and fixing precursor/inchikey issues.
}
\keyword{internal}
