% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectra.R
\name{import_spectra}
\alias{import_spectra}
\title{Import spectra}
\usage{
import_spectra(file)
}
\arguments{
\item{file}{File path of the spectrum file to be imported}
}
\value{
Spectra object containing the imported spectra
}
\description{
This function imports spectra from a file (.mgf or .sqlite)
}
\examples{
NULL
}
