% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectra.R
\name{import_spectra}
\alias{import_spectra}
\title{Import spectra}
\usage{
import_spectra(file)
}
\arguments{
\item{file}{the file path of the spectrum file to be imported}
}
\value{
a Spectra object containing the imported spectra
}
\description{
Import spectra
}
\examples{
TODO
}
