% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectra.R
\name{import_spectra}
\alias{import_spectra}
\title{Import spectra}
\usage{
import_spectra(file, cutoff = 0, dalton = 0.01, polarity = NA, ppm = 10)
}
\arguments{
\item{file}{File path of the spectrum file to be imported}

\item{cutoff}{Absolute minimal intensity}

\item{dalton}{Dalton tolerance}

\item{polarity}{Polarity}

\item{ppm}{PPM tolerance}
}
\value{
Spectra object containing the imported spectra
}
\description{
This function imports spectra from a file (.mgf or .sqlite)
}
\examples{
NULL
}
