% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectra.R
\name{import_spectra}
\alias{import_spectra}
\title{Import spectra}
\usage{
import_spectra(
  file,
  cutoff = 0,
  dalton = 0.01,
  polarity = NA,
  ppm = 10,
  sanitize = TRUE
)
}
\arguments{
\item{file}{File path of the spectrum file to be imported}

\item{cutoff}{Absolute minimal intensity}

\item{dalton}{Dalton tolerance}

\item{polarity}{Polarity}

\item{ppm}{PPM tolerance}

\item{sanitize}{Flag indicating whether to sanitize. Default TRUE}
}
\value{
Spectra object containing the imported spectra
}
\description{
This function imports spectra from a file (.mgf or .sqlite)
}
\examples{
get_file(
  url = get_default_paths()$urls$examples$spectra_mini,
  export = get_default_paths()$data$source$spectra
)
import_spectra(file = get_default_paths()$data$source$spectra)
import_spectra(
  file = get_default_paths()$data$source$spectra,
  sanitize = FALSE
)
}
