% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{INCHI_NO_STEREO_PATTERN}
\alias{INCHI_NO_STEREO_PATTERN}
\title{InChI stereochemistry layer pattern indicating no stereochemistry
Structures with "-UHFFFAOYSA-" in their InChIKey have no defined stereochemistry}
\format{
An object of class \code{character} of length 1.
}
\usage{
INCHI_NO_STEREO_PATTERN
}
\description{
InChI stereochemistry layer pattern indicating no stereochemistry
Structures with "-UHFFFAOYSA-" in their InChIKey have no defined stereochemistry
}
\keyword{internal}
