% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install}
\usage{
install(
  git_repo = "taxonomicallyinformedannotation/tima",
  r_universe_repo = "taxonomicallyinformedannotation",
  test = FALSE
)
}
\arguments{
\item{git_repo}{Git repo}

\item{r_universe_repo}{r-universe repo}

\item{test}{Flag for tests}
}
\description{
This function runs some required install
}
\examples{
NULL
}
