% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install}
\usage{
install(
  package = "tima",
  repos = c("https://taxonomicallyinformedannotation.r-universe.dev",
    "https://bioc.r-universe.dev", "https://cloud.r-project.org"),
  dependencies = TRUE,
  test = FALSE
)
}
\arguments{
\item{package}{Package name}

\item{repos}{Repositories for install.packages}

\item{dependencies}{Whether to install dependencies}

\item{test}{Whether to retry with source or test install}
}
\description{
This function runs some required install
}
\examples{
NULL
}
