% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install}
\usage{
install(
  package = "tima",
  repos = c("https://taxonomicallyinformedannotation.r-universe.dev",
    "https://bioc.r-universe.dev", "https://cloud.r-project.org"),
  test = FALSE
)
}
\arguments{
\item{package}{Package}

\item{repos}{Repos}

\item{test}{Flag for tests}
}
\description{
This function runs some required install
}
\examples{
NULL
}
