% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tima}
\alias{install_tima}
\title{Install TIMA Package and Dependencies}
\usage{
install_tima(
  package = "tima",
  repos = c("https://taxonomicallyinformedannotation.r-universe.dev",
    "https://cran.r-universe.dev",
    "https://bioc.r-universe.dev"),
  type = getOption("pkgType"),
  dependencies = TRUE
)
}
\arguments{
\item{package}{\link{character} Name of the package (default: "tima")}

\item{repos}{\link{character} Vector of repository URLs}

\item{type}{\link{character}  Type of install ("source", "binary")}

\item{dependencies}{\link{logical} Whether to install dependencies (default: TRUE)}
}
\value{
NULL (invisibly). Installs packages and sets up Python environment as
side effects.
}
\description{
Installs or updates the TIMA package from r-universe and sets up
a Python virtual environment with dependencies.
}
\examples{
\dontrun{
install_tima()
}
}
\seealso{
Other workflow:
\code{\link[=change_params_small]{change_params_small()}},
\code{\link[=create_components]{create_components()}},
\code{\link[=create_edges]{create_edges()}},
\code{\link[=create_edges_spectra]{create_edges_spectra()}},
\code{\link[=go_to_cache]{go_to_cache()}},
\code{\link[=install]{install()}},
\code{\link[=run_app]{run_app()}},
\code{\link[=run_tima]{run_tima()}},
\code{\link[=tima_full]{tima_full()}},
\code{\link[=validate_inputs]{validate_inputs()}}
}
\concept{workflow}
