% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wrappers.R
\name{join_gnps_wrapper}
\alias{join_gnps_wrapper}
\title{Wrapper for the C function "join_gnps"}
\usage{
join_gnps_wrapper(
  query_masses,
  target_masses,
  query_precursor,
  target_precursor,
  dalton,
  ppm
)
}
\arguments{
\item{query_masses}{Numeric vector or matrix for query masses.}

\item{target_masses}{Numeric vector or matrix for target masses.}

\item{query_precursor}{Numeric vector of precursor values for queries.}

\item{target_precursor}{Numeric vector of precursor values for targets.}

\item{dalton}{Numeric value specifying the tolerance in daltons.}

\item{ppm}{Numeric value specifying the tolerance in ppm.}
}
\value{
The result returned from the C function "join_gnps".
}
\description{
Wrapper for the C function "join_gnps"
}
