% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_annotations.R
\name{load_structure_organism_pairs}
\alias{load_structure_organism_pairs}
\title{Load Structure-Organism Pairs}
\usage{
load_structure_organism_pairs(library, str_stereo, org_tax_ott)
}
\arguments{
\item{library}{Character, path to library file}

\item{str_stereo}{Character, path to stereo file (optional)}

\item{org_tax_ott}{Character, path to taxonomy file (optional)}
}
\value{
Data frame with structure-organism pairs
}
\description{
Internal helper to load and join library tables.
}
\keyword{internal}
