% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_files.R
\name{load_yaml_files}
\alias{load_yaml_files}
\title{Load yaml files}
\usage{
load_yaml_files(paths = parse_yaml_paths())
}
\arguments{
\item{paths}{Paths}
}
\description{
This function load yaml files
}
\examples{
NULL
}
