% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_masses_hypothesis.R
\name{match_candidates_to_library}
\alias{match_candidates_to_library}
\title{Match candidate (feature, adduct, mass) rows against the library}
\usage{
match_candidates_to_library(candidates, library_em, tolerance_ppm,
  tolerance_dalton = NULL)
}
\description{
Match candidate (feature, adduct, mass) rows against the library
}
\keyword{internal}
