% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_MASS_DALTONS}
\alias{MAX_MASS_DALTONS}
\title{Maximum practical mass value (Da) for small molecules
Used for sanity checking, not hard limit}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_MASS_DALTONS
}
\description{
Maximum practical mass value (Da) for small molecules
Used for sanity checking, not hard limit
}
\keyword{datasets}
