% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_PATH_LENGTH}
\alias{MAX_PATH_LENGTH}
\title{Maximum path length (OS-dependent, conservative estimate)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MAX_PATH_LENGTH
}
\description{
Maximum path length (OS-dependent, conservative estimate)
}
\keyword{internal}
