% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_TOLERANCE_RT_ADDUCTS}
\alias{MAX_TOLERANCE_RT_ADDUCTS}
\title{Maximum allowed retention time tolerance for adduct grouping (minutes)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_TOLERANCE_RT_ADDUCTS
}
\description{
Maximum allowed retention time tolerance for adduct grouping (minutes)
}
\keyword{datasets}
