% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_MASS_DALTONS}
\alias{MIN_MASS_DALTONS}
\title{Minimum mass value (Da) - used for validation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_MASS_DALTONS
}
\description{
Minimum mass value (Da) - used for validation
}
\keyword{internal}
