% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_results.R
\name{minimize_results}
\alias{minimize_results}
\title{Minimize results}
\usage{
minimize_results(
  df,
  features_table,
  min_score_classes = 0.6,
  min_score_compounds = 0.4,
  best_percentile = 0.9
)
}
\arguments{
\item{df}{Dataframe}

\item{features_table}{Prepared features file}

\item{min_score_classes}{Minimal score to keep classes}

\item{min_score_compounds}{Minimal score to keep compounds}

\item{best_percentile}{Best percentile}
}
\value{
A minimized table
}
\description{
This function minimizes results
}
\examples{
NULL
}
