% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MODIFICATION_REGEX_PATTERN}
\alias{MODIFICATION_REGEX_PATTERN}
\title{Regular expression for individual adduct modifications
Matches: +H, -H2O, +Na, etc.
Groups: (1) sign, (2) count, (3) element/formula}
\format{
An object of class \code{character} of length 1.
}
\usage{
MODIFICATION_REGEX_PATTERN
}
\description{
Regular expression for individual adduct modifications
Matches: +H, -H2O, +Na, etc.
Groups: (1) sign, (2) count, (3) element/formula
}
\keyword{datasets}
