% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_annotation.R
\name{ms1_annotation}
\alias{ms1_annotation}
\title{Title}
\usage{
ms1_annotation(
  annotationTable = metadata_table_spectral_annotation,
  structureExactMassTable = structure_exact_mass_table,
  structureOrganismPairsTable = structure_organism_pairs_table,
  adducts = unlist(params$ms$adducts[[params$ms$mode]]),
  neutralLosses = neutral_losses_table
)
}
\arguments{
\item{neutralLosses}{}
}
\value{

}
\description{
Title
}
