% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_adduct.R
\name{parse_adduct}
\alias{parse_adduct}
\title{Parse adduct}
\usage{
parse_adduct(
  adduct_string,
  regex = "\\\\[(\\\\d*)M(?![a-z])(\\\\d*)([+-][\\\\w\\\\d].*)?.*\\\\](\\\\d*)([+-])?"
)
}
\arguments{
\item{adduct_string}{Adduct to be parsed}

\item{regex}{Regex used for parsing}
}
\value{
Parsed elements from adduct
}
\description{
This function parses adducts
}
\examples{
\dontrun{
parse_adduct("[M+H]+")
parse_adduct("[2M1-C6H12O6 (hexose)+NaCl+H]2+")
}
}
