% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_params.R
\name{parse_yaml_params}
\alias{parse_yaml_params}
\title{Parse YAML parameters}
\usage{
parse_yaml_params(def = default_path, par = params_path)
}
\arguments{
\item{def}{Default path}

\item{par}{Param path}
}
\value{
A list containing the parameters specified in the YAML files
}
\description{
This function parses YAML parameters
}
\examples{
NULL
}
