% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_params.R
\name{parse_yaml_params}
\alias{parse_yaml_params}
\title{Parse YAML parameter files}
\usage{
parse_yaml_params(def, usr = NULL)
}
\arguments{
\item{def}{Path to the default YAML parameters file}

\item{usr}{Path to the user-specified YAML parameters file (optional).
If it exists, it will override default values.}
}
\value{
List containing the parameters specified in the YAML files
}
\description{
Parses YAML parameter files, loading default parameters and
optionally overriding them with user-specified values.
}
\examples{
\dontrun{
# Load parameters
params <- parse_yaml_params(
  def = "params/default_params.yaml",
  usr = "params/user_params.yaml"
)
}
}
