% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_paths.R
\name{parse_yaml_paths}
\alias{parse_yaml_paths}
\title{Parse YAML paths}
\usage{
parse_yaml_paths(file = "tools/paths.yaml")
}
\arguments{
\item{file}{The file name of the YAML file containing the paths (default is "tools/paths.yaml")}
}
\value{
A list containing the paths specified in the YAML file
}
\description{
This function parses YAML paths
}
\examples{
NULL
}
