% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_harmonize_names_sirius.R
\name{pre_harmonize_names_sirius}
\alias{pre_harmonize_names_sirius}
\title{Pre harmonize names sirius}
\usage{
pre_harmonize_names_sirius(x)
}
\arguments{
\item{x}{Character string containing a name}
}
\value{
Character string with the name modified according
to the rules specified in the function
}
\description{
This function pre harmonizes Sirius names
to make them compatible
}
\examples{
prepared_name <- pre_harmonize_names_sirius("My name/suffix")
}
