% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_adducts.R
\name{prepare_adducts}
\alias{prepare_adducts}
\title{Prepare adducts}
\usage{
prepare_adducts(
  adducts_input = params$files$libraries$sop$merged,
  adducts_table_input = paths$data$source$adducts,
  config_output_path = paths$data$interim$config$path,
  adducts_output_path = paths$data$interim$adducts$path,
  output_name = params$files$libraries$adducts$processed,
  masses_pos_output_path = paths$data$interim$adducts$pos,
  masses_neg_output_path = paths$data$interim$adducts$neg
)
}
\arguments{
\item{adducts_input}{List of adducts taken as input}

\item{adducts_table_input}{Table of adducts taken as input}

\item{config_output_path}{Path where the configuration will be saved}

\item{adducts_output_path}{Path where the adducts will be saved}

\item{output_name}{Name of the file where adducts will be saved}

\item{masses_pos_output_path}{Path where positive adducts masses will be saved}

\item{masses_neg_output_path}{Path where negative adducts masses will be saved}
}
\description{
This function prepares adducts for further use
}
\examples{
NULL
}
