% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations_gnps.R
\name{prepare_annotations_gnps}
\alias{prepare_annotations_gnps}
\title{Prepare annotations GNPS}
\usage{
prepare_annotations_gnps(
  input = get_params(step =
    "prepare_annotations_gnps")$files$annotations$raw$spectral$gnps,
  output = get_params(step =
    "prepare_annotations_gnps")$files$annotations$prepared$structural$gnps,
  str_stereo = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "prepare_annotations_gnps")$files$libraries$sop$merged$structures$taxonomies$npc
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\value{
The path to the prepared GNPS annotations
}
\description{
This function prepares GNPS obtained annotations
}
\examples{
go_to_cache(dir = system.file(package = "tima"))
prepare_annotations_gnps()
unlink("data", recursive = TRUE)
}
