% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations_gnps.R
\name{prepare_annotations_gnps}
\alias{prepare_annotations_gnps}
\title{Prepare annotations GNPS}
\usage{
prepare_annotations_gnps(
  input = params$files$annotations$raw$spectral,
  output = params$files$annotations$prepared$structural,
  str_stereo = params$files$libraries$sop$merged$structures$stereo,
  str_met = params$files$libraries$sop$merged$structures$metadata,
  str_nam = params$files$libraries$sop$merged$structures$names,
  str_tax_cla = params$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = params$files$libraries$sop$merged$structures$taxonomies$npc,
  parameters = params
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}

\item{parameters}{Parameters}
}
\description{
This function prepares GNPS obtained annotations
}
\examples{
NULL
}
