% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations_sirius.R
\name{prepare_annotations_sirius}
\alias{prepare_annotations_sirius}
\title{Prepare annotations SIRIUS}
\usage{
prepare_annotations_sirius(
  input_directory = params$files$annotations$raw$sirius,
  output = params$files$annotations$prepared,
  str_2d_3d = params$files$libraries$sop$merged$structures$dd_ddd,
  str_met = params$files$libraries$sop$merged$structures$metadata,
  str_nam = params$files$libraries$sop$merged$structures$names,
  str_tax_cla = params$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = params$files$libraries$sop$merged$structures$taxonomies$npc,
  parameters = params
)
}
\arguments{
\item{input_directory}{Directory containing the Sirius results}

\item{output}{Output where to save prepared results}

\item{str_2d_3d}{File containing 2D and 3D structures}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}

\item{parameters}{Params}
}
\description{
This function prepares Sirius results to make them compatible
}
\examples{
NULL
}
