% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations_sirius.R
\name{prepare_annotations_sirius}
\alias{prepare_annotations_sirius}
\title{Prepare annotations SIRIUS}
\usage{
prepare_annotations_sirius(
  input_directory = get_params(step =
    "prepare_annotations_sirius")$files$annotations$raw$sirius,
  output_ann = get_params(step =
    "prepare_annotations_sirius")$files$annotations$prepared$structural$sirius,
  output_can = get_params(step =
    "prepare_annotations_sirius")$files$annotations$prepared$canopus,
  output_for = get_params(step =
    "prepare_annotations_sirius")$files$annotations$prepared$formula,
  str_stereo = get_params(step =
    "prepare_annotations_sirius")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "prepare_annotations_sirius")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "prepare_annotations_sirius")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "prepare_annotations_sirius")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "prepare_annotations_sirius")$files$libraries$sop$merged$structures$taxonomies$npc
)
}
\arguments{
\item{input_directory}{Directory containing the Sirius results}

\item{output_ann}{Output where to save prepared annotation results}

\item{output_can}{Output where to save prepared canopus results}

\item{output_for}{Output where to save prepared formula results}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\description{
This function prepares Sirius results to make them compatible
}
\examples{
NULL
}
