% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_closed.R
\name{prepare_closed}
\alias{prepare_closed}
\title{Prepare closed}
\usage{
prepare_closed(input = params$input, output = params$output)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}
}
\value{
TODO
}
\description{
Prepare closed
}
\examples{
TODO
}
