% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_edges.R
\name{prepare_edges}
\alias{prepare_edges}
\title{Prepare edges}
\usage{
prepare_edges(
  tool = params$tools$networks$spectral$edges,
  gnps_job_id = params$gnps$id,
  input = params$files$networks$spectral$edges$raw,
  output = params$files$networks$spectral$edges$processed,
  name_source = params$names$source,
  name_target = params$names$target
)
}
\arguments{
\item{tool}{Tool used to generate the edges}

\item{gnps_job_id}{GNPS job ID, if the tool was 'gnps'}

\item{input}{Input file if 'manual'}

\item{output}{Output file}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}
}
\description{
This function prepares edges for further use
}
\examples{
NULL
}
