% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_edges.R
\name{prepare_edges}
\alias{prepare_edges}
\title{Prepare edges}
\usage{
prepare_edges(
  tool = params$tool,
  gnps_job_id = params$gnps,
  input = params$input,
  output = params$output,
  name_source = params$source_name,
  name_target = params$target_name
)
}
\arguments{
\item{tool}{Tool used to generate the edges}

\item{gnps_job_id}{GNPS job ID, if the tool was 'gnps'}

\item{input}{Input file if 'manual'}

\item{output}{Output file}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}
}
\description{
This function prepares edges for further use
}
\examples{
NULL
}
