% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_classification.R
\name{prepare_features_classification}
\alias{prepare_features_classification}
\title{Prepare features classification}
\usage{
prepare_features_classification(
  library = params$library,
  input = params$input,
  output = params$output,
  quickmode = params$quickmode
)
}
\arguments{
\item{library}{TODO}

\item{input}{TODO}

\item{output}{TODO}

\item{quickmode}{TODO}
}
\value{
TODO
}
\description{
Prepare features classification
}
\examples{
TODO
}
