% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_components.R
\name{prepare_features_components}
\alias{prepare_features_components}
\title{Prepare features components}
\usage{
prepare_features_components(
  input = params$files$annotations$pretreated,
  output = params$files$annotations$filled,
  tool = params$tools$networks$spectral$components,
  components = params$files$networks$spectral$components$raw,
  gnps_job_id = params$gnps$id,
  ms_mode = params$ms$polarity,
  parameters = params
)
}
\arguments{
\item{input}{Input file if tool == 'manual'}

\item{output}{Output file}

\item{tool}{Tool used to generate components}

\item{components}{File containing the components if tool == 'manual'}

\item{gnps_job_id}{GNPS job ID}

\item{ms_mode}{Ionization mode. Must be 'pos' or 'neg'}

\item{parameters}{Params}
}
\description{
This function prepares the components (clusters in molecular network) for further use
}
\examples{
NULL
}
