% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_components.R
\name{prepare_features_components}
\alias{prepare_features_components}
\title{Prepare features components}
\usage{
prepare_features_components(
  input = params$files$annotations$pretreated,
  output = params$files$annotations$filled,
  components = params$files$networks$spectral$components$raw,
  parameters = params
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{components}{File containing the components}

\item{parameters}{Params}
}
\description{
This function prepares the components (clusters in molecular network) for further use
}
\examples{
NULL
}
