% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_components.R
\name{prepare_features_components}
\alias{prepare_features_components}
\title{Prepare features components}
\usage{
prepare_features_components(
  input = params$input,
  output = params$output,
  tool = params$tool,
  components = params$components,
  gnps_job_id = params$gnps,
  ms_mode = params$mode
)
}
\arguments{
\item{input}{Input file if tool == 'manual'}

\item{output}{Output file}

\item{tool}{Tool used to generate components}

\item{components}{File containing the components if tool == 'manual'}

\item{gnps_job_id}{GNPS job ID}

\item{ms_mode}{Ionization mode. Must be 'pos' or 'neg'}
}
\description{
TODO
}
\examples{
NULL
}
