% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_components.R
\name{prepare_features_components}
\alias{prepare_features_components}
\title{Prepare features components}
\usage{
prepare_features_components(
  input = get_params(step =
    "prepare_features_components")$files$networks$spectral$components$raw,
  output = get_params(step =
    "prepare_features_components")$files$networks$spectral$components$prepared
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\description{
This function prepares the components
(clusters in molecular network) for further use
}
\examples{
NULL
}
