% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_components.R
\name{prepare_features_components}
\alias{prepare_features_components}
\title{Title}
\usage{
prepare_features_components(
  input = params$input,
  output = params$output,
  tool = params$tool,
  components = params$components,
  gnps_job_id = params$gnps,
  ms_mode = params$mode
)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}

\item{tool}{TODO}

\item{components}{TODO}

\item{gnps_job_id}{TODO}

\item{ms_mode}{TODO}
}
\value{
TODO
}
\description{
Title
}
