% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_edges.R
\name{prepare_features_edges}
\alias{prepare_features_edges}
\title{Prepare edges}
\usage{
prepare_features_edges(
  input = params$files$networks$spectral$edges$raw,
  output = params$files$networks$spectral$edges$processed,
  name_source = params$names$source,
  name_target = params$names$target,
  parameters = params
)
}
\arguments{
\item{input}{Input file if 'manual'}

\item{output}{Output file}

\item{name_source}{Name of the source features column}

\item{name_target}{Name of the target features column}

\item{parameters}{Params}
}
\description{
This function prepares edges for further use
}
\examples{
NULL
}
