% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_tables.R
\name{prepare_features_tables}
\alias{prepare_features_tables}
\title{Prepare features table}
\usage{
prepare_features_tables(
  features = params$files$features$raw,
  output = params$files$features$prepared,
  name_features = params$names$features,
  name_rt = params$names$rt,
  name_mz = params$names$precursor,
  parameters = params
)
}
\arguments{
\item{features}{Path to the file containing the features data}

\item{output}{Path to the file to export the merged data to}

\item{name_features}{Name of the features column in the features data}

\item{name_rt}{Name of the retention time column in the features data}

\item{name_mz}{Name of the m/z column in the features data}

\item{parameters}{params}
}
\description{
This function prepares features
}
\examples{
NULL
}
