% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features_tables.R
\name{prepare_features_tables}
\alias{prepare_features_tables}
\title{Prepare features table}
\usage{
prepare_features_tables(
  features = get_params(step = "prepare_features_tables")$files$features$raw,
  output = get_params(step = "prepare_features_tables")$files$features$prepared,
  candidates = get_params(step = "prepare_features_tables")$annotations$canidates$samples,
  name_adduct = get_params(step = "prepare_features_tables")$names$adduct,
  name_features = get_params(step = "prepare_features_tables")$names$features,
  name_rt = get_params(step = "prepare_features_tables")$names$rt$features,
  name_mz = get_params(step = "prepare_features_tables")$names$precursor
)
}
\arguments{
\item{features}{Path to the file containing the features data}

\item{output}{Path to the file to export the merged data to}

\item{candidates}{Number of samples to be retained per feature top intensities}

\item{name_adduct}{Name of the adduct column in the features data}

\item{name_features}{Name of the features column in the features data}

\item{name_rt}{Name of the retention time column in the features data}

\item{name_mz}{Name of the m/z column in the features data}
}
\value{
The path to the prepared feature table
}
\description{
This function prepares features
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
get_file(
  url = get_default_paths()$urls$examples$features,
  export = get_params(step = "prepare_features_tables")$files$features$raw
)
prepare_features_tables()
unlink("data", recursive = TRUE)
}
}
