% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gnps.R
\name{prepare_gnps}
\alias{prepare_gnps}
\title{Title}
\usage{
prepare_gnps(
  gnps_job_id = params$gnps,
  nap_job_id = params$nap,
  output = params$output
)
}
\arguments{
\item{gnps_job_id}{TODO}

\item{nap_job_id}{TODO}

\item{output}{TODO}
}
\value{
TODO
}
\description{
Title
}
