% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gnps.R
\name{prepare_gnps}
\alias{prepare_gnps}
\title{Prepare GNPS}
\usage{
prepare_gnps(
  gnps_job_id = params$gnps$id,
  output = params$files$annotations$pretreated,
  str_2D_3D = paths$data$interim$libraries$merged$structures$dd_ddd,
  str_met = paths$data$interim$libraries$merged$structures$metadata,
  str_nam = paths$data$interim$libraries$merged$structures$names,
  str_tax_cla = paths$data$interim$libraries$merged$structures$taxonomies$classyfire,
  str_tax_npc = paths$data$interim$libraries$merged$structures$taxonomies$npc,
  parameters = params
)
}
\arguments{
\item{gnps_job_id}{GNPS job ID}

\item{output}{Output file}

\item{str_2D_3D}{File containing 2D and 3D structures}

\item{str_met}{File containing structures metadata}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}

\item{parameters}{Parameters}
}
\description{
This function prepares GNPS obtained annotations for further use
}
\examples{
NULL
}
