% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isdb_lotus.R
\name{prepare_isdb_lotus}
\alias{prepare_isdb_lotus}
\title{Prepare LOTUS In Silico DataBase}
\usage{
prepare_isdb_lotus(
  input_pos = paths$data$source$spectra$lotus$pos,
  input_neg = paths$data$source$spectra$lotus$neg,
  output_pos = paths$data$interim$spectra$lotus$pos,
  output_neg = paths$data$interim$spectra$lotus$neg,
  export_sqlite = FALSE
)
}
\arguments{
\item{input_pos}{TODO}

\item{input_neg}{TODO}

\item{output_pos}{TODO}

\item{output_neg}{TODO}

\item{export_sqlite}{TODO}
}
\value{
TODO
}
\description{
Prepare LOTUS In Silico DataBase
}
\examples{
TODO
}
