% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isdb_lotus.R
\name{prepare_isdb_lotus}
\alias{prepare_isdb_lotus}
\title{Prepare LOTUS In Silico DataBase}
\usage{
prepare_isdb_lotus(
  input = paths$data$source$spectra$lotus$pos,
  output = paths$data$interim$spectra$lotus$pos,
  polarity = "pos",
  export_sqlite = TRUE
)
}
\arguments{
\item{input}{File containing LOTUS ISDB}

\item{output}{Output file}

\item{polarity}{MS polarity}

\item{export_sqlite}{Boolean. Export to SQLite? TRUE or FALSE}
}
\description{
This function prepares the LOTUS in silico generated spectra to be used for spectral matching
}
\examples{
NULL
}
