% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_rt.R
\name{prepare_libraries_rt}
\alias{prepare_libraries_rt}
\title{Prepare libraries of retention times}
\usage{
prepare_libraries_rt(
  mgf_exp = params$files$libraries$spectral$exp,
  mgf_is = params$files$libraries$spectral$is,
  temp_exp = params$files$libraries$temporal$exp,
  temp_is = params$files$libraries$temporal$is,
  output = params$files$libraries$temporal$prepared,
  library = paths$data$interim$libraries$sop$merged$keys,
  col_ik = params$names$mgf$inchikey,
  col_rt = params$names$mgf$retention_time,
  col_sm = params$names$mgf$smiles,
  name_inchikey = params$names$inchikey,
  name_rt = params$names$rt,
  name_smiles = params$names$smiles,
  unit_rt = params$units$rt,
  parameters = params
)
}
\arguments{
\item{mgf_exp}{MGF containing experimental retention times}

\item{mgf_is}{MGF containing in silico predicted retention times}

\item{temp_exp}{File containing experimental retention times}

\item{temp_is}{File containing in silico predicted retention times}

\item{output}{Output file}

\item{library}{Library containing the keys}

\item{col_ik}{Name of the InChIKey in mgf}

\item{col_rt}{Name of the retention time in mgf}

\item{col_sm}{Name of the SMILES in mgf}

\item{name_inchikey}{Name of the InChIKey in file}

\item{name_rt}{Name of the retention time in file}

\item{name_smiles}{Name of the SMILES in file}

\item{unit_rt}{Unit of the retention time. Must be "seconds" or "minutes"}

\item{parameters}{Params}
}
\description{
This function prepares retention times libraries
to be used for later
}
\examples{
NULL
}
