% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_rt.R
\name{prepare_libraries_rt}
\alias{prepare_libraries_rt}
\title{Prepare libraries of retention times}
\usage{
prepare_libraries_rt(
  mgf_exp = get_params(step = "prepare_libraries_rt")$files$libraries$temporal$exp$mgf,
  mgf_is = get_params(step = "prepare_libraries_rt")$files$libraries$temporal$is$mgf,
  temp_exp = get_params(step = "prepare_libraries_rt")$files$libraries$temporal$exp$csv,
  temp_is = get_params(step = "prepare_libraries_rt")$files$libraries$temporal$is$csv,
  output_rt = get_params(step = "prepare_libraries_rt")$files$libraries$temporal$prepared,
  output_sop = get_params(step = "prepare_libraries_rt")$files$libraries$sop$prepared$rt,
  col_ik = get_params(step = "prepare_libraries_rt")$names$mgf$inchikey,
  col_rt = get_params(step = "prepare_libraries_rt")$names$mgf$retention_time,
  col_sm = get_params(step = "prepare_libraries_rt")$names$mgf$smiles,
  name_inchikey = get_params(step = "prepare_libraries_rt")$names$inchikey,
  name_rt = get_params(step = "prepare_libraries_rt")$names$rt$library,
  name_smiles = get_params(step = "prepare_libraries_rt")$names$smiles,
  unit_rt = get_params(step = "prepare_libraries_rt")$units$rt
)
}
\arguments{
\item{mgf_exp}{MGF containing experimental retention times}

\item{mgf_is}{MGF containing in silico predicted retention times}

\item{temp_exp}{File containing experimental retention times}

\item{temp_is}{File containing in silico predicted retention times}

\item{output_rt}{Output retention time file}

\item{output_sop}{Output pseudo sop file}

\item{col_ik}{Name of the InChIKey in mgf}

\item{col_rt}{Name of the retention time in mgf}

\item{col_sm}{Name of the SMILES in mgf}

\item{name_inchikey}{Name of the InChIKey in file}

\item{name_rt}{Name of the retention time in file}

\item{name_smiles}{Name of the SMILES in file}

\item{unit_rt}{Unit of the retention time. Must be "seconds" or "minutes"}
}
\value{
The path to the prepared retention time library
}
\description{
This function prepares retention times libraries
to be used for later
}
\examples{
tima:::copy_backbone()
go_to_cache()
prepare_libraries_rt()
unlink("data", recursive = TRUE)
}
