% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_closed.R
\name{prepare_libraries_sop_closed}
\alias{prepare_libraries_sop_closed}
\title{Prepare libraries of structure organism pairs CLOSED}
\usage{
prepare_libraries_sop_closed(
  input = params$files$libraries$sop$raw$closed,
  output = params$files$libraries$sop$processed,
  parameters = params
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{parameters}{params}
}
\description{
Prepare libraries of structure organism pairs CLOSED
}
\examples{
NULL
}
