% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_closed.R
\name{prepare_libraries_sop_closed}
\alias{prepare_libraries_sop_closed}
\title{Prepare libraries of structure organism pairs CLOSED}
\usage{
prepare_libraries_sop_closed(
  input = get_params(step =
    "prepare_libraries_sop_closed")$files$libraries$sop$raw$closed,
  output = get_params(step =
    "prepare_libraries_sop_closed")$files$libraries$sop$prepared$closed
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\value{
The path to the prepared structure-organism pairs library CLOSED
}
\description{
Prepare libraries of structure organism pairs CLOSED
}
\examples{
\donttest{
tima:::copy_backbone()
go_to_cache()
prepare_libraries_sop_closed()
unlink("data", recursive = TRUE)
}
}
