% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_ecmdb.R
\name{prepare_libraries_sop_ecmdb}
\alias{prepare_libraries_sop_ecmdb}
\title{Prepare libraries of structure organism pairs ECMDB}
\usage{
prepare_libraries_sop_ecmdb(
  input = get_params(step = "prepare_libraries_sop_ecmdb")$files$libraries$sop$raw$ecmdb,
  output = get_params(step =
    "prepare_libraries_sop_ecmdb")$files$libraries$sop$prepared$ecmdb
)
}
\arguments{
\item{input}{Character string path to ECMDB JSON zip file}

\item{output}{Character string path for prepared ECMDB library output}
}
\value{
Character string path to prepared ECMDB structure-organism pairs
}
\description{
This function prepares ECMDB (E. coli Metabolome Database)
structure-organism pairs by parsing JSON data, extracting metabolite
information, and formatting for TIMA workflows. Handles E. coli
metabolite data with structures.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_ecmdb()
unlink("data", recursive = TRUE)
}
}
\seealso{
Other preparation: 
\code{\link{prepare_annotations_gnps}()},
\code{\link{prepare_annotations_mzmine}()},
\code{\link{prepare_annotations_sirius}()},
\code{\link{prepare_annotations_spectra}()},
\code{\link{prepare_features_components}()},
\code{\link{prepare_features_edges}()},
\code{\link{prepare_features_tables}()},
\code{\link{prepare_libraries_rt}()},
\code{\link{prepare_libraries_sop_bigg}()},
\code{\link{prepare_libraries_sop_closed}()},
\code{\link{prepare_libraries_sop_hmdb}()},
\code{\link{prepare_libraries_sop_lotus}()},
\code{\link{prepare_libraries_sop_merged}()},
\code{\link{prepare_libraries_spectra}()},
\code{\link{prepare_params}()},
\code{\link{prepare_taxa}()}
}
\concept{preparation}
