% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_ecmdb.R
\name{prepare_libraries_sop_ecmdb}
\alias{prepare_libraries_sop_ecmdb}
\title{Prepare libraries of structure organism pairs ECMDB}
\usage{
prepare_libraries_sop_ecmdb(
  input = params$files$libraries$sop$raw$ecmdb,
  output = params$files$libraries$sop$prepared,
  parameters = params
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}

\item{parameters}{params}
}
\description{
Prepare libraries of structure organism pairs ECMDB
}
\examples{
NULL
}
