% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_hmdb.R
\name{prepare_libraries_sop_hmdb}
\alias{prepare_libraries_sop_hmdb}
\title{Prepare libraries of structure organism pairs HMDB}
\usage{
prepare_libraries_sop_hmdb(
  input = get_params(step = "prepare_libraries_sop_hmdb")$files$libraries$sop$raw$hmdb,
  output = get_params(step =
    "prepare_libraries_sop_hmdb")$files$libraries$sop$prepared$hmdb
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\value{
The path to the prepared structure-organism pairs library HMDB
}
\description{
This function prepares the HMDB structure-organism pairs
}
\examples{
go_to_cache(dir = system.file(package = "tima"))
prepare_libraries_sop_hmdb()
unlink("data", recursive = TRUE)
}
