% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_hmdb.R
\name{prepare_libraries_sop_hmdb}
\alias{prepare_libraries_sop_hmdb}
\title{Prepare libraries of structure organism pairs HMDB}
\usage{
prepare_libraries_sop_hmdb(
  input = get_params(step = "prepare_libraries_sop_hmdb")$files$libraries$sop$raw$hmdb,
  output = get_params(step =
    "prepare_libraries_sop_hmdb")$files$libraries$sop$prepared$hmdb
)
}
\arguments{
\item{input}{Character string path to HMDB SDF zip file}

\item{output}{Character string path for prepared HMDB library output}
}
\value{
Character string path to prepared HMDB structure-organism pairs
}
\description{
This function prepares HMDB (Human Metabolome Database)
structure-organism pairs by parsing SDF files, extracting metadata,
and formatting for TIMA annotation workflows.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_hmdb()
unlink("data", recursive = TRUE)
}
}
