% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_lotus.R
\name{prepare_libraries_sop_lotus}
\alias{prepare_libraries_sop_lotus}
\title{Prepare libraries of structure organism pairs LOTUS}
\usage{
prepare_libraries_sop_lotus(
  input = get_params(step = "prepare_libraries_sop_lotus")$files$libraries$sop$raw$lotus,
  output = get_params(step =
    "prepare_libraries_sop_lotus")$files$libraries$sop$prepared$lotus
)
}
\arguments{
\item{input}{Character string path to the raw LOTUS data file}

\item{output}{Character string path for the prepared output file}
}
\value{
Character string path to the prepared structure-organism pairs
library file
}
\description{
This function prepares the LOTUS (LOng-lasting, cUraTed collection
of cOnnectivity daTa for natural products) structure-organism pairs database.
It standardizes columns, extracts 2D InChIKeys, rounds numeric values,
and removes duplicates.
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_lotus()
unlink("data", recursive = TRUE)
}
}
