% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_lotus.R
\name{prepare_libraries_sop_lotus}
\alias{prepare_libraries_sop_lotus}
\title{Prepare libraries of structure organism pairs LOTUS}
\usage{
prepare_libraries_sop_lotus(
  input = get_params(step = "prepare_libraries_sop_lotus")$files$libraries$sop$raw$lotus,
  output = get_params(step =
    "prepare_libraries_sop_lotus")$files$libraries$sop$prepared$lotus
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\value{
The path to the prepared structure-organism pairs library LOTUS
}
\description{
This function prepares the LOTUS structure-organism pairs
}
\examples{
\dontrun{
copy_backbone()
go_to_cache()
prepare_libraries_sop_lotus()
unlink("data", recursive = TRUE)
}
}
