% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_lotus.R
\name{prepare_libraries_sop_lotus}
\alias{prepare_libraries_sop_lotus}
\title{Prepare libraries of structure organism pairs LOTUS}
\usage{
prepare_libraries_sop_lotus(
  input = paths$data$source$libraries$sop$lotus,
  output = paths$data$interim$libraries$sop$lotus
)
}
\arguments{
\item{input}{Input file}

\item{output}{Output file}
}
\description{
This function prepares the LOTUS structure-organism pairs
}
\examples{
NULL
}
