% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_sop_merged.R
\name{prepare_libraries_sop_merged}
\alias{prepare_libraries_sop_merged}
\title{Prepare merged structure organism pairs libraries}
\usage{
prepare_libraries_sop_merged(
  files = params$files$libraries$sop$prepared,
  filter = params$organisms$filter$mode,
  level = params$organisms$filter$level,
  value = params$organisms$filter$value,
  output_key = paths$data$interim$libraries$sop$merged$keys,
  output_org_tax_ott = paths$data$interim$libraries$sop$merged$organisms$taxonomies$ott,
  output_str_2D_3D = paths$data$interim$libraries$sop$merged$structures$dd_ddd,
  output_str_met = paths$data$interim$libraries$sop$merged$structures$metadata,
  output_str_nam = paths$data$interim$libraries$sop$merged$structures$names,
  output_str_tax_cla =
    paths$data$interim$libraries$sop$merged$structures$taxonomies$classyfire,
  output_str_tax_npc = paths$data$interim$libraries$sop$merged$structures$taxonomies$npc,
  parameters = params
)
}
\arguments{
\item{files}{List of libraries to be merged}

\item{filter}{Boolean. TRUE or FALSE if you want to filter the library}

\item{level}{Biological rank to be filtered. Kingdom, phylum, family, genus, ...}

\item{value}{Name of the taxon or taxa to be kept, e.g. 'Gentianaceae|Apocynaceae'}

\item{output_key}{Output file for keys}

\item{output_org_tax_ott}{Output file for organisms taxonomy (OTT)}

\item{output_str_2D_3D}{Output file for structures (2D + 3D)}

\item{output_str_met}{Output file for structures metadata}

\item{output_str_nam}{Output file for structures names}

\item{output_str_tax_cla}{Output file for structures taxonomy (Classyifre)}

\item{output_str_tax_npc}{Output file for structures taxonomy (NPC)}

\item{parameters}{Param}
}
\description{
This function prepares the libraries made of all sub-libraries containing structure-organism pairs
}
\details{
It can be restricted to specific taxa to have more biologically meaningful annotation.
}
\examples{
NULL
}
