% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_libraries_spectra.R
\name{prepare_libraries_spectra}
\alias{prepare_libraries_spectra}
\title{Prepare libraries of spectra}
\usage{
prepare_libraries_spectra(
  input = get_params(step = "prepare_libraries_spectra")$files$libraries$spectral$raw,
  nam_lib = get_params(step = "prepare_libraries_spectra")$names$libraries,
  col_ad = get_params(step = "prepare_libraries_spectra")$names$mgf$adduct,
  col_ce = get_params(step = "prepare_libraries_spectra")$names$mgf$collision_energy,
  col_ci = get_params(step = "prepare_libraries_spectra")$names$mgf$compound_id,
  col_em = get_params(step = "prepare_libraries_spectra")$names$mgf$exact_mass,
  col_in = get_params(step = "prepare_libraries_spectra")$names$mgf$inchi,
  col_io = get_params(step = "prepare_libraries_spectra")$names$mgf$inchi_no_stereo,
  col_ik = get_params(step = "prepare_libraries_spectra")$names$mgf$inchikey,
  col_il = get_params(step = "prepare_libraries_spectra")$names$mgf$inchikey_no_stereo,
  col_mf = get_params(step = "prepare_libraries_spectra")$names$mgf$molecular_formula,
  col_na = get_params(step = "prepare_libraries_spectra")$names$mgf$name,
  col_po = get_params(step = "prepare_libraries_spectra")$names$mgf$polarity,
  col_sm = get_params(step = "prepare_libraries_spectra")$names$mgf$smiles,
  col_sn = get_params(step = "prepare_libraries_spectra")$names$mgf$smiles_no_stereo,
  col_si = get_params(step = "prepare_libraries_spectra")$names$mgf$spectrum_id,
  col_sp = get_params(step = "prepare_libraries_spectra")$names$mgf$splash,
  col_sy = get_params(step = "prepare_libraries_spectra")$names$mgf$synonyms,
  col_xl = get_params(step = "prepare_libraries_spectra")$names$mgf$xlogp
)
}
\arguments{
\item{input}{File containing spectra}

\item{nam_lib}{Metadata to identify the library}

\item{col_ad}{Name of the adduct in mgf}

\item{col_ce}{Name of the collision energy in mgf}

\item{col_ci}{Name of the compound id in mgf}

\item{col_em}{Name of the exact mass in mgf}

\item{col_in}{Name of the InChI in mgf}

\item{col_io}{Name of the InChI without stereo in mgf}

\item{col_ik}{Name of the InChIKey in mgf}

\item{col_il}{Name of the InChIKey without stereo in mgf}

\item{col_mf}{Name of the molecular formula in mgf}

\item{col_na}{Name of the name in mgf}

\item{col_po}{Name of the polarity in mgf}

\item{col_sm}{Name of the SMILES in mgf}

\item{col_sn}{Name of the SMILES without stereo in mgf}

\item{col_si}{Name of the spectrum id in mgf}

\item{col_sp}{Name of the SPLASH in mgf}

\item{col_sy}{Name of the synonyms in mgf}

\item{col_xl}{Name of the xlogp in mgf}

\item{polarity}{MS polarity}
}
\value{
The path to the prepared spectral library
}
\description{
This function prepares spectra
to be used for spectral matching
}
\examples{
go_to_cache(dir = system.file(package = "tima"))
prepare_libraries_spectra()
unlink("data", recursive = TRUE)
}
