% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_library.R
\name{prepare_library}
\alias{prepare_library}
\title{Prepare library}
\usage{
prepare_library(
  filter = params$filter$mode,
  level = params$filter$level,
  value = params$filter$value,
  output = params$output
)
}
\arguments{
\item{filter}{Boolean. TRUE or FALSE if you want to filter the library}

\item{level}{Biological rank to be filtered. Kingdom, phylum, family, genus, ...}

\item{value}{Name of the taxon or taxa to be kept, e.g. 'Gentianaceae|Apocynaceae'}

\item{output}{Output file}
}
\description{
TODO
}
\examples{
NULL
}
