% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_library.R
\name{prepare_library}
\alias{prepare_library}
\title{Title}
\usage{
prepare_library(
  filter = params$filter$mode,
  level = params$filter$level,
  value = params$filter$value,
  output = params$output
)
}
\arguments{
\item{filter}{TODO}

\item{level}{TODO}

\item{value}{TODO}

\item{output}{TODO}
}
\value{
TODO
}
\description{
Title
}
