% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_params.R
\name{prepare_params}
\alias{prepare_params}
\title{Title}
\usage{
prepare_params(
  filename = params$filename,
  gnps_job_id = params$gnps,
  candidates_initial = params$top_k$initial,
  candidates_final = params$top_k$final,
  weight_biological = params$weights$biological,
  weight_chemical = params$weights$chemical,
  weight_spectral = params$weights$spectral,
  ms_mode = params$ms$mode,
  annotate = params$ms$annotate,
  tolerance_ppm = params$ms$tolerance$ppm,
  tolerance_rt = params$ms$tolerance$rt,
  taxon = params$taxon,
  force = params$force
)
}
\arguments{
\item{filename}{TODO}

\item{gnps_job_id}{TODO}

\item{candidates_initial}{TODO}

\item{candidates_final}{TODO}

\item{weight_biological}{TODO}

\item{weight_chemical}{TODO}

\item{weight_spectral}{TODO}

\item{ms_mode}{TODO}

\item{annotate}{TODO}

\item{tolerance_ppm}{TODO}

\item{tolerance_rt}{TODO}

\item{force}{TODO}
}
\value{
TODO
}
\description{
Title
}
