% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_params.R
\name{prepare_params}
\alias{prepare_params}
\title{Prepare params}
\usage{
prepare_params(
  params_small = get_params(step = "prepare_params"),
  params_advanced = get_params(step = "prepare_params_advanced"),
  step = NA
)
}
\arguments{
\item{params_small}{params_small}

\item{params_advanced}{params_advanced}

\item{step}{Step}
}
\value{
The path to the yaml files containing prepared parameters
}
\description{
This function prepares main parameters
}
\examples{
prepare_params()
}
