% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_params.R
\name{prepare_params}
\alias{prepare_params}
\title{Prepare params}
\usage{
prepare_params(
  filename = params$files$pattern,
  gnps_job_id = params$gnps$id,
  ms_mode = params$ms$polarity,
  taxon = params$organisms$taxon,
  summarise = params$options$summarise,
  parameters = params,
  step = NA
)
}
\arguments{
\item{filename}{Name of the file}

\item{gnps_job_id}{GNPS job ID}

\item{ms_mode}{MS ionization mode. 'pos' or 'neg'}

\item{taxon}{Name of a taxon you want to enforce}

\item{summarise}{Summarise results to one row per feature. BOOLEAN}

\item{parameters}{Params}

\item{step}{Step}
}
\description{
This function prepares main parameters
}
\examples{
NULL
}
