% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_params.R
\name{prepare_params}
\alias{prepare_params}
\title{Prepare params}
\usage{
prepare_params(
  filename = get_params(step = "prepare_params")$files$pattern,
  features = get_params(step = "prepare_params")$files$features$raw,
  spectra = get_params(step = "prepare_params")$files$spectral$raw,
  gnps_job_id = get_params(step = "prepare_params")$gnps$id,
  gnps_example_id = parse_yaml_paths()$gnps$example,
  ms_mode = get_params(step = "prepare_params")$ms$polarity,
  taxon = get_params(step = "prepare_params")$organisms$taxon,
  summarise = get_params(step = "prepare_params")$options$summarise,
  step = NA
)
}
\arguments{
\item{filename}{Name of the file}

\item{features}{File containing the features table}

\item{spectra}{File containing the spectra}

\item{gnps_job_id}{GNPS job ID}

\item{gnps_example_id}{GNPS example job ID}

\item{ms_mode}{MS ionization mode. 'pos' or 'neg'}

\item{taxon}{Name of a taxon you want to enforce}

\item{summarise}{Summarise results to one row per feature. BOOLEAN}

\item{step}{Step}
}
\description{
This function prepares main parameters
}
\examples{
NULL
}
