% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sirius.R
\name{prepare_sirius}
\alias{prepare_sirius}
\title{Title}
\usage{
prepare_sirius(input_directory = params$directory, output = params$output)
}
\arguments{
\item{input_directory}{TODO}

\item{output}{TODO}
}
\value{
TODO
}
\description{
Title
}
