% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sirius.R
\name{prepare_sirius}
\alias{prepare_sirius}
\title{Prepare sirius}
\usage{
prepare_sirius(
  input_directory = params$directory,
  npc = params$npc,
  output = params$output
)
}
\arguments{
\item{input_directory}{Directory containing the Sirius results}

\item{npc}{Boolean. NPClassifier classes computed. TRUE or FALSE}

\item{output}{Output where to save prepared results}
}
\description{
TODO
}
\examples{
NULL
}
