% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_taxa.R
\name{prepare_taxa}
\alias{prepare_taxa}
\title{Title}
\usage{
prepare_taxa(
  input = params$input,
  tool = params$tool,
  extension = params$extension,
  colname = params$column_name,
  gnps_job_id = params$gnps,
  metadata = params$metadata,
  top_k = params$top_k,
  output = params$output,
  taxon = params$taxon
)
}
\arguments{
\item{input}{TODO}

\item{tool}{TODO}

\item{extension}{TODO}

\item{colname}{TODO}

\item{gnps_job_id}{TODO}

\item{metadata}{TODO}

\item{top_k}{TODO}

\item{output}{TODO}

\item{taxon}{TODO}
}
\value{
TODO
}
\description{
Title
}
