% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_taxa.R
\name{prepare_taxa}
\alias{prepare_taxa}
\title{Prepare taxa}
\usage{
prepare_taxa(
  input = params$files$features$raw,
  extension = params$names$extension,
  colname = params$names$taxon,
  metadata = params$files$taxa$raw,
  top_k = params$organisms$candidates,
  org_tax_ott = params$files$libraries$sop$merged$organisms$taxonomies$ott,
  output = params$files$taxa$prepared,
  taxon = params$organisms$taxon,
  parameters = params
)
}
\arguments{
\item{input}{File containing your features intensities}

\item{extension}{Does your colun names contain the file extension? (MZmine mainly)}

\item{colname}{Name of the column containing biological source information}

\item{metadata}{File containing your metadata including biological source}

\item{top_k}{Number of organisms to be retained per feature top intensities}

\item{org_tax_ott}{File containing Open Tree of Life Taxonomy}

\item{output}{Output file}

\item{taxon}{If you want to enforce all features to a given taxon, put its name here.}

\item{parameters}{Params}
}
\description{
This function performs taxon name preparation to match the Open Tree of Life taxonomy
}
\details{
Depending if the features are aligned between samples originating from various organisms or not,
It can either attribute all features to a single organism, or attribute them to multiple ones,
according to their relative intensities among the samples.
}
\examples{
NULL
}
