% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreat_names_sirius.R
\name{pretreat_names_sirius}
\alias{pretreat_names_sirius}
\title{Pretreat names sirius}
\usage{
pretreat_names_sirius(x)
}
\arguments{
\item{x}{Character string containing a name}
}
\value{
Character string with the name modified according to the rules specified in the function
}
\description{
This function pretreats Sirius names to make them compatible
}
\examples{
pretreated_name <- pretreat_names_sirius("My name/suffix")
}
