% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{process_annotations}
\alias{process_annotations}
\title{Title}
\usage{
process_annotations(
  library = params$library,
  name = params$name,
  gnps = params$annotation$gnps,
  isdb = params$annotation$isdb,
  sirius = params$annotation$sirius,
  taxa = params$taxa,
  edges = params$edges,
  output = params$output,
  candidates_initial = params$top_k$initial,
  candidates_final = params$top_k$final,
  weight_spectral = params$weights$spectral,
  weight_chemical = params$weights$chemical,
  weight_biological = params$weights$biological,
  score_chemical_pathway = params$scores$chemical$pathway,
  score_chemical_superclass = params$scores$chemical$superclass,
  score_chemical_class = params$scores$chemical$class,
  score_biological_domain = params$scores$biological$domain,
  score_biological_kingdom = params$scores$biological$kingdom,
  score_biological_phylum = params$scores$biological$phylum,
  score_biological_class = params$scores$biological$class,
  score_biological_order = params$scores$biological$order,
  score_biological_infraorder = params$scores$biological$infraorder,
  score_biological_family = params$scores$biological$family,
  score_biological_subfamily = params$scores$biological$subfamily,
  score_biological_tribe = params$scores$biological$tribe,
  score_biological_subtribe = params$scores$biological$subtribe,
  score_biological_genus = params$scores$biological$genus,
  score_biological_subgenus = params$scores$biological$subgenus,
  score_biological_species = params$scores$biological$species,
  score_biological_subspecies = params$scores$biological$subspecies,
  score_biological_variety = params$scores$biological$variety,
  ms_mode = params$ms$mode,
  ms_level = params$ms$level,
  annotate = params$ms$annotate,
  tolerance_ppm = params$ms$tolerance$ppm,
  tolerance_rt = params$ms$tolerance$rt,
  adducts_list = params$ms$adducts,
  minimal_ms1_bio = params$scores$biological$minimal,
  minimal_ms1_chemo = params$scores$chemical$minimal,
  ms1_only = params$ms$ms1only,
  force = params$force
)
}
\arguments{
\item{library}{TODO}

\item{name}{TODO}

\item{gnps}{TODO}

\item{isdb}{TODO}

\item{sirius}{TODO}

\item{taxa}{TODO}

\item{edges}{TODO}

\item{output}{TODO}

\item{candidates_initial}{TODO}

\item{candidates_final}{TODO}

\item{weight_spectral}{TODO}

\item{weight_chemical}{TODO}

\item{weight_biological}{TODO}

\item{score_chemical_pathway}{TODO}

\item{score_chemical_superclass}{TODO}

\item{score_chemical_class}{TODO}

\item{score_biological_domain}{TODO}

\item{score_biological_kingdom}{TODO}

\item{score_biological_phylum}{TODO}

\item{score_biological_class}{TODO}

\item{score_biological_order}{TODO}

\item{score_biological_infraorder}{TODO}

\item{score_biological_family}{TODO}

\item{score_biological_subfamily}{TODO}

\item{score_biological_tribe}{TODO}

\item{score_biological_subtribe}{TODO}

\item{score_biological_genus}{TODO}

\item{score_biological_subgenus}{TODO}

\item{score_biological_species}{TODO}

\item{score_biological_subspecies}{TODO}

\item{score_biological_variety}{TODO}

\item{ms_mode}{TODO}

\item{ms_level}{TODO}

\item{annotate}{TODO}

\item{tolerance_ppm}{TODO}

\item{tolerance_rt}{TODO}

\item{adducts_list}{TODO}

\item{minimal_ms1_bio}{TODO}

\item{minimal_ms1_chemo}{TODO}

\item{ms1_only}{TODO}

\item{force}{TODO}
}
\value{
TODO
}
\description{
Title
}
