% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_smiles.R
\name{process_smiles}
\alias{process_smiles}
\title{Process SMILES}
\usage{
process_smiles(df, smiles_colname = "structure_smiles_initial", cache = NULL)
}
\arguments{
\item{df}{Dataframe to use}

\item{smiles_colname}{Column name of the SMILES variable}

\item{cache}{Cache where already processed SMILES are located}
}
\description{
This function processes SMILES
}
\examples{
data.frame(
  "structure_smiles_initial" = "C[C@H]1C=C(C(=O)[C@]2([C@H]1C[C@H]3[C@]4([C@H]2C(=O)C(=C([C@H]4CC(=O)O3)C)OC)C)C)OC"
) |>
  process_smiles()
}
