% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_smiles.R
\name{process_smiles}
\alias{process_smiles}
\title{Process SMILES}
\usage{
process_smiles(df, smiles_colname = "structure_smiles_initial", cache = NULL)
}
\arguments{
\item{df}{Dataframe to use}

\item{smiles_colname}{Column name of the SMILES variable}

\item{cache}{Cache where already processed SMILES are located}
}
\description{
This function processes SMILES
}
\examples{
NULL
}
