% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectra.R
\name{process_spectra}
\alias{process_spectra}
\title{Process spectra}
\usage{
process_spectra(
  input = params$input,
  library = params$library,
  output = params$output,
  method = params$ms$similarity$method,
  threshold = params$ms$similarity$threshold,
  ppm = params$ms$tolerance$ppm,
  dalton = params$ms$tolerance$dalton,
  npeaks = params$ms$peaks$absolute,
  rpeaks = params$ms$peaks$ratio,
  condition = params$ms$condition,
  quickmode = params$quickmode,
  parallel = params$parallel
)
}
\arguments{
\item{input}{TODO}

\item{library}{TODO}

\item{output}{TODO}

\item{method}{TODO}

\item{threshold}{TODO}

\item{ppm}{TODO}

\item{dalton}{TODO}

\item{npeaks}{TODO}

\item{rpeaks}{TODO}

\item{condition}{TODO}

\item{quickmode}{TODO}

\item{parallel}{TODO}
}
\value{
TODO
}
\description{
Process spectra
}
\examples{
TODO
}
