% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusters.R
\name{read_clusters}
\alias{read_clusters}
\title{Read clusters}
\usage{
read_clusters(id, workflow = params$gnps$workflow)
}
\arguments{
\item{id}{Character string of length 32 representing the job ID}

\item{workflow}{Character string indicating the type of workflow, either "fbmn" or "classical"}
}
\value{
A data frame containing cluster information
}
\description{
This function reads the clusters table from GNPS
}
\examples{
NULL
}
