% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_features.R
\name{read_features}
\alias{read_features}
\title{Read features}
\usage{
read_features(id)
}
\arguments{
\item{id}{a character string containing a GNPS job ID}
}
\value{
a data frame containing the features for the specified GNPS job
}
\description{
Read features
}
\examples{
features <- read_features(id = "abcdefghijklmnopqrstuvwxyz123456")
}
