% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_features.R
\name{read_features}
\alias{read_features}
\title{Read features}
\usage{
read_features(id)
}
\arguments{
\item{id}{Character string containing a GNPS job ID}
}
\value{
Data frame containing the features for the specified GNPS job
}
\description{
This function reads the features table from GNPS
}
\examples{
NULL
}
