% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_sirius_zip.R
\name{read_from_sirius_zip}
\alias{read_from_sirius_zip}
\title{Read from SIRIUS zip}
\usage{
read_from_sirius_zip(sirius_zip, file)
}
\arguments{
\item{sirius_zip}{Compressed directory containing the Sirius results}

\item{file}{File to be read}
}
\description{
This function reads files from Sirius compressed workspace
}
\examples{
NULL
}
