% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(id)
}
\arguments{
\item{id}{a character string containing a GNPS job ID}
}
\value{
a data frame containing the metadata for the specified GNPS job
}
\description{
TODO
}
\examples{
NULL
}
