% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{Read results}
\usage{
read_results(id, workflow)
}
\arguments{
\item{id}{a character string containing the job ID of the GNPS task}

\item{workflow}{a character string indicating the type of workflow used in the GNPS task. Supported values are 'fbmn' and 'classical'}
}
\value{
a data frame containing the results of the GNPS task
}
\description{
Read results
}
\examples{
read_results("abcdefghijklmnopqrstuvwxyz1234", workflow = "fbmn")
}
