% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_above_precursor.R
\name{remove_above_precursor}
\alias{remove_above_precursor}
\title{Remove peaks above precursor in MS2 spectra}
\usage{
remove_above_precursor(tol_mz = 0.5)
}
\arguments{
\item{tol_mz}{m/z tolerance}
}
\description{
This function remove peaks above precursor in MS2 spectra
}
\details{
Credit goes to Carolin Huber (0000-0002-9355-8948)
With fine tuning of Michele Stravs (0000-0002-1426-8572)
}
\examples{
NULL
}
