% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_gnps_job_id.R
\name{replace_gnps_job_id}
\alias{replace_gnps_job_id}
\title{Replace GNPS job ID}
\usage{
replace_gnps_job_id(
  x,
  default = yamls_default$prepare_params$gnps,
  params = params$gnps
)
}
\arguments{
\item{x}{a character string containing a GNPS job ID}

\item{default}{a default values for a GNPS job ID}

\item{params}{a user-specified values for a GNPS job ID}
}
\value{
a character string with the GNPS job ID modified according to the rules specified in the function
}
\description{
Replace GNPS job ID
}
\examples{
replaced_job_id <- replace_gnps_job_id("MyGNPSJobID", yamls_default, params)
}
