% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_id.R
\name{replace_id}
\alias{replace_id}
\title{Replace ID in file paths}
\usage{
replace_id(
  x,
  user_filename = get_params(step = "prepare_params")$files$pattern,
  user_gnps = get_params(step = "prepare_params")$gnps$id,
  example_gnps = get_default_paths()$gnps$example
)
}
\arguments{
\item{x}{a character string containing the default ID}

\item{user_filename}{a user-specified value for a file name job ID}

\item{user_gnps}{a user-specified value for a GNPS job ID}

\item{example_gnps}{an example value for a GNPS job ID}
}
\value{
Character string with the GNPS job ID modified according
to the rules specified in the function
}
\description{
This function replaces the default ID
in the example by a user-specified one
}
\examples{
tima:::replace_id(
  x = "example/123456_features.tsv",
  user_gnps = NULL,
  user_filename = "Foo"
)
}
