% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_reals.R
\name{round_reals}
\alias{round_reals}
\title{Round reals}
\usage{
round_reals(df, dig = 5)
}
\arguments{
\item{df}{Dataframe to use}

\item{dig}{Number of digits}
}
\description{
This function rounds some reals in a dataframe
}
\examples{
NULL
}
