% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run app}
\usage{
run_app(dir = "inst", host = "127.0.0.1", port = 3838, browser = TRUE)
}
\arguments{
\item{dir}{Directory of the app}

\item{host}{Host. Default to 127.0.0.1}

\item{port}{Port. Default to 3838}

\item{browser}{Flag for browser use. Default to TRUE}
}
\value{
Opens the app
}
\description{
This function runs the app
}
\examples{
NULL
}
