% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_spectra.R
\name{sanitize_spectra}
\alias{sanitize_spectra}
\title{Sanitize spectra}
\usage{
sanitize_spectra(
  spectra,
  ratio = 10000,
  cutoff = 10,
  fragments = 2,
  deeper = FALSE
)
}
\arguments{
\item{spectra}{Spectra object}

\item{ratio}{Minimal ratio to the max peak}

\item{cutoff}{Absolute minimal intensity}

\item{fragments}{Minimal number of fragments}

\item{deeper}{Deep sanitization using quantiles}
}
\description{
This function sanitizes spectra
}
\examples{
NULL
}
