% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_spectra.R
\name{sanitize_spectra}
\alias{sanitize_spectra}
\title{Sanitize spectra}
\usage{
sanitize_spectra(spectra, cutoff = 0, dalton = 0.01, polarity = NA, ppm = 10)
}
\arguments{
\item{spectra}{Spectra object}

\item{cutoff}{Absolute minimal intensity}

\item{dalton}{Dalton tolerance}

\item{polarity}{Polarity}

\item{ppm}{PPM tolerance}
}
\value{
The sanitized spectra
}
\description{
This function sanitizes spectra
}
\examples{
NULL
}
