% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_annotations_columns.R
\name{select_annotations_columns}
\alias{select_annotations_columns}
\title{Select annotations columns}
\usage{
select_annotations_columns(
  df,
  str_stereo = get("str_stereo", envir = parent.frame()),
  str_met = get("str_met", envir = parent.frame()),
  str_nam = get("str_nam", envir = parent.frame()),
  str_tax_cla = get("str_tax_cla", envir = parent.frame()),
  str_tax_npc = get("str_tax_npc", envir = parent.frame())
)
}
\arguments{
\item{df}{Dataframe}

\item{str_stereo}{File containing structures stereo}

\item{str_met}{File containing structures metadata}

\item{str_nam}{File containing structures names}

\item{str_tax_cla}{File containing Classyfire taxonomy}

\item{str_tax_npc}{File containing NPClassifier taxonomy}
}
\value{
The dataframe with annotation columns selected
}
\description{
This function selects annotations columns
}
\examples{
NULL
}
