% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_annotations_columns.R
\name{select_annotations_columns}
\alias{select_annotations_columns}
\title{Select annotations columns}
\usage{
select_annotations_columns(df)
}
\arguments{
\item{df}{Dataframe}
}
\description{
This function selects annotations columns
}
\examples{
NULL
}
