% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sirius_columns.R
\name{select_sirius_columns_formulas}
\alias{select_sirius_columns_formulas}
\title{Select sirius columns (formulas)}
\usage{
select_sirius_columns_formulas(df, sirius_version)
}
\arguments{
\item{df}{Dataframe}

\item{sirius_version}{Sirius version}
}
\value{
The dataframe with selected sirius columns
}
\description{
This function selects sirius columns (formulas)
}
\examples{
NULL
}
