% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sirius_columns.R
\name{select_sirius_columns_structures}
\alias{select_sirius_columns_structures}
\title{Select sirius columns (structures)}
\usage{
select_sirius_columns_structures(df, sirius_version)
}
\arguments{
\item{df}{Dataframe}

\item{sirius_version}{Sirius version}
}
\description{
This function selects sirius columns (structures)
}
\examples{
NULL
}
