% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sop_columns.R
\name{select_sop_columns}
\alias{select_sop_columns}
\title{Select SOP columns}
\usage{
select_sop_columns(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
The dataframe with selected structure organism pairs columns
}
\description{
This function selects sop columns
}
\examples{
NULL
}
