% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_logger.R
\name{setup_logger}
\alias{setup_logger}
\title{Setup logger}
\usage{
setup_logger(filename)
}
\arguments{
\item{filename}{Log file name. Default: \code{"tima.log"}.}
}
\description{
Setup logger
}
\keyword{internal}
