% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_logger.R
\name{setup_logger}
\alias{setup_logger}
\title{Setup logger}
\usage{
setup_logger(filename = "tima.log")
}
\arguments{
\item{filename}{Character string specifying the log file name.
Default: "tima.log"}
}
\value{
NULL (invisibly). Sets up logger as side effect.
}
\description{
Configures the logger for the package with trace-level logging
and output to both console and a log file
}
\examples{
\dontrun{
setup_logger("analysis.log")
}
}
\keyword{internal}
