% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables_sop.R
\name{split_tables_sop}
\alias{split_tables_sop}
\title{Split Structure Organism Pairs table}
\usage{
split_tables_sop(table, cache)
}
\arguments{
\item{table}{Table to split}

\item{cache}{Cache where already processed SMILES are located}
}
\value{
A list of tables from the structure organism pairs tables
}
\description{
This function splits the structure organism table.
}
\examples{
NULL
}
