% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summarize_results}
\alias{summarize_results}
\title{Summarize results}
\usage{
summarize_results(
  df,
  features_table,
  components_table,
  structure_organism_pairs_table,
  annot_table_wei_chemo,
  remove_ties,
  summarize
)
}
\arguments{
\item{df}{Dataframe}

\item{features_table}{Prepared features file}

\item{components_table}{Prepared components file}

\item{structure_organism_pairs_table}{Table containing the
structure - organism pairs}

\item{annot_table_wei_chemo}{Table containing your
chemically weighted annotation}

\item{remove_ties}{Remove ties. BOOLEAN}

\item{summarize}{Boolean. summarize results (1 row per feature)}
}
\value{
A summarized table
}
\description{
This function summarizes results
}
\examples{
NULL
}
