% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tima-package.R
\docType{package}
\name{tima-package}
\alias{tima}
\alias{tima-package}
\title{tima: Taxonomically Informed Metabolite Annotation}
\description{
This package provides the infrastructure to perform Taxonomically Informed Metabolite Annotation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/taxonomicallyinformedannotation/tima}
  \item \url{https://taxonomicallyinformedannotation.github.io/tima-shinylive}
  \item \url{https://taxonomicallyinformedannotation.github.io/tima}
  \item Report bugs at \url{https://github.com/taxonomicallyinformedannotation/tima/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{adafede@gmail.com} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

Other contributors:
\itemize{
  \item Pierre-Marie Allard \email{pierre-marie.allard@unifr.ch} (\href{https://orcid.org/0000-0003-3389-2191}{ORCID}) [contributor]
}

}
\keyword{internal}
