% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_score_sirius_csi.R
\name{transform_score_sirius_csi}
\alias{transform_score_sirius_csi}
\title{Transform score sirius CSI}
\usage{
transform_score_sirius_csi(csi_score, K = 50, scale = 10)
}
\arguments{
\item{csi_score}{Original CSI score}

\item{K}{Shift}

\item{scale}{Scale}
}
\value{
A mass
}
\description{
This function calculates the mass of M
}
\examples{
NULL
}
