% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_annotations.R
\name{weight_annotations}
\alias{weight_annotations}
\title{Weight annotations}
\usage{
weight_annotations(
  library = paths$data$interim$libraries$sop$merged$keys,
  org_tax_ott = paths$data$interim$libraries$sop$merged$organisms$taxonomies$ott,
  str_2d_3d = paths$data$interim$libraries$sop$merged$structures$dd_ddd,
  annotations = params$files$annotations$filtered,
  components = params$files$networks$spectral$components$prepared,
  edges = params$files$networks$spectral$edges$prepared,
  taxa = params$files$taxa$prepared,
  output = params$files$annotations$processed,
  candidates_initial = params$annotations$candidates$initial,
  candidates_final = params$annotations$candidates$final,
  weight_spectral = params$weights$global$spectral,
  weight_chemical = params$weights$global$chemical,
  weight_biological = params$weights$global$biological,
  score_biological_domain = params$weights$biological$domain,
  score_biological_kingdom = params$weights$biological$kingdom,
  score_biological_phylum = params$weights$biological$phylum,
  score_biological_class = params$weights$biological$class,
  score_biological_order = params$weights$biological$order,
  score_biological_infraorder = params$weights$biological$infraorder,
  score_biological_family = params$weights$biological$family,
  score_biological_subfamily = params$weights$biological$subfamily,
  score_biological_tribe = params$weights$biological$tribe,
  score_biological_subtribe = params$weights$biological$subtribe,
  score_biological_genus = params$weights$biological$genus,
  score_biological_subgenus = params$weights$biological$subgenus,
  score_biological_species = params$weights$biological$species,
  score_biological_subspecies = params$weights$biological$subspecies,
  score_biological_variety = params$weights$biological$variety,
  score_chemical_cla_kingdom = params$weights$chemical$cla$kingdom,
  score_chemical_cla_superclass = params$weights$chemical$cla$superclass,
  score_chemical_cla_class = params$weights$chemical$cla$class,
  score_chemical_cla_parent = params$weights$chemical$cla$parent,
  score_chemical_npc_pathway = params$weights$chemical$npc$pathway,
  score_chemical_npc_superclass = params$weights$chemical$npc$superclass,
  score_chemical_npc_class = params$weights$chemical$npc$class,
  minimal_ms1_bio = params$annotations$ms1$thresholds$biological,
  minimal_ms1_chemo = params$annotations$ms1$thresholds$chemical,
  ms1_only = params$annotations$ms1only,
  summarise = params$options$summarise,
  pattern = params$files$pattern,
  force = params$options$force,
  parameters = params
)
}
\arguments{
\item{library}{Library containing the keys}

\item{org_tax_ott}{File containing organisms taxonomy (OTT)}

\item{str_2d_3d}{File containing 2D and 3D structures}

\item{annotations}{Prepared annotations file}

\item{components}{Prepared components file}

\item{edges}{Prepared edges file}

\item{taxa}{Prepared taxed features file}

\item{output}{Output file}

\item{candidates_initial}{Number of initial candidates to keep}

\item{candidates_final}{Number of final candidates to keep}

\item{weight_spectral}{Weight for the spectral score}

\item{weight_chemical}{Weight for the biological score}

\item{weight_biological}{Weight for the chemical consistency score}

\item{score_biological_domain}{Score for a \code{domain} match
(should be lower than \code{kingdom})}

\item{score_biological_kingdom}{Score for a \code{kingdom} match
(should be lower than \code{phylum})}

\item{score_biological_phylum}{Score for a \code{phylum} match
(should be lower than \code{class})}

\item{score_biological_class}{Score for a \code{class} match
(should be lower than \code{order})}

\item{score_biological_order}{Score for a \code{order} match
(should be lower than \code{infraorder})}

\item{score_biological_infraorder}{Score for a \code{infraorder} match
(should be lower than \code{order})}

\item{score_biological_family}{Score for a \code{family} match
(should be lower than \code{subfamily})}

\item{score_biological_subfamily}{Score for a \code{subfamily} match
(should be lower than \code{family})}

\item{score_biological_tribe}{Score for a \code{tribe} match
(should be lower than \code{subtribe})}

\item{score_biological_subtribe}{Score for a \code{subtribe} match
(should be lower than \code{genus})}

\item{score_biological_genus}{Score for a \code{genus} match
(should be lower than \code{subgenus})}

\item{score_biological_subgenus}{Score for a \code{subgenus} match
(should be lower than \code{species})}

\item{score_biological_species}{Score for a \code{species} match
(should be lower than \code{subspecies})}

\item{score_biological_subspecies}{Score for a \code{subspecies} match
(should be lower than \code{variety})}

\item{score_biological_variety}{Score for a \code{variety} match
(should be the highest)}

\item{score_chemical_cla_kingdom}{Score for a \verb{Classyfire kingdom} match
(should be lower than \verb{ Classyfire superclass})}

\item{score_chemical_cla_superclass}{Score for a \verb{Classyfire superclass} match
(should be lower than \verb{Classyfire class})}

\item{score_chemical_cla_class}{Score for a \verb{Classyfire class} match
(should be lower than \verb{Classyfire parent})}

\item{score_chemical_cla_parent}{Score for a \verb{Classyfire parent} match
(should be the highest)}

\item{score_chemical_npc_pathway}{Score for a \verb{NPC pathway} match
(should be lower than \verb{ NPC superclass})}

\item{score_chemical_npc_superclass}{Score for a \verb{NPC superclass}
match (should be lower than \verb{NPC class})}

\item{score_chemical_npc_class}{Score for a \verb{NPC class} match
(should be the highest)}

\item{minimal_ms1_bio}{Minimal biological score to keep MS1 based annotation}

\item{minimal_ms1_chemo}{Minimal chemical score to keep MS1 based annotation}

\item{ms1_only}{Boolean. Keep only MS1 annotations}

\item{summarise}{Boolean. Summarize results (1 row per feature)}

\item{pattern}{Pattern to identify your job. STRING}

\item{force}{Force parameters. Use it at your own risk}

\item{parameters}{Params}
}
\description{
This function weights annotations.
}
\examples{
NULL
}
\seealso{
annotate_masses weight_bio weight_chemo
}
