% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_bio.R
\name{weight_bio}
\alias{weight_bio}
\title{Weight bio}
\usage{
weight_bio(
  annotationTable = annotation_table_taxed,
  structureOrganismPairsTable = structure_organism_pairs_table,
  weightSpectral = weight_spectral,
  weightBiological = weight_biological,
  scoreBiologicalDomain = score_biological_domain,
  scoreBiologicalKingdom = score_biological_kingdom,
  scoreBiologicalPhylum = score_biological_phylum,
  scoreBiologicalClass = score_biological_class,
  scoreBiologicalOrder = score_biological_order,
  scoreBiologicalFamily = score_biological_family,
  scoreBiologicalTribe = score_biological_tribe,
  scoreBiologicalGenus = score_biological_genus,
  scoreBiologicalSpecies = score_biological_species,
  scoreBiologicalVariety = score_biological_variety
)
}
\arguments{
\item{annotationTable}{Table containing the initial annotation eventually complemented by additional MS1 annotations}

\item{structureOrganismPairsTable}{Table containing the structure - organism pairs}

\item{weightSpectral}{Weight for the spectral score}

\item{weightBiological}{Weight for the biological score}

\item{scoreBiologicalDomain}{Score for a \code{domain} match (should be lower than \code{kingdom})}

\item{scoreBiologicalKingdom}{Score for a \code{kingdom} match (should be lower than \code{phylum})}

\item{scoreBiologicalPhylum}{Score for a \code{phylum} match (should be lower than \code{class})}

\item{scoreBiologicalClass}{Score for a \code{class} match (should be lower than \code{order})}

\item{scoreBiologicalOrder}{Score for a \code{order} match (should be lower than \code{family})}

\item{scoreBiologicalFamily}{Score for a \code{family} match (should be lower than \code{tribe})}

\item{scoreBiologicalTribe}{Score for a \code{tribe} match (should be lower than \code{genus})}

\item{scoreBiologicalGenus}{Score for a \code{genus} match (should be lower than \code{species})}

\item{scoreBiologicalSpecies}{Score for a \code{species} match (should be lower than \code{variety})}

\item{scoreBiologicalVariety}{Score for a \code{variety} match (should be the highest)}
}
\value{
A table containing the biologically weighted annotation
}
\description{
This function weights the eventually MS1 complemented annotations according their biological source
}
\examples{
NULL
}
