% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_bio.R
\name{weight_bio}
\alias{weight_bio}
\title{Weight bio}
\usage{
weight_bio(
  annotation_table_taxed = get("annotation_table_taxed", envir = parent.frame()),
  structure_organism_pairs_table = get("structure_organism_pairs_table", envir =
    parent.frame()),
  weight_spectral = get("weight_spectral", envir = parent.frame()),
  weight_biological = get("weight_biological", envir = parent.frame()),
  score_biological_domain = get("score_biological_domain", envir = parent.frame()),
  score_biological_kingdom = get("score_biological_kingdom", envir = parent.frame()),
  score_biological_phylum = get("score_biological_phylum", envir = parent.frame()),
  score_biological_class = get("score_biological_class", envir = parent.frame()),
  score_biological_order = get("score_biological_order", envir = parent.frame()),
  score_biological_family = get("score_biological_family", envir = parent.frame()),
  score_biological_tribe = get("score_biological_tribe", envir = parent.frame()),
  score_biological_genus = get("score_biological_genus", envir = parent.frame()),
  score_biological_species = get("score_biological_species", envir = parent.frame()),
  score_biological_variety = get("score_biological_variety", envir = parent.frame())
)
}
\arguments{
\item{annotation_table_taxed}{Table containing the initial annotation eventually complemented by additional MS1 annotations}

\item{structure_organism_pairs_table}{Table containing the structure - organism pairs}

\item{weight_spectral}{Weight for the spectral score}

\item{weight_biological}{Weight for the biological score}

\item{score_biological_domain}{Score for a \code{domain} match (should be lower than \code{kingdom})}

\item{score_biological_kingdom}{Score for a \code{kingdom} match (should be lower than \code{phylum})}

\item{score_biological_phylum}{Score for a \code{phylum} match (should be lower than \code{class})}

\item{score_biological_class}{Score for a \code{class} match (should be lower than \code{order})}

\item{score_biological_order}{Score for a \code{order} match (should be lower than \code{family})}

\item{score_biological_family}{Score for a \code{family} match (should be lower than \code{tribe})}

\item{score_biological_tribe}{Score for a \code{tribe} match (should be lower than \code{genus})}

\item{score_biological_genus}{Score for a \code{genus} match (should be lower than \code{species})}

\item{score_biological_species}{Score for a \code{species} match (should be lower than \code{variety})}

\item{score_biological_variety}{Score for a \code{variety} match (should be the highest)}
}
\value{
A table containing the biologically weighted annotation
}
\description{
This function weights the eventually MS1 complemented annotations according their biological source
}
\examples{
NULL
}
