% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_chemo.R
\name{weight_chemo}
\alias{weight_chemo}
\title{Weight chemo}
\usage{
weight_chemo(
  annotationTableWeightedBioCleaned = annotation_table_weighted_bio_cleaned,
  weightSpectral = weight_spectral,
  weightBiological = weight_biological,
  weightChemical = weight_chemical,
  scoreChemicalPathway = score_chemical_pathway,
  scoreChemicalSuperclass = score_chemical_superclass,
  scoreChemicalClass = score_chemical_class
)
}
\arguments{
\item{annotationTableWeightedBioCleaned}{TODO}

\item{weightSpectral}{TODO}

\item{weightBiological}{TODO}

\item{weightChemical}{TODO}

\item{scoreChemicalPathway}{TODO}

\item{scoreChemicalSuperclass}{TODO}

\item{scoreChemicalClass}{TODO}
}
\value{
TODO
}
\description{
Weight chemo
}
\examples{
TODO
}
