% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_chemo.R
\name{weight_chemo}
\alias{weight_chemo}
\title{Weight chemo}
\usage{
weight_chemo(
  annotationTableWeightedBioCleaned = annotation_table_weighted_bio_cleaned,
  weightSpectral = weight_spectral,
  weightBiological = weight_biological,
  weightChemical = weight_chemical,
  scoreChemicalPathway = score_chemical_pathway,
  scoreChemicalSuperclass = score_chemical_superclass,
  scoreChemicalClass = score_chemical_class
)
}
\arguments{
\item{annotationTableWeightedBioCleaned}{Table containing the biologically weighted annotation}

\item{weightSpectral}{Weight for the spectral score}

\item{weightBiological}{Weight for the biological score}

\item{weightChemical}{Weight for the chemical consistency score}

\item{scoreChemicalPathway}{Score for a \code{pathway} match (should be lower than \code{superclass})}

\item{scoreChemicalSuperclass}{Score for a \code{superclass} match (should be lower than \code{class})}

\item{scoreChemicalClass}{Score for a \code{class} match (should be the highest)}
}
\value{
A table containing the chemically weighted annotation
}
\description{
This function weights the biologically weigthed annotations according their chemical consistency
}
\examples{
NULL
}
