% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_chemo.R
\name{weight_chemo}
\alias{weight_chemo}
\title{Weight chemo}
\usage{
weight_chemo(
  annotationTableWeightedBioCleaned = annotation_table_weighted_bio_cleaned,
  weightSpectral = weight_spectral,
  weightBiological = weight_biological,
  weightChemical = weight_chemical,
  scoreChemicalClaKingdom = score_chemical_cla_kingdom,
  scoreChemicalClaSuperclass = score_chemical_cla_superclass,
  scoreChemicalClaClass = score_chemical_cla_class,
  scoreChemicalClaParent = score_chemical_cla_parent,
  scoreChemicalNpcPathway = score_chemical_npc_pathway,
  scoreChemicalNpcSuperclass = score_chemical_npc_superclass,
  scoreChemicalNpcClass = score_chemical_npc_class
)
}
\arguments{
\item{annotationTableWeightedBioCleaned}{Table containing the biologically weighted annotation}

\item{weightSpectral}{Weight for the spectral score}

\item{weightBiological}{Weight for the biological score}

\item{weightChemical}{Weight for the chemical consistency score}

\item{scoreChemicalClaKingdom}{Score for a \verb{Classyfire kingdom} match (should be lower than \verb{ Classyfire superclass})}

\item{scoreChemicalClaSuperclass}{Score for a \verb{Classyfire superclass} match (should be lower than \verb{Classyfire class})}

\item{scoreChemicalClaClass}{Score for a \verb{Classyfire class} match (should be lower than \verb{Classyfire parent})}

\item{scoreChemicalClaParent}{Score for a \verb{Classyfire parent} match (should be the highest)}

\item{scoreChemicalNpcPathway}{Score for a \code{pathway} match (should be lower than \code{superclass})}

\item{scoreChemicalNpcSuperclass}{Score for a \code{superclass} match (should be lower than \code{class})}

\item{scoreChemicalNpcClass}{Score for a \code{class} match (should be the highest)}
}
\value{
A table containing the chemically weighted annotation
}
\description{
This function weights the biologically weighted annotations according their chemical consistency
}
\examples{
NULL
}
