% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WEIGHT_SUM_TOLERANCE}
\alias{WEIGHT_SUM_TOLERANCE}
\title{Tolerance for floating point weight sum comparisons
Used when validating that weights sum to 1.0}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WEIGHT_SUM_TOLERANCE
}
\description{
Tolerance for floating point weight sum comparisons
Used when validating that weights sum to 1.0
}
\keyword{datasets}
